Convert flytable cell type information into a neuroglancer info file

Convert flytable cell type information into a neuroglancer info file

fct2nginfo(f, ids = NULL, version = NULL, sep = "_", gluestr = NULL, ...)

Arguments

f

Path to a json file. write_nginfo will create the enclosing directory if necessary.

ids

FlyWire root ids in any form understood by flywire_ids

version

Optional numeric CAVE version (see flywire_cave_query). The special signalling value of TRUE uses the current default data dump as returned by flywire_connectome_data_version.

sep

The separator used to paste multiple columns of annotations together.

gluestr

Optional string passed to glue::glue which is interpreted in the context of the annotation data.frame produced by flytable_meta. This allows arbitrary formatting for

...

Additional arguments passed to flytable_meta and then eventually flytable_cell_types.

Value

The path f invisibly

See also

Examples

if (FALSE) {
# all neurons in info table
fct2nginfo(f='path/to/flytablev526/info', version=526)
fct2nginfo(f='path/to/hemilineagev526/info', version=526,
  gluestr="{ito_lee_hemilineage}_{toupper(substr(side,1,1))}")
fct2nginfo("MBON%", 'path/to/mboninfo/info')
}