Find all merged segment ids for a given starting google segment id

This is a naive implementation that only depends only on the mergeinfo object created by read_mergeinfo. It is retained only to verify correctness. It is slow but does not require any pre-processing of the table of merge information.

find_merged_segments_slow(x, mergeinfo)

Arguments

x

a segment id or any other input that can be interpreted by ngl_segments

mergeinfo

The merge information data.frame created by read_mergeinfo

Value

vector of segment ids

See also