Flytable database queries

flytable_query performs a SQL query against a flytable database. You can omit the base argument unless you have tables of the same name in different bases.

Usage

flytable_list_rows(
  table,
  base = NULL,
  view_name = NULL,
  order_by = NULL,
  desc = FALSE,
  start = 0L,
  limit = Inf,
  collapse_lists = TRUE,
  python = FALSE,
  chunksize = NULL
)

flytable_query(
  sql,
  limit = 100000L,
  base = NULL,
  python = FALSE,
  convert = TRUE,
  collapse_lists = TRUE
)

Arguments

table

The name of a table inside your database

base

Character vector naming a seatable base (recommended) or a Base object returned by flytable_base (expert use).

view_name

An optional view which may limit the rows/columns displayed.

order_by

Optional name of columns to order results

desc

Whether to use descending order (default FALSE => ascending order)

start

Optional starting row

limit

An optional limit, which only applies if you do not specify a limit directly in the sql query. By default seatable limits SQL queries to 100 rows. We increase the limit to 100000 rows by default.

collapse_lists

Whether to collapse any list multi-select columns into simple strings. The default value of collapse_lists=TRUE will comma separate them.

python

Whether to return a Python pandas DataFrame. The default of FALSE returns an R data.frame

chunksize

Optional The maximum number of rows to request in one web request. For advanced use only as the default value of NULL will fetch as many as possible.

sql

A SQL query string. See examples and seatable docs.

convert

Expert use only: Whether or not to allow the Python seatable module to process raw output from the database. This is is principally for debugging purposes. NB this imposes a requirement of seatable_api >=2.4.0.

Value

An R data.frame or Pandas DataFrame depending on the value of the python argument.

a data.frame of results. There should be 0 rows if no rows matched query.

Details

Flytable uses programmatic access to the seatable API.

See also

Examples

# \donttest{
flytable_list_rows(table = "testfruit")
#>                       _id              _mtime              _ctime  fruit_name
#> 1  H8BckTnXRL2PaVuLDcRPMA 2026-08-04 14:22:21 2021-12-17 13:36:17       apple
#> 2  SrBW3vFLRxafKRYHoPrGQQ 2022-05-12 16:58:08 2021-12-17 13:36:17      banana
#> 3  F1h-TZKpTcWiHujZ2xnF8Q 2024-08-28 01:30:00 2021-12-17 13:36:17  clementine
#> 4  dwTLFmsDRoCxhE3BwY-ovQ 2026-08-07 12:15:23 2024-08-28 01:30:00        kiwi
#> 5  Im6VZG_DQ7mRJh6BAUwW7A 2024-08-28 01:31:13 2024-08-28 01:31:13        kiwi
#> 6  YsZ-5JndSpW96JW9cZZ8dw 2026-08-04 14:11:02 2026-05-16 16:31:01        kiwi
#> 7  eQVD8GmEQmq11XxfggkoVA 2026-08-04 13:58:58 2026-05-16 16:31:05        kiwi
#> 8  AtYZUS-EQT2NGEpMBco_iw 2026-08-04 13:59:02 2026-05-16 16:35:36        kiwi
#> 9  FWYTJZgDQyioqcRnVvM4TA 2026-08-04 02:40:41 2026-08-04 02:40:41        kiwi
#> 10 NpVTQf72REeeDYiQHhD7bA 2026-08-04 02:43:29 2026-08-04 02:43:29        kiwi
#> 11 cMs7rchlTOO6ArPnVtIbwQ 2026-08-04 07:33:15 2026-08-04 07:33:15        kiwi
#> 12 NkDj38vMR5yHz2yNvJWQug 2026-08-04 07:34:16 2026-08-04 07:34:16        kiwi
#> 13 YMZSVgaMS3KdvoVJ15gkaQ 2026-08-04 07:34:32 2026-08-04 07:34:32        kiwi
#> 14 NE2aRAPpQqyaSFau39x-NQ 2026-08-04 15:25:33 2026-08-04 15:25:33        kiwi
#> 15 Ugt6lOQ4SPGSIRN3KYjZaQ 2026-08-04 15:35:21 2026-08-04 15:35:21        kiwi
#> 16 JCkxWltZTQ-YpYQOumajvQ 2026-08-04 16:20:32 2026-08-04 16:20:32        kiwi
#> 17 OimH9KKhSUWbRSDjw4f3Sg 2026-08-05 11:38:09 2026-08-05 11:38:09        kiwi
#> 18 Z6a6aS1pSfqxNor6mYy_2A 2026-08-05 22:09:33 2026-08-05 22:09:33        kiwi
#> 19 YtGyNFbeQLWYmS8uOiyaQA 2026-08-05 22:28:04 2026-08-05 22:28:04        kiwi
#> 20 Y-7LDG-3QtePwU-3lQIJew 2026-08-05 22:37:32 2026-08-05 22:37:32        kiwi
#> 21 IGCLe4OyQ8SEbX8W4dNnXQ 2026-08-05 22:53:32 2026-08-05 22:53:32        kiwi
#> 22 L5_EZ_8SQIiEiwGmiUN68A 2026-08-06 01:35:28 2026-08-06 01:35:28        kiwi
#> 23 WOz9dZSYSxqJ2PASy1L9vQ 2026-08-06 01:57:27 2026-08-06 01:57:27        kiwi
#> 24 KHa1G7FUQQqbspZyQfqPBw 2026-08-06 02:00:20 2026-08-06 02:00:20        kiwi
#> 25 M7fnq2OVQXK1w5HJfXiSAQ 2026-08-06 02:01:24 2026-08-06 02:01:24        kiwi
#> 26 EHyFiZJfQ9is_kZWETSO4Q 2026-08-06 02:04:20 2026-08-06 02:04:20        kiwi
#> 27 EHl71b0ZTMKgzk3068dQrw 2026-08-06 02:29:36 2026-08-06 02:29:36        kiwi
#> 28 Bx-VOyOxTxazlZZ7s0_CdQ 2026-08-06 02:30:45 2026-08-06 02:30:45        kiwi
#> 29 A99gi51PQ7-ceAnmF7LMPw 2026-08-06 02:37:03 2026-08-06 02:37:03        kiwi
#> 30 HmI5aEbtT6ayx6xH4VnOaw 2026-08-06 07:27:33 2026-08-06 07:27:33 dragonfruit
#> 31 Hi9ONOFDTTKp_Yl-aSctgw 2026-08-06 07:27:36 2026-08-06 07:27:34   starfruit
#> 32 eUB140NYS4GkF7DPSx1thw 2026-08-06 07:27:59 2026-08-06 07:27:59 dragonfruit
#> 33 FzJthRgqT8Wfi6dKux8UsQ 2026-08-06 07:28:02 2026-08-06 07:28:00   starfruit
#> 34 dPWqcY8NSXaxXvlKrm2eeg 2026-08-06 07:30:41 2026-08-06 07:30:41        kiwi
#> 35 ZZJWrJGrRme2W45B_B-_uA 2026-08-06 07:32:14 2026-08-06 07:32:14        kiwi
#> 36 DeaPuxpDSAadYNV5gPM8Hw 2026-08-06 08:08:31 2026-08-06 08:08:31        kiwi
#> 37 I1GuWY5cT4-PP7jB-FTBiw 2026-08-06 08:10:00 2026-08-06 08:10:00        kiwi
#> 38 bJN3l4UwS1KEXeHpC7w5_g 2026-08-06 08:10:54 2026-08-06 08:10:54        kiwi
#> 39 WYB5jGVnSymeVx8oGKeA_g 2026-08-06 08:12:18 2026-08-06 08:12:18        kiwi
#> 40 Yxb_aAdETfG5jVWB7tE9xg 2026-08-06 22:45:59 2026-08-06 22:45:59        kiwi
#> 41 a5U8meTJSNaitX85ZYaSwQ 2026-08-06 23:15:23 2026-08-06 23:15:23        kiwi
#> 42 CVuZJLScQL-huoErGQQUSg 2026-08-06 23:18:44 2026-08-06 23:18:44        kiwi
#> 43 S63Uerd9QbyZaZRdKTen9g 2026-08-06 23:21:04 2026-08-06 23:21:04        kiwi
#> 44 aFFQR0MHT7SwBzHVOg7n_w 2026-08-07 03:32:39 2026-08-07 03:32:39        kiwi
#> 45 X6re_fP1QiCToV7y6omezg 2026-08-07 03:34:57 2026-08-07 03:34:57        kiwi
#> 46 L_cTHvP0S6653S6y40u3Dg 2026-08-07 03:35:43 2026-08-07 03:35:43        kiwi
#> 47 Syyip8pSSPKecbf5M3v2hg 2026-08-07 05:03:49 2026-08-07 05:03:49        kiwi
#> 48 ZdCzpd4VSzylY_0J53839g 2026-08-07 05:06:17 2026-08-07 05:06:17        kiwi
#> 49 N8D0-hAIRuueA9YBzywQSw 2026-08-07 05:08:35 2026-08-07 05:08:35        kiwi
#> 50 cb8aKpHZS5aSR6-y0sUWzw 2026-08-07 11:36:13 2026-08-07 11:36:13        kiwi
#> 51 D90iNFwXRk-9vwjhaUIr2w 2026-08-07 11:38:48 2026-08-07 11:38:48        kiwi
#> 52 A-41sdysQ2iJKzCY4HLQGw 2026-08-07 11:39:09 2026-08-07 11:39:09        kiwi
#> 53 d6D46fD8SUyMzkWaOB4Pyg 2026-08-07 11:54:14 2026-08-07 11:54:14        kiwi
#> 54 XBSX9jgBQUaYnOm_3k0KRQ 2026-08-07 11:55:24 2026-08-07 11:55:24        kiwi
#> 55 a52x6PXEQhO0wa4z9NG5Cw 2026-08-07 11:58:13 2026-08-07 11:58:13        kiwi
#> 56 LXBPookCSDiKWfzH1RdEBg 2026-08-07 12:14:56 2026-08-07 12:14:56        kiwi
#> 57 N2ioPI_sRtqZx4zhoRejCQ 2026-08-07 12:15:23 2026-08-07 12:15:23        kiwi
#>        nid              person       last_modified date_nominute
#> 1        1               Alice 2026-08-04 14:22:21    2022-01-06
#> 2        2                 Bob 2022-05-12 16:58:08    2022-01-03
#> 3        3               Clara 2024-08-28 01:30:00    2021-08-05
#> 4   976376 Frederick the Great 2026-08-07 12:15:23          <NA>
#> 5  7706772 Frederick the Great 2024-08-28 01:31:13          <NA>
#> 6  7592831 Frederick the Great 2026-08-04 14:11:02          <NA>
#> 7  8095327 Frederick the Great 2026-08-04 13:58:58          <NA>
#> 8  3462194 Frederick the Great 2026-08-04 13:59:02          <NA>
#> 9  2442444 Frederick the Great 2026-08-04 02:40:41          <NA>
#> 10 4079011 Frederick the Great 2026-08-04 02:43:29          <NA>
#> 11 3924912 Frederick the Great 2026-08-04 07:33:15          <NA>
#> 12 9139682 Frederick the Great 2026-08-04 07:34:16          <NA>
#> 13 5170183 Frederick the Great 2026-08-04 07:34:32          <NA>
#> 14 6151012 Frederick the Great 2026-08-04 15:25:33          <NA>
#> 15 4401168 Frederick the Great 2026-08-04 15:35:21          <NA>
#> 16 4361950 Frederick the Great 2026-08-04 16:20:32          <NA>
#> 17 5312251 Frederick the Great 2026-08-05 11:38:09          <NA>
#> 18 7868204 Frederick the Great 2026-08-05 22:09:33          <NA>
#> 19 1890153 Frederick the Great 2026-08-05 22:28:04          <NA>
#> 20 6994939 Frederick the Great 2026-08-05 22:37:32          <NA>
#> 21 7694898 Frederick the Great 2026-08-05 22:53:32          <NA>
#> 22 7255705 Frederick the Great 2026-08-06 01:35:28          <NA>
#> 23  329511 Frederick the Great 2026-08-06 01:57:27          <NA>
#> 24 9321031 Frederick the Great 2026-08-06 02:00:20          <NA>
#> 25 4387483 Frederick the Great 2026-08-06 02:01:24          <NA>
#> 26 8784907 Frederick the Great 2026-08-06 02:04:20          <NA>
#> 27 9478480 Frederick the Great 2026-08-06 02:29:36          <NA>
#> 28 9338701 Frederick the Great 2026-08-06 02:30:45          <NA>
#> 29 8411708 Frederick the Great 2026-08-06 02:37:03          <NA>
#> 30 1683741     Delta Sync Test 2026-08-06 07:27:33          <NA>
#> 31 2339781   Multi Select Test 2026-08-06 07:27:36          <NA>
#> 32 3508782     Delta Sync Test 2026-08-06 07:27:59          <NA>
#> 33 6158314   Multi Select Test 2026-08-06 07:28:02          <NA>
#> 34 5578437 Frederick the Great 2026-08-06 07:30:41          <NA>
#> 35 8801360 Frederick the Great 2026-08-06 07:32:14          <NA>
#> 36 7882694 Frederick the Great 2026-08-06 08:08:31          <NA>
#> 37 9472171 Frederick the Great 2026-08-06 08:10:00          <NA>
#> 38 1173603 Frederick the Great 2026-08-06 08:10:54          <NA>
#> 39 9855950 Frederick the Great 2026-08-06 08:12:18          <NA>
#> 40 4860898 Frederick the Great 2026-08-06 22:45:59          <NA>
#> 41 6449285 Frederick the Great 2026-08-06 23:15:23          <NA>
#> 42 3955252 Frederick the Great 2026-08-06 23:18:44          <NA>
#> 43 2255968 Frederick the Great 2026-08-06 23:21:04          <NA>
#> 44 9442450 Frederick the Great 2026-08-07 03:32:39          <NA>
#> 45 9014656 Frederick the Great 2026-08-07 03:34:57          <NA>
#> 46 7039635 Frederick the Great 2026-08-07 03:35:43          <NA>
#> 47 5103791 Frederick the Great 2026-08-07 05:03:49          <NA>
#> 48 3864773 Frederick the Great 2026-08-07 05:06:17          <NA>
#> 49 9911958 Frederick the Great 2026-08-07 05:08:35          <NA>
#> 50 1547744 Frederick the Great 2026-08-07 11:36:13          <NA>
#> 51 3256988 Frederick the Great 2026-08-07 11:38:48          <NA>
#> 52 5457118 Frederick the Great 2026-08-07 11:39:09          <NA>
#> 53 6517708 Frederick the Great 2026-08-07 11:54:14          <NA>
#> 54 6034946 Frederick the Great 2026-08-07 11:55:24          <NA>
#> 55 9981019 Frederick the Great 2026-08-07 11:58:13          <NA>
#> 56 6856049 Frederick the Great 2026-08-07 12:14:56          <NA>
#> 57 3784415 Frederick the Great 2026-08-07 12:15:23          <NA>
#>           date_wminute                                        user
#> 1  2022-01-12 09:30:00 8adf4f5dd661449fa6cc1f5a0b1815c0@auth.local
#> 2  2022-01-03 07:56:00 c7efb8019da54923a9b04d4a74f0fde8@auth.local
#> 3  2021-08-05 08:30:00 c7efb8019da54923a9b04d4a74f0fde8@auth.local
#> 4                 <NA>                                         NaN
#> 5                 <NA>                                         NaN
#> 6                 <NA>                                         NaN
#> 7                 <NA>                                         NaN
#> 8                 <NA>                                         NaN
#> 9                 <NA>                                         NaN
#> 10                <NA>                                         NaN
#> 11                <NA>                                         NaN
#> 12                <NA>                                         NaN
#> 13                <NA>                                         NaN
#> 14                <NA>                                         NaN
#> 15                <NA>                                         NaN
#> 16                <NA>                                         NaN
#> 17                <NA>                                         NaN
#> 18                <NA>                                         NaN
#> 19                <NA>                                         NaN
#> 20                <NA>                                         NaN
#> 21                <NA>                                         NaN
#> 22                <NA>                                         NaN
#> 23                <NA>                                         NaN
#> 24                <NA>                                         NaN
#> 25                <NA>                                         NaN
#> 26                <NA>                                         NaN
#> 27                <NA>                                         NaN
#> 28                <NA>                                         NaN
#> 29                <NA>                                         NaN
#> 30                <NA>                                         NaN
#> 31                <NA>                                         NaN
#> 32                <NA>                                         NaN
#> 33                <NA>                                         NaN
#> 34                <NA>                                         NaN
#> 35                <NA>                                         NaN
#> 36                <NA>                                         NaN
#> 37                <NA>                                         NaN
#> 38                <NA>                                         NaN
#> 39                <NA>                                         NaN
#> 40                <NA>                                         NaN
#> 41                <NA>                                         NaN
#> 42                <NA>                                         NaN
#> 43                <NA>                                         NaN
#> 44                <NA>                                         NaN
#> 45                <NA>                                         NaN
#> 46                <NA>                                         NaN
#> 47                <NA>                                         NaN
#> 48                <NA>                                         NaN
#> 49                <NA>                                         NaN
#> 50                <NA>                                         NaN
#> 51                <NA>                                         NaN
#> 52                <NA>                                         NaN
#> 53                <NA>                                         NaN
#> 54                <NA>                                         NaN
#> 55                <NA>                                         NaN
#> 56                <NA>                                         NaN
#> 57                <NA>                                         NaN
#>                    initials  camid
#> 1                     AB,CD 100001
#> 2                      <NA> 100002
#> 3                      <NA> 100003
#> 4                      <NA> 100908
#> 5                      <NA> 100909
#> 6                      <NA> 101479
#> 7                     AB,CD 101480
#> 8                        EF 101483
#> 9                      <NA> 101692
#> 10                     <NA> 101694
#> 11                     <NA> 101697
#> 12                     <NA> 101698
#> 13                     <NA> 101700
#> 14                     <NA> 101703
#> 15                     <NA> 101706
#> 16                     <NA> 101709
#> 17                     <NA> 101712
#> 18                     <NA> 101715
#> 19                     <NA> 101718
#> 20                     <NA> 101721
#> 21                     <NA> 101724
#> 22                     <NA> 101727
#> 23                     <NA> 101730
#> 24                     <NA> 101733
#> 25                     <NA> 101736
#> 26                     <NA> 101739
#> 27                     <NA> 101742
#> 28                     <NA> 101745
#> 29                     <NA> 101748
#> 30                     <NA> 101751
#> 31 zztest-allow-new-options 101752
#> 32                     <NA> 101753
#> 33 zztest-allow-new-options 101754
#> 34                     <NA> 101755
#> 35                     <NA> 101758
#> 36                     <NA> 101761
#> 37                     <NA> 101764
#> 38                     <NA> 101767
#> 39                     <NA> 101770
#> 40                     <NA> 101773
#> 41                     <NA> 101776
#> 42                     <NA> 101779
#> 43                     <NA> 101782
#> 44                     <NA> 101785
#> 45                     <NA> 101788
#> 46                     <NA> 101790
#> 47                     <NA> 101794
#> 48                     <NA> 101797
#> 49                     <NA> 101800
#> 50                     <NA> 101803
#> 51                     <NA> 101806
#> 52                     <NA> 101807
#> 53                     <NA> 101812
#> 54                     <NA> 101815
#> 55                     <NA> 101818
#> 56                     <NA> 101821
#> 57                     <NA> 101822
# }
# \donttest{
flytable_query("SELECT person, fruit_name FROM testfruit WHERE person!='Bob'")
#>                 person  fruit_name
#> 1                Alice       apple
#> 2                Clara  clementine
#> 3  Frederick the Great        kiwi
#> 4  Frederick the Great        kiwi
#> 5  Frederick the Great        kiwi
#> 6  Frederick the Great        kiwi
#> 7  Frederick the Great        kiwi
#> 8  Frederick the Great        kiwi
#> 9  Frederick the Great        kiwi
#> 10 Frederick the Great        kiwi
#> 11 Frederick the Great        kiwi
#> 12 Frederick the Great        kiwi
#> 13 Frederick the Great        kiwi
#> 14 Frederick the Great        kiwi
#> 15 Frederick the Great        kiwi
#> 16 Frederick the Great        kiwi
#> 17 Frederick the Great        kiwi
#> 18 Frederick the Great        kiwi
#> 19 Frederick the Great        kiwi
#> 20 Frederick the Great        kiwi
#> 21 Frederick the Great        kiwi
#> 22 Frederick the Great        kiwi
#> 23 Frederick the Great        kiwi
#> 24 Frederick the Great        kiwi
#> 25 Frederick the Great        kiwi
#> 26 Frederick the Great        kiwi
#> 27 Frederick the Great        kiwi
#> 28 Frederick the Great        kiwi
#> 29     Delta Sync Test dragonfruit
#> 30   Multi Select Test   starfruit
#> 31     Delta Sync Test dragonfruit
#> 32   Multi Select Test   starfruit
#> 33 Frederick the Great        kiwi
#> 34 Frederick the Great        kiwi
#> 35 Frederick the Great        kiwi
#> 36 Frederick the Great        kiwi
#> 37 Frederick the Great        kiwi
#> 38 Frederick the Great        kiwi
#> 39 Frederick the Great        kiwi
#> 40 Frederick the Great        kiwi
#> 41 Frederick the Great        kiwi
#> 42 Frederick the Great        kiwi
#> 43 Frederick the Great        kiwi
#> 44 Frederick the Great        kiwi
#> 45 Frederick the Great        kiwi
#> 46 Frederick the Great        kiwi
#> 47 Frederick the Great        kiwi
#> 48 Frederick the Great        kiwi
#> 49 Frederick the Great        kiwi
#> 50 Frederick the Great        kiwi
#> 51 Frederick the Great        kiwi
#> 52 Frederick the Great        kiwi
#> 53 Frederick the Great        kiwi
#> 54 Frederick the Great        kiwi
#> 55 Frederick the Great        kiwi
#> 56 Frederick the Great        kiwi
# }
if (FALSE) { # \dontrun{
flytable_query(paste("SELECT root_id, supervoxel_id FROM info limit 5"))
} # }