Query the FlyWire CAVE annotation system

Query the FlyWire CAVE annotation system

flywire_cave_query(
  table,
  datastack_name = getOption("fafbseg.cave.datastack_name", "flywire_fafb_production"),
  live = TRUE,
  ...
)

Arguments

table
The name of the table to query

datastack_name
defaults to "flywire_fafb_production". See https://global.daf-apis.com/info/ for other options.

live
Whether to use live query mode, which updates any root ids to their current value.

...
Additional arguments to the query method. See examples and details.

Value

A tibble. Note that xyzmatrix can be used on single columns containing XYZ locations.

Details

CAVE (Connectome Annotation Versioning Engine) provides a shared infrastructure for a number of connectomics projects involving Sebastian Seung's group at Princeton and collaborators at the Allen Institute. There is both a backend system running on their servers and a Python client for end users.

You can find out more at https://caveclient.readthedocs.io/ as well as looking at the Python notebooks on the github repo https://github.com/seung-lab/CAVEclient.

The annotation system shares authentication infrastructure with the rest of the FlyWire API (see flywire_set_token).

See also

Examples

# \donttest{
# note use of limit to restrict the number of rows
n10=flywire_cave_query(table = 'nuclei_v1', limit=10)
head(as.data.frame(n10))

#>        id valid  pt_supervoxel_id         pt_root_id    volume
#> 1 7393349     t 82827379285852979 720575940630194247 26.141245
#> 2 7416439     t 82827998029690530 720575940626850921 11.523400
#> 3 7415038     t 83038623024880664 720575940626046919 32.895959
#> 4 7415013     t 83038760463398837 720575940632921242 53.711176
#> 5 7415848     t 83038554439606237 720575940612954553  9.280717
#> 6 7415851     t 83038691811464073 720575940603963070 19.018342
#>              pt_position      bb_start_position        bb_end_position
#> 1  709888, 227744, 57160  708032, 226144, 54760  711904, 229440, 59280
#> 2 710592, 263392, 129800 708800, 262112, 128200 712064, 264832, 131080
#> 3  722528, 234656, 77000  720768, 232896, 74480  724224, 236576, 79360
#> 4  722912, 244032, 65200  720480, 242144, 62480  725600, 246208, 67760
#> 5 721984, 229792, 119560 720832, 228320, 118240 723232, 231264, 120840
#> 6  722432, 239520, 98440  720864, 237888, 96200 724256, 241056, 100760

# }
if (FALSE) {
nuclei_v1=flywire_cave_query(table = 'nuclei_v1')
points3d(xyzmatrix(nuclei_v1$pt_position))

library(elmr)
# calculate signed distance to FAFB surface
# NB this surface is not a perfect fit in the optic lobes
nuclei_v1$d=pointsinside(xyzmatrix(nuclei_v1$pt_position), FAFB.surf,
  rval = 'dist')
points3d(xyzmatrix(nuclei_v1$pt_position),
  col=matlab::jet.colors(20)[cut(nuclei_v1$d,20)])
plot3d(FAFB)
}