Summarise edits for one or more FlyWire segment ids

Summarise edits for one or more FlyWire segment ids

flywire_change_log(
  x,
  root_ids = FALSE,
  filtered = TRUE,
  tz = "UTC",
  OmitFailures = TRUE,
  ...
)

Arguments

x
One or more flywire segment ids (character vector strongly recommended but any format accepted by ngl_segments will work).

root_ids
Whether to look up the root ids before/after each change. Default FALSE. NB this is quite resource intensive so please do not flood the server with requests of this sort.

filtered
Whether to filter out edits unlikely to relate to the current state of the neuron (default TRUE, see details).

tz
Time zone for edit timestamps. Defaults to "UTC" i.e. Universal Time, Coordinated. Set to "" for your current timezone. See as.POSIXct for more details.

OmitFailures
Whether to omit neurons for which there is an API timeout or error. The default value (TRUE) will skip over errors, while NA) will result in a hard stop on error. See nlapply for more details.

...
Additional arguments passed to flywire_fetch

Value

A data frame with values itemize

In addition when filtered=FALSE, in_neuron is_relevant

In addition when root_ids=TRUE, before_root_ids after_root_ids, as space separated strings.

Details

When filtered=TRUE the following are removed

edits that are outside of the given root id - meaning these edits are in the history of the cell but the affected part of the arbour was since cut off.

edits that were since undone by an inverse operation (often temporary splits made while reviewing a neuron.)

Examples

# \donttest{
flywire_change_log("720575940619010932")

#>   operation_id           timestamp user_id is_merge        user_name
#> 0       194549 2020-09-03 19:33:23     397    FALSE Marlon Blanquart
#> 1       194550 2020-09-03 19:37:26     397     TRUE Marlon Blanquart
#> 2       194551 2020-09-03 19:38:12     397    FALSE Marlon Blanquart
#> 3       194552 2020-09-03 19:40:33     397     TRUE Marlon Blanquart
#> 4       194582 2020-09-03 20:05:41     397     TRUE Marlon Blanquart
#> 5       194584 2020-09-03 20:06:51     397    FALSE Marlon Blanquart

flywire_change_log("720575940619010932", root_ids = TRUE)

#>   operation_id           timestamp user_id is_merge
#> 0       194549 2020-09-03 19:33:23     397    FALSE
#> 1       194550 2020-09-03 19:37:26     397     TRUE
#> 2       194551 2020-09-03 19:38:12     397    FALSE
#> 3       194552 2020-09-03 19:40:33     397     TRUE
#> 4       194582 2020-09-03 20:05:41     397     TRUE
#> 5       194584 2020-09-03 20:06:51     397    FALSE
#>                         before_root_ids                        after_root_ids
#> 0                    720575940615250973 720575940611598545 720575940625952497
#> 1 720575940622305363 720575940625952497                    720575940603341158
#> 2                    720575940603341158 720575940618539430 720575940620629528
#> 3 720575940620629528 720575940637849454                    720575940625315922
#> 4 720575940625315922 720575940635457646                    720575940610162858
#> 5                    720575940610162858 720575940618698480 720575940619010932
#>          user_name
#> 0 Marlon Blanquart
#> 1 Marlon Blanquart
#> 2 Marlon Blanquart
#> 3 Marlon Blanquart
#> 4 Marlon Blanquart
#> 5 Marlon Blanquart

flywire_change_log("720575940619010932", filtered = FALSE)

#>   operation_id           timestamp user_id is_merge in_neuron is_relevant
#> 0       194549 2020-09-03 19:33:23     397    FALSE      TRUE        TRUE
#> 1       194550 2020-09-03 19:37:26     397     TRUE      TRUE        TRUE
#> 2       194551 2020-09-03 19:38:12     397    FALSE      TRUE        TRUE
#> 3       194552 2020-09-03 19:40:33     397     TRUE      TRUE        TRUE
#> 4       194582 2020-09-03 20:05:41     397     TRUE      TRUE        TRUE
#> 5       194584 2020-09-03 20:06:51     397    FALSE      TRUE        TRUE
#>          user_name
#> 0 Marlon Blanquart
#> 1 Marlon Blanquart
#> 2 Marlon Blanquart
#> 3 Marlon Blanquart
#> 4 Marlon Blanquart
#> 5 Marlon Blanquart

# with a flywire URL
u="https://ngl.flywire.ai/?json_url=https://globalv1.flywire-daf.com/nglstate/5409525645443072"
flywire_change_log(u)

#>                     id operation_id           timestamp user_id is_merge
#> 1   720575940621039145         1710 2019-08-03 23:17:28       4    FALSE
#> 2   720575940621039145         1715 2019-08-03 23:21:22       4    FALSE
#> 3   720575940621039145         1716 2019-08-03 23:22:16       4    FALSE
#> 4   720575940621039145         1735 2019-08-03 23:31:40       4    FALSE
#> 5   720575940621039145         1736 2019-08-03 23:31:57       4    FALSE
#> 6   720575940621039145         1737 2019-08-03 23:32:08       4    FALSE
#> 7   720575940621039145         1738 2019-08-03 23:32:20       4    FALSE
#> 8   720575940621039145         1739 2019-08-03 23:32:30       4    FALSE
#> 9   720575940621039145         1742 2019-08-03 23:33:22       4    FALSE
#> 10  720575940621039145         1744 2019-08-03 23:33:56       4    FALSE
#> 11  720575940621039145         1748 2019-08-03 23:34:34       4    FALSE
#> 12  720575940621039145         1750 2019-08-03 23:34:48       4    FALSE
#> 13  720575940621039145         1752 2019-08-03 23:35:15       4    FALSE
#> 14  720575940621039145         1753 2019-08-03 23:35:31       4    FALSE
#> 15  720575940621039145         1755 2019-08-03 23:35:50       4    FALSE
#> 16  720575940621039145         1757 2019-08-03 23:36:12       4    FALSE
#> 17  720575940621039145         1759 2019-08-03 23:36:26       4    FALSE
#> 18  720575940621039145         1761 2019-08-03 23:36:45       4    FALSE
#> 19  720575940621039145         1763 2019-08-03 23:37:06       4    FALSE
#> 20  720575940621039145         1766 2019-08-03 23:37:33       4    FALSE
#> 21  720575940621039145         1768 2019-08-03 23:37:55       4    FALSE
#> 22  720575940621039145         1770 2019-08-03 23:38:14       4    FALSE
#> 23  720575940621039145         1771 2019-08-03 23:38:27       4    FALSE
#> 24  720575940621039145         1776 2019-08-03 23:39:39       4    FALSE
#> 25  720575940621039145         1778 2019-08-03 23:42:10       4    FALSE
#> 26  720575940621039145         1781 2019-08-03 23:47:50       4    FALSE
#> 27  720575940621039145         1782 2019-08-03 23:49:56       4     TRUE
#> 28  720575940621039145         1784 2019-08-03 23:55:56       4    FALSE
#> 29  720575940621039145         1785 2019-08-03 23:56:10       4    FALSE
#> 30  720575940621039145         1786 2019-08-03 23:56:36       4    FALSE
#> 31  720575940621039145         1787 2019-08-03 23:56:48       4    FALSE
#> 32  720575940621039145         1788 2019-08-03 23:57:10       4    FALSE
#> 33  720575940621039145         1896 2019-08-04 02:42:31       4    FALSE
#> 34  720575940621039145         2231 2019-08-08 16:45:15       4    FALSE
#> 35  720575940621039145         2300 2019-08-08 19:39:12       4    FALSE
#> 36  720575940621039145         2346 2019-08-09 21:51:43       4    FALSE
#> 37  720575940621039145         2377 2019-08-12 18:19:12       4    FALSE
#> 38  720575940621039145         2423 2019-08-14 17:16:15       4    FALSE
#> 39  720575940621039145         2742 2019-08-16 17:14:54       4     TRUE
#> 40  720575940621039145         8042 2019-09-05 23:10:59       4    FALSE
#> 41  720575940621039145         8046 2019-09-05 23:22:12       4    FALSE
#> 42  720575940621039145         8047 2019-09-05 23:26:27       4    FALSE
#> 43  720575940621039145         8049 2019-09-05 23:29:06       4    FALSE
#> 44  720575940621039145         8052 2019-09-05 23:32:15       4    FALSE
#> 45  720575940621039145         8085 2019-09-06 00:28:46       4    FALSE
#> 46  720575940621039145         8086 2019-09-06 00:30:04       4    FALSE
#> 47  720575940621039145         8091 2019-09-06 00:34:02       4     TRUE
#> 48  720575940621039145        11049 2019-09-16 11:57:40      13    FALSE
#> 49  720575940621039145        11194 2019-09-16 19:22:17      13    FALSE
#> 50  720575940621039145        20627 2019-10-18 18:04:25      16    FALSE
#> 51  720575940621039145        40816 2020-01-07 18:48:38      13    FALSE
#> 52  720575940621039145        40819 2020-01-07 18:50:06      13     TRUE
#> 53  720575940621039145        52005 2020-02-03 18:42:45      26     TRUE
#> 54  720575940621039145        52065 2020-02-03 19:42:24      26    FALSE
#> 55  720575940621039145        52135 2020-02-03 20:34:35      26     TRUE
#> 56  720575940621039145        52141 2020-02-03 20:35:32      26     TRUE
#> 57  720575940621039145        52142 2020-02-03 20:36:01      26     TRUE
#> 58  720575940621039145        52146 2020-02-03 20:40:29      26    FALSE
#> 59  720575940626877799         4729 2019-08-23 21:17:22      12     TRUE
#> 60  720575940626877799         6163 2019-08-28 20:48:48      12    FALSE
#> 61  720575940626877799        10877 2019-09-13 16:00:25      13    FALSE
#> 62  720575940626877799        10903 2019-09-13 16:52:47      13    FALSE
#> 63  720575940626877799        10904 2019-09-13 16:57:06      13     TRUE
#> 64  720575940626877799        10906 2019-09-13 16:58:52      13    FALSE
#> 65  720575940626877799        10907 2019-09-13 17:01:45      13    FALSE
#> 66  720575940626877799        10909 2019-09-13 17:11:08      13    FALSE
#> 67  720575940626877799        10912 2019-09-13 17:14:06      13    FALSE
#> 68  720575940626877799        10937 2019-09-13 17:37:41      13    FALSE
#> 69  720575940626877799        10939 2019-09-13 17:43:53      13     TRUE
#> 70  720575940626877799        10940 2019-09-13 17:44:28      13     TRUE
#> 71  720575940626877799        10947 2019-09-13 18:00:31      13    FALSE
#> 72  720575940626877799        10949 2019-09-13 18:04:35      13    FALSE
#> 73  720575940626877799        10952 2019-09-13 18:08:03      13    FALSE
#> 74  720575940626877799        10956 2019-09-13 18:12:46      13    FALSE
#> 75  720575940626877799        10967 2019-09-13 18:19:11      13    FALSE
#> 76  720575940626877799        10969 2019-09-13 18:20:48      13     TRUE
#> 77  720575940626877799        10971 2019-09-13 18:22:16      13    FALSE
#> 78  720575940626877799        10972 2019-09-13 18:24:20      13    FALSE
#> 79  720575940626877799        10981 2019-09-13 18:26:26      13     TRUE
#> 80  720575940626877799        11363 2019-09-17 14:53:32      17    FALSE
#> 81  720575940626877799        11414 2019-09-17 16:03:08      17    FALSE
#> 82  720575940626877799        33530 2019-12-13 19:16:56      84    FALSE
#> 83  720575940626877799        46345 2020-01-17 18:27:27      84    FALSE
#> 84  720575940626877799        46350 2020-01-17 18:28:49      84    FALSE
#> 85  720575940626877799        46351 2020-01-17 18:30:32      84    FALSE
#> 86  720575940626877799        46352 2020-01-17 18:33:23      84    FALSE
#> 87  720575940626877799       107453 2020-04-27 18:44:42     104    FALSE
#> 88  720575940626877799       107454 2020-04-27 18:44:46     104    FALSE
#> 89  720575940626877799       107455 2020-04-27 18:44:53     104    FALSE
#> 90  720575940626877799       107456 2020-04-27 18:45:00     104    FALSE
#> 91  720575940626877799       107460 2020-04-27 18:45:43     104     TRUE
#> 92  720575940626877799       107461 2020-04-27 18:45:46     104     TRUE
#> 93  720575940626877799       107466 2020-04-27 18:46:08     104     TRUE
#> 94  720575940626877799       107467 2020-04-27 18:46:14     104     TRUE
#> 95  720575940626877799       107468 2020-04-27 18:48:06     104    FALSE
#> 96  720575940626877799       107480 2020-04-27 19:03:54     104    FALSE
#> 97  720575940626877799       107482 2020-04-27 19:07:39     104     TRUE
#> 98  720575940626877799       107487 2020-04-27 19:17:14     104    FALSE
#> 99  720575940626877799       107494 2020-04-27 19:30:34     104    FALSE
#> 100 720575940626877799       107495 2020-04-27 19:30:47     104     TRUE
#> 101 720575940626877799       107557 2020-04-27 20:40:33     104    FALSE
#> 102 720575940626877799       107589 2020-04-27 21:10:52     104    FALSE
#> 103 720575940626877799       107590 2020-04-27 21:11:03     104    FALSE
#> 104 720575940626877799       107592 2020-04-27 21:11:47     104    FALSE
#> 105 720575940626877799       107603 2020-04-27 21:20:09     104     TRUE
#> 106 720575940626877799       107604 2020-04-27 21:20:13     104     TRUE
#> 107 720575940626877799       107635 2020-04-27 21:31:33     104    FALSE
#> 108 720575940626877799       107650 2020-04-27 22:15:31     104    FALSE
#> 109 720575940626877799       107651 2020-04-27 22:15:46     104    FALSE
#> 110 720575940626877799       107652 2020-04-27 22:15:51     104    FALSE
#> 111 720575940626877799       107653 2020-04-27 22:15:57     104    FALSE
#> 112 720575940626877799       107656 2020-04-27 22:27:41     104    FALSE
#> 113 720575940626877799       107657 2020-04-27 22:29:06     104    FALSE
#> 114 720575940626877799       107660 2020-04-27 22:31:44     104     TRUE
#> 115 720575940626877799       107661 2020-04-27 22:37:56     104    FALSE
#> 116 720575940626877799       107667 2020-04-27 22:57:59     104    FALSE
#> 117 720575940626877799       107668 2020-04-27 22:58:05     104    FALSE
#> 118 720575940626877799       107699 2020-04-27 23:36:50     104    FALSE
#> 119 720575940626877799       107710 2020-04-27 23:42:21     104    FALSE
#> 120 720575940626877799       107726 2020-04-28 00:43:21     104    FALSE
#> 121 720575940626877799       107729 2020-04-28 00:48:13     104    FALSE
#> 122 720575940626877799       107735 2020-04-28 01:00:38     104    FALSE
#> 123 720575940626877799       107737 2020-04-28 01:09:32     104    FALSE
#> 124 720575940626877799       107738 2020-04-28 01:11:31     104    FALSE
#> 125 720575940626877799       107739 2020-04-28 01:12:12     104    FALSE
#> 126 720575940626877799       107742 2020-04-28 01:21:57     104    FALSE
#> 127 720575940626877799       107766 2020-04-28 01:47:17     104    FALSE
#> 128 720575940626877799       107767 2020-04-28 01:47:22     104    FALSE
#> 129 720575940626877799       107783 2020-04-28 02:06:45     104    FALSE
#> 130 720575940626877799       107797 2020-04-28 02:55:10     104    FALSE
#> 131 720575940626877799       107798 2020-04-28 02:59:17     104    FALSE
#> 132 720575940626877799       107799 2020-04-28 03:01:32     104    FALSE
#> 133 720575940626877799       107800 2020-04-28 03:02:02     104    FALSE
#> 134 720575940626877799       107801 2020-04-28 03:03:27     104    FALSE
#> 135 720575940626877799       107803 2020-04-28 03:08:37     104    FALSE
#> 136 720575940626877799       107806 2020-04-28 03:22:03     104     TRUE
#> 137 720575940626877799       107807 2020-04-28 03:22:07     104     TRUE
#> 138 720575940626877799       107808 2020-04-28 03:22:49     104     TRUE
#> 139 720575940626877799       107809 2020-04-28 03:22:52     104     TRUE
#> 140 720575940626877799       107810 2020-04-28 03:22:56     104     TRUE
#> 141 720575940626877799       107948 2020-04-28 18:23:53     104    FALSE
#> 142 720575940626877799       107949 2020-04-28 18:24:11     104     TRUE
#> 143 720575940626877799       107950 2020-04-28 18:34:57     104    FALSE
#> 144 720575940626877799       107966 2020-04-28 19:01:09     104    FALSE
#> 145 720575940626877799       107981 2020-04-28 19:07:41     104    FALSE
#> 146 720575940626877799       108009 2020-04-28 19:33:43     104    FALSE
#> 147 720575940626877799       108010 2020-04-28 19:33:53     104    FALSE
#> 148 720575940626877799       108012 2020-04-28 19:34:00     104    FALSE
#> 149 720575940626877799       108013 2020-04-28 19:34:03     104    FALSE
#> 150 720575940626877799       108014 2020-04-28 19:34:06     104    FALSE
#> 151 720575940626877799       108034 2020-04-28 20:22:10     104    FALSE
#> 152 720575940626877799       108036 2020-04-28 20:31:23     104    FALSE
#> 153 720575940626877799       108037 2020-04-28 20:32:38     104     TRUE
#> 154 720575940626877799       108042 2020-04-28 20:40:41     104    FALSE
#> 155 720575940626877799       108046 2020-04-28 20:46:59     104    FALSE
#> 156 720575940626877799       108054 2020-04-28 20:56:18     104    FALSE
#> 157 720575940626877799       108056 2020-04-28 20:56:26     104    FALSE
#> 158 720575940626877799       108057 2020-04-28 20:59:15     104     TRUE
#> 159 720575940626877799       108058 2020-04-28 20:59:18     104     TRUE
#> 160 720575940626877799       108059 2020-04-28 21:01:17     104    FALSE
#> 161 720575940626877799       108060 2020-04-28 21:01:32     104     TRUE
#> 162 720575940626877799       108070 2020-04-28 21:26:49     104    FALSE
#> 163 720575940626877799       108071 2020-04-28 21:28:23     104     TRUE
#> 164 720575940626877799       108072 2020-04-28 21:28:24     104     TRUE
#> 165 720575940626877799       108073 2020-04-28 21:28:28     104     TRUE
#> 166 720575940626877799       108094 2020-04-28 21:58:38     104    FALSE
#> 167 720575940626877799       108095 2020-04-28 21:59:26     104    FALSE
#> 168 720575940626877799       108101 2020-04-29 00:47:33     104    FALSE
#> 169 720575940626877799       108102 2020-04-29 00:53:33     104    FALSE
#> 170 720575940626877799       108107 2020-04-29 01:16:53     104    FALSE
#> 171 720575940626877799       108109 2020-04-29 01:17:03     104    FALSE
#> 172 720575940626877799       108122 2020-04-29 01:32:15     104    FALSE
#> 173 720575940626877799       108128 2020-04-29 01:37:54     104    FALSE
#> 174 720575940626877799       108129 2020-04-29 01:38:22     104     TRUE
#> 175 720575940626877799       108130 2020-04-29 01:44:41     104    FALSE
#> 176 720575940626877799       108132 2020-04-29 01:57:21     104    FALSE
#> 177 720575940626877799       108133 2020-04-29 02:27:13     104    FALSE
#> 178 720575940626877799       108134 2020-04-29 02:36:06     104    FALSE
#> 179 720575940626877799       108135 2020-04-29 03:23:50     104    FALSE
#> 180 720575940626877799       108136 2020-04-29 03:28:56     104    FALSE
#> 181 720575940626877799       108137 2020-04-29 03:32:06     104    FALSE
#> 182 720575940626877799       108138 2020-04-29 03:32:24     104     TRUE
#> 183 720575940626877799       108139 2020-04-29 03:33:11     104    FALSE
#> 184 720575940626877799       108140 2020-04-29 03:33:23     104    FALSE
#> 185 720575940626877799       108259 2020-04-29 17:12:45     104    FALSE
#> 186 720575940626877799       108260 2020-04-29 17:12:49     104    FALSE
#> 187 720575940626877799       108263 2020-04-29 17:13:05     104    FALSE
#> 188 720575940626877799       108264 2020-04-29 17:13:12     104    FALSE
#> 189 720575940626877799       108266 2020-04-29 17:13:32     104    FALSE
#> 190 720575940626877799       108270 2020-04-29 17:14:03     104     TRUE
#> 191 720575940626877799       108273 2020-04-29 17:14:23     104     TRUE
#> 192 720575940626877799       108274 2020-04-29 17:14:25     104     TRUE
#> 193 720575940626877799       108275 2020-04-29 17:14:35     104     TRUE
#> 194 720575940626877799       108276 2020-04-29 17:14:44     104    FALSE
#> 195 720575940626877799       108277 2020-04-29 17:15:58     104     TRUE
#> 196 720575940626877799       108278 2020-04-29 17:16:15     104     TRUE
#> 197 720575940626877799       108279 2020-04-29 17:16:20     104     TRUE
#> 198 720575940626877799       108280 2020-04-29 17:16:33     104     TRUE
#> 199 720575940626877799       108281 2020-04-29 17:16:39     104     TRUE
#> 200 720575940626877799       108282 2020-04-29 17:16:43     104     TRUE
#> 201 720575940626877799       108283 2020-04-29 17:16:59     104     TRUE
#> 202 720575940626877799       108284 2020-04-29 17:17:25     104     TRUE
#> 203 720575940626877799       108285 2020-04-29 17:17:36     104     TRUE
#> 204 720575940626877799       108286 2020-04-29 17:17:44     104     TRUE
#> 205 720575940626877799       108287 2020-04-29 17:18:38     104     TRUE
#> 206 720575940626877799       108288 2020-04-29 17:18:50     104     TRUE
#> 207 720575940626877799       108289 2020-04-29 17:19:04     104     TRUE
#> 208 720575940626877799       108290 2020-04-29 17:19:16     104     TRUE
#> 209 720575940626877799       108291 2020-04-29 17:19:26     104     TRUE
#> 210 720575940626877799       108292 2020-04-29 17:21:44     104     TRUE
#> 211 720575940626877799       108293 2020-04-29 17:21:48     104     TRUE
#> 212 720575940626877799       108294 2020-04-29 17:22:01     104     TRUE
#> 213 720575940626877799       108295 2020-04-29 17:22:04     104     TRUE
#> 214 720575940626877799       108296 2020-04-29 17:22:40     104     TRUE
#> 215 720575940626877799       108298 2020-04-29 17:23:02     104     TRUE
#> 216 720575940626877799       108299 2020-04-29 17:23:05     104     TRUE
#> 217 720575940626877799       108301 2020-04-29 17:23:23     104     TRUE
#> 218 720575940626877799       108302 2020-04-29 17:23:38     104     TRUE
#> 219 720575940626877799       108303 2020-04-29 17:23:50     104     TRUE
#> 220 720575940626877799       108304 2020-04-29 17:23:53     104     TRUE
#> 221 720575940626877799       108305 2020-04-29 17:24:02     104     TRUE
#> 222 720575940626877799       108307 2020-04-29 17:24:23     104     TRUE
#> 223 720575940626877799       108308 2020-04-29 17:24:34     104     TRUE
#> 224 720575940626877799       108310 2020-04-29 17:24:47     104     TRUE
#> 225 720575940626877799       108312 2020-04-29 17:25:02     104    FALSE
#> 226 720575940626877799       108313 2020-04-29 17:25:29     104     TRUE
#> 227 720575940626877799       108314 2020-04-29 17:25:47     104     TRUE
#> 228 720575940626877799       108315 2020-04-29 17:25:54     104     TRUE
#> 229 720575940626877799       108316 2020-04-29 17:26:06     104     TRUE
#> 230 720575940626877799       108317 2020-04-29 17:26:21     104     TRUE
#> 231 720575940626877799       108318 2020-04-29 17:33:43     104     TRUE
#> 232 720575940626877799       108319 2020-04-29 17:33:51     104     TRUE
#> 233 720575940626877799       108320 2020-04-29 17:33:59     104     TRUE
#> 234 720575940626877799       108321 2020-04-29 17:34:05     104     TRUE
#> 235 720575940626877799       108322 2020-04-29 17:34:12     104     TRUE
#> 236 720575940626877799       108323 2020-04-29 17:34:21     104     TRUE
#> 237 720575940626877799       108324 2020-04-29 17:34:26     104     TRUE
#> 238 720575940626877799       108325 2020-04-29 17:34:33     104     TRUE
#> 239 720575940626877799       108326 2020-04-29 17:34:38     104     TRUE
#> 240 720575940626877799       108327 2020-04-29 17:34:44     104     TRUE
#> 241 720575940626877799       108328 2020-04-29 17:34:48     104     TRUE
#> 242 720575940626877799       108329 2020-04-29 17:34:58     104     TRUE
#> 243 720575940626877799       108331 2020-04-29 17:35:13     104     TRUE
#> 244 720575940626877799       108332 2020-04-29 17:35:19     104     TRUE
#> 245 720575940626877799       108333 2020-04-29 17:35:25     104     TRUE
#> 246 720575940626877799       108334 2020-04-29 17:35:32     104     TRUE
#> 247 720575940626877799       108335 2020-04-29 17:35:41     104     TRUE
#> 248 720575940626877799       108336 2020-04-29 17:35:46     104     TRUE
#> 249 720575940626877799       108337 2020-04-29 17:36:01     104     TRUE
#> 250 720575940626877799       108338 2020-04-29 17:37:18     104     TRUE
#> 251 720575940626877799       108479 2020-04-29 20:48:54     104    FALSE
#> 252 720575940626877799       111739 2020-05-06 16:16:58      84    FALSE
#> 253 720575940626877799       111740 2020-05-06 16:17:10      84    FALSE
#> 254 720575940626877799       111766 2020-05-06 16:48:07      84    FALSE
#> 255 720575940626877799       111819 2020-05-06 17:05:02      84    FALSE
#> 256 720575940626877799       111824 2020-05-06 17:13:58      84    FALSE
#> 257 720575940626877799       111836 2020-05-06 17:22:18      84    FALSE
#> 258 720575940626877799       111837 2020-05-06 17:22:28      84    FALSE
#> 259 720575940626877799       111842 2020-05-06 17:26:58      84    FALSE
#> 260 720575940626877799       111845 2020-05-06 17:29:53      84    FALSE
#> 261 720575940626877799       111849 2020-05-06 17:32:11      84    FALSE
#> 262 720575940626877799       112050 2020-05-06 18:47:44      84    FALSE
#> 263 720575940626877799       112095 2020-05-06 19:17:35      84    FALSE
#> 264 720575940626877799       112118 2020-05-06 19:50:26      84    FALSE
#> 265 720575940626877799       112136 2020-05-06 20:02:49      84    FALSE
#> 266 720575940626877799       112137 2020-05-06 20:03:05      84    FALSE
#> 267 720575940626877799       112140 2020-05-06 20:07:48      84    FALSE
#> 268 720575940626877799       112145 2020-05-06 20:17:07      84    FALSE
#> 269 720575940626877799       112176 2020-05-06 20:42:38      84    FALSE
#> 270 720575940626877799       112207 2020-05-06 21:10:54      84    FALSE
#> 271 720575940626877799       112208 2020-05-06 21:14:47      84    FALSE
#> 272 720575940626877799       112210 2020-05-06 21:15:08      84    FALSE
#> 273 720575940626877799       112211 2020-05-06 21:17:43      84    FALSE
#> 274 720575940626877799       112225 2020-05-06 21:41:35      84    FALSE
#> 275 720575940626877799       112226 2020-05-06 21:41:48      84    FALSE
#> 276 720575940626877799       112667 2020-05-07 14:11:01      84    FALSE
#> 277 720575940626877799       112682 2020-05-07 14:37:31      84     TRUE
#> 278 720575940626877799       112683 2020-05-07 14:37:34      84     TRUE
#> 279 720575940626877799       112685 2020-05-07 14:37:51      84     TRUE
#> 280 720575940626877799       112686 2020-05-07 14:38:30      84     TRUE
#> 281 720575940626877799       112687 2020-05-07 14:38:34      84     TRUE
#> 282 720575940626877799       112688 2020-05-07 14:38:48      84     TRUE
#> 283 720575940626877799       112689 2020-05-07 14:39:03      84     TRUE
#> 284 720575940626877799       112690 2020-05-07 14:39:05      84     TRUE
#> 285 720575940626877799       112691 2020-05-07 14:39:15      84     TRUE
#> 286 720575940626877799       112692 2020-05-07 14:39:29      84     TRUE
#> 287 720575940626877799       112695 2020-05-07 14:40:02      84     TRUE
#> 288 720575940626877799       112696 2020-05-07 14:40:16      84     TRUE
#>           user_name
#> 1   Claire McKellar
#> 2   Claire McKellar
#> 3   Claire McKellar
#> 4   Claire McKellar
#> 5   Claire McKellar
#> 6   Claire McKellar
#> 7   Claire McKellar
#> 8   Claire McKellar
#> 9   Claire McKellar
#> 10  Claire McKellar
#> 11  Claire McKellar
#> 12  Claire McKellar
#> 13  Claire McKellar
#> 14  Claire McKellar
#> 15  Claire McKellar
#> 16  Claire McKellar
#> 17  Claire McKellar
#> 18  Claire McKellar
#> 19  Claire McKellar
#> 20  Claire McKellar
#> 21  Claire McKellar
#> 22  Claire McKellar
#> 23  Claire McKellar
#> 24  Claire McKellar
#> 25  Claire McKellar
#> 26  Claire McKellar
#> 27  Claire McKellar
#> 28  Claire McKellar
#> 29  Claire McKellar
#> 30  Claire McKellar
#> 31  Claire McKellar
#> 32  Claire McKellar
#> 33  Claire McKellar
#> 34  Claire McKellar
#> 35  Claire McKellar
#> 36  Claire McKellar
#> 37  Claire McKellar
#> 38  Claire McKellar
#> 39  Claire McKellar
#> 40  Claire McKellar
#> 41  Claire McKellar
#> 42  Claire McKellar
#> 43  Claire McKellar
#> 44  Claire McKellar
#> 45  Claire McKellar
#> 46  Claire McKellar
#> 47  Claire McKellar
#> 48        Jay Gager
#> 49        Jay Gager
#> 50    Sandeep Kumar
#> 51        Jay Gager
#> 52        Jay Gager
#> 53   Sarah Morejohn
#> 54   Sarah Morejohn
#> 55   Sarah Morejohn
#> 56   Sarah Morejohn
#> 57   Sarah Morejohn
#> 58   Sarah Morejohn
#> 59    Christa Baker
#> 60    Christa Baker
#> 61        Jay Gager
#> 62        Jay Gager
#> 63        Jay Gager
#> 64        Jay Gager
#> 65        Jay Gager
#> 66        Jay Gager
#> 67        Jay Gager
#> 68        Jay Gager
#> 69        Jay Gager
#> 70        Jay Gager
#> 71        Jay Gager
#> 72        Jay Gager
#> 73        Jay Gager
#> 74        Jay Gager
#> 75        Jay Gager
#> 76        Jay Gager
#> 77        Jay Gager
#> 78        Jay Gager
#> 79        Jay Gager
#> 80   Austin T Burke
#> 81   Austin T Burke
#> 82     Merlin Moore
#> 83     Merlin Moore
#> 84     Merlin Moore
#> 85     Merlin Moore
#> 86     Merlin Moore
#> 87      Ryan Willie
#> 88      Ryan Willie
#> 89      Ryan Willie
#> 90      Ryan Willie
#> 91      Ryan Willie
#> 92      Ryan Willie
#> 93      Ryan Willie
#> 94      Ryan Willie
#> 95      Ryan Willie
#> 96      Ryan Willie
#> 97      Ryan Willie
#> 98      Ryan Willie
#> 99      Ryan Willie
#> 100     Ryan Willie
#> 101     Ryan Willie
#> 102     Ryan Willie
#> 103     Ryan Willie
#> 104     Ryan Willie
#> 105     Ryan Willie
#> 106     Ryan Willie
#> 107     Ryan Willie
#> 108     Ryan Willie
#> 109     Ryan Willie
#> 110     Ryan Willie
#> 111     Ryan Willie
#> 112     Ryan Willie
#> 113     Ryan Willie
#> 114     Ryan Willie
#> 115     Ryan Willie
#> 116     Ryan Willie
#> 117     Ryan Willie
#> 118     Ryan Willie
#> 119     Ryan Willie
#> 120     Ryan Willie
#> 121     Ryan Willie
#> 122     Ryan Willie
#> 123     Ryan Willie
#> 124     Ryan Willie
#> 125     Ryan Willie
#> 126     Ryan Willie
#> 127     Ryan Willie
#> 128     Ryan Willie
#> 129     Ryan Willie
#> 130     Ryan Willie
#> 131     Ryan Willie
#> 132     Ryan Willie
#> 133     Ryan Willie
#> 134     Ryan Willie
#> 135     Ryan Willie
#> 136     Ryan Willie
#> 137     Ryan Willie
#> 138     Ryan Willie
#> 139     Ryan Willie
#> 140     Ryan Willie
#> 141     Ryan Willie
#> 142     Ryan Willie
#> 143     Ryan Willie
#> 144     Ryan Willie
#> 145     Ryan Willie
#> 146     Ryan Willie
#> 147     Ryan Willie
#> 148     Ryan Willie
#> 149     Ryan Willie
#> 150     Ryan Willie
#> 151     Ryan Willie
#> 152     Ryan Willie
#> 153     Ryan Willie
#> 154     Ryan Willie
#> 155     Ryan Willie
#> 156     Ryan Willie
#> 157     Ryan Willie
#> 158     Ryan Willie
#> 159     Ryan Willie
#> 160     Ryan Willie
#> 161     Ryan Willie
#> 162     Ryan Willie
#> 163     Ryan Willie
#> 164     Ryan Willie
#> 165     Ryan Willie
#> 166     Ryan Willie
#> 167     Ryan Willie
#> 168     Ryan Willie
#> 169     Ryan Willie
#> 170     Ryan Willie
#> 171     Ryan Willie
#> 172     Ryan Willie
#> 173     Ryan Willie
#> 174     Ryan Willie
#> 175     Ryan Willie
#> 176     Ryan Willie
#> 177     Ryan Willie
#> 178     Ryan Willie
#> 179     Ryan Willie
#> 180     Ryan Willie
#> 181     Ryan Willie
#> 182     Ryan Willie
#> 183     Ryan Willie
#> 184     Ryan Willie
#> 185     Ryan Willie
#> 186     Ryan Willie
#> 187     Ryan Willie
#> 188     Ryan Willie
#> 189     Ryan Willie
#> 190     Ryan Willie
#> 191     Ryan Willie
#> 192     Ryan Willie
#> 193     Ryan Willie
#> 194     Ryan Willie
#> 195     Ryan Willie
#> 196     Ryan Willie
#> 197     Ryan Willie
#> 198     Ryan Willie
#> 199     Ryan Willie
#> 200     Ryan Willie
#> 201     Ryan Willie
#> 202     Ryan Willie
#> 203     Ryan Willie
#> 204     Ryan Willie
#> 205     Ryan Willie
#> 206     Ryan Willie
#> 207     Ryan Willie
#> 208     Ryan Willie
#> 209     Ryan Willie
#> 210     Ryan Willie
#> 211     Ryan Willie
#> 212     Ryan Willie
#> 213     Ryan Willie
#> 214     Ryan Willie
#> 215     Ryan Willie
#> 216     Ryan Willie
#> 217     Ryan Willie
#> 218     Ryan Willie
#> 219     Ryan Willie
#> 220     Ryan Willie
#> 221     Ryan Willie
#> 222     Ryan Willie
#> 223     Ryan Willie
#> 224     Ryan Willie
#> 225     Ryan Willie
#> 226     Ryan Willie
#> 227     Ryan Willie
#> 228     Ryan Willie
#> 229     Ryan Willie
#> 230     Ryan Willie
#> 231     Ryan Willie
#> 232     Ryan Willie
#> 233     Ryan Willie
#> 234     Ryan Willie
#> 235     Ryan Willie
#> 236     Ryan Willie
#> 237     Ryan Willie
#> 238     Ryan Willie
#> 239     Ryan Willie
#> 240     Ryan Willie
#> 241     Ryan Willie
#> 242     Ryan Willie
#> 243     Ryan Willie
#> 244     Ryan Willie
#> 245     Ryan Willie
#> 246     Ryan Willie
#> 247     Ryan Willie
#> 248     Ryan Willie
#> 249     Ryan Willie
#> 250     Ryan Willie
#> 251     Ryan Willie
#> 252    Merlin Moore
#> 253    Merlin Moore
#> 254    Merlin Moore
#> 255    Merlin Moore
#> 256    Merlin Moore
#> 257    Merlin Moore
#> 258    Merlin Moore
#> 259    Merlin Moore
#> 260    Merlin Moore
#> 261    Merlin Moore
#> 262    Merlin Moore
#> 263    Merlin Moore
#> 264    Merlin Moore
#> 265    Merlin Moore
#> 266    Merlin Moore
#> 267    Merlin Moore
#> 268    Merlin Moore
#> 269    Merlin Moore
#> 270    Merlin Moore
#> 271    Merlin Moore
#> 272    Merlin Moore
#> 273    Merlin Moore
#> 274    Merlin Moore
#> 275    Merlin Moore
#> 276    Merlin Moore
#> 277    Merlin Moore
#> 278    Merlin Moore
#> 279    Merlin Moore
#> 280    Merlin Moore
#> 281    Merlin Moore
#> 282    Merlin Moore
#> 283    Merlin Moore
#> 284    Merlin Moore
#> 285    Merlin Moore
#> 286    Merlin Moore
#> 287    Merlin Moore
#> 288    Merlin Moore

# }