Return a sample Neuroglancer scene URL for FlyWire dataset

Return a sample Neuroglancer scene URL for FlyWire dataset

Usage

flywire_scene(
  ids = NULL,
  annotations = NULL,
  open = FALSE,
  shorten = FALSE,
  segmentation = "flywire31",
  ...
)

Arguments

ids

A set of root ids to include in the scene. Also accepts a data.frame containing a column rootid, root_id, id or any form acceptable to ngl_segments including neuroglancer scene URLs.

annotations

data.frame or matrix of position and other information for annotation layers. See ngl_annotation_layers for details.

open

Whether to open the scene in your default browser

shorten

Not currently implemented

segmentation

Defaults to 'flywire31'. See choose_segmentation for other options.

...

Passed to ngl_annotation_layers

Value

A character vector containing a single Neuroglancer URL (invisibly when open=TRUE)

See also

Examples

if (FALSE) {
flywire_scene(open=T)
# top 20 partners of a neuron
flywire_scene(flywire_partner_summary("720575940621039145", partners='out')$partner[1:20], open=T)

# using the ability to query flytable for cell types
flywire_scene('DA2_lPN', open=TRUE)
flywire_scene('class:MBON_R', open=TRUE)
}