Shorten or expand neuroglancer URLs

flywire_shortenurl makes short URLs from long URLs or ngscene objects that you may have constructed in R.

flywire_expandurl expands shortened URLs into a full neuroglancer JSON scene specification. If the active segmentation (choose_segmentation) is a flywire segmentation then that is used to define the initial part of the output URL, otherwise the flywire31 segmentation is used.

flywire_shortenurl(x, include_base = TRUE, baseurl = NULL, cache = TRUE, ...)

flywire_expandurl(x, json.only = FALSE, cache = TRUE, ...)

Arguments

x
One or more neuroglancer URLs or (for flywire_expandurl) ngscene parsed scene description.

include_base
Whether to return a full URL that will open a neuroglancer session (the default) or only the component that defines the scene (which would display JSON in your browser).

baseurl
Optional URL defining the neuroglancer browser to use with shortened URLs.

cache
Whether to cache any calls to the flywire state server shortening or expanding URLs. Default is TRUE. NB this cache is only active for the current session.

...
Additional arguments passed to pbsapply (when multiple URLs to process) and then to ngl_encode_url (when generating a short URL for an ngscene list object) or to flywire_fetch when using flywire_expandurl.

json.only
Only return the JSON fragment rather than the neuroglancer URL

Value

A character vector containing one or more URLs.

Examples

if (FALSE) {
sc=ngl_blank_scene()
short=flywire_shortenurl(sc)
long=flywire_expandurl(short)
}
# \donttest{
flywire_expandurl("https://globalv1.flywire-daf.com/nglstate/5747205470158848")

#> [1] "https://ngl.flywire.ai/#!%7B%22layers%22:[%7B%22source%22:%22precomputed://gs://microns-seunglab/drosophila_v0/alignment/image_rechunked%22,%22type%22:%22image%22,%22blend%22:%22default%22,%22shaderControls%22:%7B%7D,%22name%22:%22Production-image%22%7D,%7B%22tab%22:%22graph%22,%22source%22:%22graphene://https://prodv1.flywire-daf.com/segmentation/1.0/fly_v31%22,%22type%22:%22segmentation_with_graph%22,%22selectedAlpha%22:0.62,%22colorSeed%22:1358768360,%22segments%22:[%22720575940619527173%22,%22720575940628130268%22,%22720575940630484179%22],%22hiddenSegments%22:[%22720575940634222734%22,%22720575940636364870%22],%22skeletonRendering%22:%7B%22mode2d%22:%22lines_and_points%22,%22mode3d%22:%22lines%22%7D,%22graphOperationMarker%22:[%7B%22annotations%22:[],%22tags%22:[]%7D,%7B%22annotations%22:[],%22tags%22:[]%7D],%22pathFinder%22:%7B%22color%22:%22#ffff00%22,%22pathObject%22:%7B%22annotationPath%22:%7B%22annotations%22:[],%22tags%22:[]%7D,%22hasPath%22:false%7D%7D,%22name%22:%22Production-segmentation_with_graph%22%7D],%22navigation%22:%7B%22pose%22:%7B%22position%22:%7B%22voxelSize%22:[4,4,40],%22voxelCoordinates%22:[115198,55453,1537]%7D%7D,%22zoomFactor%22:4.145162302892657%7D,%22perspectiveOrientation%22:[-0.22566604614257812,-0.034010399132966995,-0.07959923148155212,0.97035151720047],%22perspectiveZoom%22:4888.123757966904,%22showSlices%22:false,%22jsonStateServer%22:%22https://globalv1.flywire-daf.com/nglstate/post%22,%22selectedLayer%22:%7B%22layer%22:%22Production-segmentation_with_graph%22,%22visible%22:true%7D,%22layout%22:%22xy-3d%22%7D"

# }