Update root ids for flywire neurons using XYZ or supervoxel ids

Update root ids for flywire neurons using XYZ or supervoxel ids

flywire_updateids(
  x,
  svids = NULL,
  xyz = NULL,
  rawcoords = FALSE,
  voxdims = c(4, 4, 40),
  Verbose = TRUE,
  ...
)

Arguments

x
Current root ids

svids
optional supervoxel ids

xyz
optional xyz locations in any form understood by xyzmatrix

rawcoords
whether the input values are raw voxel indices or in nm

voxdims
voxel dimensions in nm used to convert raw coordinates.

Verbose
Whether to print a message to the console when updates are required.

...
Additional arguments passed to flywire_islatest or flywire_latestid

Value

See also

Examples

kcs=data.frame(
rootid=c("720575940602553568", "720575940602564320", "720575940602605536"),
xyz=c("(159284,42762,3594)", "(159035,41959,3594)", "(157715,44345,3594)")
)
# update root ids
kcs$rootid=flywire_updateids(kcs$rootid, xyz=kcs$xyz)