Return a blank neuroglancer scene based on a specified segmentation

defaults to the current segmentation defined by choose_segmentation when release=NULL

ngl_blank_scene(release = NULL, return.url = FALSE)

Arguments

release

character vector specifying a released segmentation via a known short name or a sample neuroglancer URL.

return.url

Whether to return a URL rather than a ngscene object.

See also

Examples

# blank scene for current segmentation
ngl_blank_scene()
#> neuroglancer scene with 2 layers and 0 segments (of which 0 shown)
#>                                 name                    type visible nsegs
#> 1                   Production-image                   image    TRUE     0
#> 2 Production-segmentation_with_graph segmentation_with_graph    TRUE     0
#>   nhidden
#> 1       0
#> 2       0
#>                                                                        source
#> 1 precomputed://gs://microns-seunglab/drosophila_v0/alignment/image_rechunked
#> 2        graphene://https://prodv1.flywire-daf.com/segmentation/table/fly_v31
# add a specific id
ngl_blank_scene()+"720575940623755722"
#> neuroglancer scene with 2 layers and 1 segments (of which 1 shown)
#>                                 name                    type visible nsegs
#> 1                   Production-image                   image    TRUE     0
#> 2 Production-segmentation_with_graph segmentation_with_graph    TRUE     1
#>   nhidden
#> 1       0
#> 2       0
#>                                                                        source
#> 1 precomputed://gs://microns-seunglab/drosophila_v0/alignment/image_rechunked
#> 2        graphene://https://prodv1.flywire-daf.com/segmentation/table/fly_v31
# a different segmentation
ngl_blank_scene("202004")
#> neuroglancer scene with 2 layers and 0 segments (of which 0 shown)
#>                 name         type visible nsegs nhidden
#> 1     fafb_v14_clahe        image    TRUE     0       0
#> 2 fafb-ffn1-20200412 segmentation    TRUE     0       0
#>                                                              source
#> 1 precomputed://gs://neuroglancer-fafb-data/fafb_v14/fafb_v14_clahe
#> 2                precomputed://gs://fafb-ffn1-20200412/segmentation
u=ngl_blank_scene("202004", return.url=TRUE)
if (FALSE) {
u=ngl_blank_scene("sandbox")
browseURL(u)
}