Skeletonise neuron meshes using skeletor

skeletor(
  segments = NULL,
  obj = NULL,
  mesh3d = TRUE,
  cloudvolume.url = getOption("fafbseg.cloudvolume.url"),
  clean = TRUE,
  radius = TRUE,
  ratio = 0.1,
  SL = 10,
  WHO = 2,
  iter_lim = 4,
  epsilon = 0.05,
  precision = 1e-06,
  validate = TRUE,
  method.radii = c("knn", "ray"),
  method = c("vertex_clusters", "edge_collapse"),
  sampling_dist = 500,
  ...
)

Arguments

segments
The segment ids to fetch (probably as a character vector). Meshes are read from the specified CloudVolume (cloudvolume.url).

obj
character. Path of a obj file or a folder of such files. These files are read as meshes and then skeletonised. If segments is given, this argument is overrriden.

mesh3d
logical. If TRUE then the neuron's volume is added to each neuron object in the resultant neuronlist] at neuron$mesh3d.

cloudvolume.url
Optional url from which to fetch meshes normally specified by the fafbseg.cloudvolume.url option.

clean
Logical. If TRUE then, in python, skeletor.clean is used to collapse twigs that have line of sight to each other and ove nodes outside the mesh back inside. Note that this is not a magic bullet and some of this will not work (well) if the original mesh was degenerate (e.g. internal faces or not watertight) to begin with.

radius
Logical. Whether or not to return radius information for each skeleton node. If you want to make use of radii, you will need to have the ncollpyde python3 module installed. You can get this with pip3 install ncollpyde.

ratio
Factor to which to reduce mesh faces. For example, a ratio of 0.5 will reduce the number of faces to 50

SLnumeric. Factor by which the contraction matrix is multiplied for each iteration. In theory, lower values are more likely to get you an optimal contraction at the cost of needing more iterations.

WHOnumeric. Initial weight factor for the attraction constraints. The ratio of the initial weights ``WL0`` (``1e-3 * sqrt(A)``) and ``WH0`` controls the smoothness and the degree of contraction of the first iteration result, thus it determines the amount of details retained in subsequent and final contracted meshes.

iter_liminteger. Maximum rounds of contractions.

epsilonnumeric. Target contraction rate as measured by the sum of all face areas in the contracted versus the original mesh. Algorithm will stop once mesh is contracted below this threshold. Depending on your mesh (number of faces, shape) reaching a strong contraction can be extremely costly with comparatively little benefit for the subsequent skeletonization. Note that the algorithm might stop short of this target if iter_lim is reached first or if the sum of face areas is increasing from one iteration to the next instead of decreasing.

precisionnumeric. Sets the precision for finding the least-square solution. This is the main determinant for speed vs quality: lower values will take (much) longer but will get you closer to an optimally contracted mesh. Higher values will be faster but the iterative contractions might stop early.

validateIf True, will try to fix potential issues with the mesh (e.g. infinite values, duplicate vertices, degenerate faces) before collapsing. Degenerate meshes can lead to effectively infinite runtime for this function!

method.radiithe method by which to determine each node's radius. "knn" uses k-nearest-neighbors to get radii: fast but potential for being very wrong. "ray" uses ray-casting to get radii: slower but sometimes less wrong.

methodSkeletonisation comes in two flavours with different Pros and Cons. "vertex_clusters" groups and collapses vertices based on their geodesic distance along the mesh's surface. It's fast and scales well but can lead to oversimplification. Good for quick & dirty skeletonisations. "edge_collapse" implements skeleton extraction by edge collapse described Au et al. 2008. It's rather slow and doesn't scale well but is really good at preserving topology.

sampling_distnumeric. Maximal distance at which vertices are clustered. This parameter should be tuned based on the resolution of your mesh.

...Additional arguments passed to reticulate::py_run_string.