Add flytable cell type information to a dataframe with flywire ids

Add flytable cell type information to a dataframe with flywire ids

add_celltype_info(x, idcol = NULL, version = NULL, ...)

Arguments

x

a data.frame containing root ids

idcol

Optional character vector specifying the column containing ids of the neurons for which cell type information should be provided.

version

Optional numeric CAVE version (see flywire_cave_query). The special signalling value of TRUE uses the current default data dump as returned by flywire_connectome_data_version.

...

additional arguments passed to flytable_cell_types

Value

a data.frame with extra columns

Details

the root ids must be in a column called one of "pre_id", "post_id", "root_id", "post_pt_root_id", "pre_pt_root_id". If you do not have exactly one of these columns present then you must specify your preferred column with the idcol argument.

See also

Examples

# \donttest{
kcin=flywire_partner_summary("720575940626474889", partners = 'in',
  cleft.threshold = 50)
#> Fetching supervoxel ids for id: 720575940626474889
#> Finding synapses for supervoxels
#> Finding additional details for synapses
#> Fetching root ids
kcin
#> # A tibble: 103 × 3
#>    pre_id             weight     n
#>    <chr>               <int> <int>
#>  1 720575940630547909     29     1
#>  2 720575940621280688     28     1
#>  3 720575940632698797     26     1
#>  4 720575940637628762     25     1
#>  5 720575940607687260     19     1
#>  6 720575940625081895     14     1
#>  7 720575940622673703      8     1
#>  8 720575940614137809      6     1
#>  9 720575940621178206      6     1
#> 10 720575940619665023      5     1
#> # … with 93 more rows
kcin2=add_celltype_info(kcin)
kcin2
#> # A tibble: 103 × 11
#>    pre_id     weight     n super…¹ side  cell_…² cell_…³ ito_l…⁴ hemib…⁵ hemib…⁶
#>    <chr>       <int> <int> <chr>   <chr> <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 720575940…     29     1 807870… left  ALPN    VA1d_a… ALad1   VA1d_a… NA     
#>  2 720575940…     28     1 813486… left  MBIN    APL     NA      APL     NA     
#>  3 720575940…     26     1 803649… left  ALPN    VM6_ad… ALad1   VC5_ad… NA     
#>  4 720575940…     25     1 802240… left  ALPN    DM2_lPN ALl1_d… DM2_lPN NA     
#>  5 720575940…     19     1 802240… left  ALPN    VM2_ad… ALad1   VM2_ad… NA     
#>  6 720575940…     14     1 804353… left  ALPN    VA4_lPN ALl1_d… VA4_lPN NA     
#>  7 720575940…      8     1 816302… left  Kenyon… KCab    MBp2    KCab-s  NA     
#>  8 720575940…      6     1 813494… left  NA      LHPV12… NA      LHPV12… NA     
#>  9 720575940…      6     1 812080… left  Kenyon… KCab    MBp2    KCab-s  NA     
#> 10 720575940…      5     1 814192… left  Kenyon… KCab    MBp2    KCab-s  NA     
#> # … with 93 more rows, 1 more variable: vfb_id <chr>, and abbreviated variable
#> #   names ¹​supervoxel_id, ²​cell_class, ³​cell_type, ⁴​ito_lee_hemilineage,
#> #   ⁵​hemibrain_type, ⁶​hemibrain_match
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following object is masked from ‘package:testthat’:
#> 
#>     matches
#> The following objects are masked from ‘package:nat’:
#> 
#>     intersect, setdiff, union
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
kcin2 %>%
  group_by(cell_type) %>%
  summarise(wt = sum(weight),n=n()) %>%
  arrange(desc(wt))
#> # A tibble: 20 × 3
#>    cell_type            wt     n
#>    <chr>             <int> <int>
#>  1 KCab                144    78
#>  2 VA1d_adPN            29     1
#>  3 APL                  28     1
#>  4 DM2_lPN              26     2
#>  5 VM6_adPN             26     1
#>  6 VM2_adPN             19     1
#>  7 VA4_lPN              14     1
#>  8 LHPV12a1              9     2
#>  9 PAM11                 4     2
#> 10 DPM                   3     1
#> 11 Morpholino (PPL2)     3     1
#> 12 PAM10                 3     2
#> 13 PAM04                 2     2
#> 14 PPL105                2     2
#> 15 KCapbp-m              1     1
#> 16 MBON02                1     1
#> 17 MBON06                1     1
#> 18 MBON07                1     1
#> 19 PPL106                1     1
#> 20 NA                    1     1
kcin2 %>%
  count(cell_class, wt = weight)
#> # A tibble: 6 × 2
#>   cell_class      n
#>   <chr>       <int>
#> 1 ALPN          114
#> 2 DAN            15
#> 3 Kenyon_Cell   145
#> 4 MBIN           31
#> 5 MBON            3
#> 6 NA             10
# }