Add flytable cell type information to a dataframe with flywire ids

Add flytable cell type information to a dataframe with flywire ids

add_celltype_info(x, ...)

Arguments

x

a data.frame containing root ids

...

additional arguments passed to flytable_cell_types

Value

a data.frame with extra columns

Details

the root ids must be in a column called one of "pre_id", "post_id", "root_id"

See also

Examples

# \donttest{
kcin=flywire_partner_summary("720575940626474889", partners = 'in',
  cleft.threshold = 50)
#> Fetching supervoxel ids for id: 720575940626474889
#> Finding synapses for supervoxels
#> Finding additional details for synapses
#> Fetching root ids
#> Error in py_call_impl(x, dots$args, dots$keywords): requests.exceptions.HTTPError: 502 Server Error: Bad Gateway for url: https://prod.flywire-daf.com/segmentation/api/v1/table/fly_v31/roots_binary?timestamp=1663168351.83536 content:b'<html>\r\n<head><title>502 Bad Gateway</title></head>\r\n<body>\r\n<center><h1>502 Bad Gateway</h1></center>\r\n<hr><center>nginx/1.17.10</center>\r\n</body>\r\n</html>\r\n'
kcin
#> Error in eval(expr, envir, enclos): object 'kcin' not found
kcin2=add_celltype_info(kcin)
#> Error in is.data.frame(x): object 'kcin' not found
kcin2
#> Error in eval(expr, envir, enclos): object 'kcin2' not found
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following object is masked from ‘package:testthat’:
#> 
#>     matches
#> The following objects are masked from ‘package:nat’:
#> 
#>     intersect, setdiff, union
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
kcin2 %>%
  group_by(cell_type) %>%
  summarise(wt = sum(weight),n=n()) %>%
  arrange(desc(wt))
#> Error in group_by(., cell_type): object 'kcin2' not found
kcin2 %>%
  count(cell_class, wt = weight)
#> Error in count(., cell_class, wt = weight): object 'kcin2' not found
# }