Fetch flytable cell type information to a dataframe with flywire ids

add_celltype_info will add information to an existing dataframe.

flytable_meta will fetch a data.frame of metadata from flytable for a given set of identifiers.

add_celltype_info(x, idcol = NULL, version = NULL, ...)

flytable_meta(ids = NULL, version = NULL, ...)

Arguments

x

a data.frame containing root ids

idcol

Optional character vector specifying the column containing ids of the neurons for which cell type information should be provided.

version

Optional numeric CAVE version (see flywire_cave_query). The special signalling value of TRUE uses the current default data dump as returned by flywire_connectome_data_version.

...

additional arguments passed to flytable_cell_types

ids

Flywire identifiers/query in any form understood by flywire_ids

Value

a data.frame with extra columns

Details

the root ids must be in a column called one of "pre_id", "post_id", "root_id", "post_pt_root_id", "pre_pt_root_id". If you do not have exactly one of these columns present then you must specify your preferred column with the idcol argument.

See also

Examples

# \donttest{
kcin=flywire_partner_summary("720575940626474889", partners = 'in',
  cleft.threshold = 50)
#> Fetching supervoxel ids for id: 720575940626474889
#> Finding synapses for supervoxels
#> Finding additional details for synapses
#> Fetching root ids
kcin
#> # A tibble: 103 × 3
#>    pre_id             weight     n
#>    <chr>               <int> <int>
#>  1 720575940630547909     29     1
#>  2 720575940621280688     28     1
#>  3 720575940632698797     26     1
#>  4 720575940637628762     25     1
#>  5 720575940607687260     19     1
#>  6 720575940625081895     14     1
#>  7 720575940622673703      8     1
#>  8 720575940614137809      6     1
#>  9 720575940621178206      6     1
#> 10 720575940619665023      5     1
#> # … with 93 more rows
kcin2=add_celltype_info(kcin)
kcin2
#> # A tibble: 103 × 13
#>    pre_id       weight     n super…¹ side  flow  super…² cell_…³ cell_…⁴ ito_l…⁵
#>    <chr>         <int> <int> <chr>   <chr> <chr> <chr>   <chr>   <chr>   <chr>  
#>  1 72057594063…     29     1 807870… left  intr… central ALPN    VA1d_a… ALad1  
#>  2 72057594062…     28     1 813486… left  intr… central MBIN    APL     NA     
#>  3 72057594063…     26     1 803649… left  intr… central ALPN    VM6_ad… ALad1  
#>  4 72057594063…     25     1 802240… left  intr… central ALPN    DM2_lPN ALl1_d…
#>  5 72057594060…     19     1 802240… left  intr… central ALPN    VM2_ad… ALad1  
#>  6 72057594062…     14     1 804353… left  intr… central ALPN    VA4_lPN ALl1_d…
#>  7 72057594062…      8     1 816302… left  intr… central Kenyon… KCab    MBp2   
#>  8 72057594061…      6     1 813494… left  intr… central NA      LHPV12… NA     
#>  9 72057594062…      6     1 812080… left  intr… central Kenyon… KCab    MBp2   
#> 10 72057594061…      5     1 814192… left  intr… central Kenyon… KCab    MBp2   
#> # … with 93 more rows, 3 more variables: hemibrain_type <chr>,
#> #   hemibrain_match <chr>, vfb_id <chr>, and abbreviated variable names
#> #   ¹​supervoxel_id, ²​super_class, ³​cell_class, ⁴​cell_type, ⁵​ito_lee_hemilineage
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following object is masked from ‘package:testthat’:
#> 
#>     matches
#> The following objects are masked from ‘package:nat’:
#> 
#>     intersect, setdiff, union
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
kcin2 %>%
  group_by(cell_type) %>%
  summarise(wt = sum(weight),n=n()) %>%
  arrange(desc(wt))
#> # A tibble: 20 × 3
#>    cell_type            wt     n
#>    <chr>             <int> <int>
#>  1 KCab                144    78
#>  2 VA1d_adPN            29     1
#>  3 APL                  28     1
#>  4 DM2_lPN              26     2
#>  5 VM6_adPN             26     1
#>  6 VM2_adPN             19     1
#>  7 VA4_lPN              14     1
#>  8 LHPV12a1              9     2
#>  9 PAM11                 4     2
#> 10 DPM                   3     1
#> 11 Morpholino (PPL2)     3     1
#> 12 PAM10                 3     2
#> 13 PAM04                 2     2
#> 14 PPL105                2     2
#> 15 KCapbp-m              1     1
#> 16 MBON02                1     1
#> 17 MBON06                1     1
#> 18 MBON07                1     1
#> 19 PPL106                1     1
#> 20 NA                    1     1
kcin2 %>%
  count(cell_class, wt = weight)
#> # A tibble: 6 × 2
#>   cell_class      n
#>   <chr>       <int>
#> 1 ALPN          114
#> 2 DAN            15
#> 3 Kenyon_Cell   145
#> 4 MBIN           31
#> 5 MBON            3
#> 6 NA             10
# }
# \donttest{
flytable_meta("class:MBON")
#>               root_id     supervoxel_id  side      flow super_class cell_class
#> 1  720575940624185095 79448854932107696  left intrinsic     central       MBON
#> 2  720575940624117245 80011942324010318  left intrinsic     central       MBON
#> 3  720575940638526278 80715286101144676  left intrinsic     central       MBON
#> 4  720575940629585602 81137979670302132  left intrinsic     central       MBON
#> 5  720575940637934308 81349085835581377  left intrinsic     central       MBON
#> 6  720575940631318031 78603948898007978 right intrinsic     central       MBON
#> 7  720575940638163428 78111986298788517 right intrinsic     central       MBON
#> 8  720575940630496374 81349223341739625  left intrinsic     central       MBON
#> 9  720575940642142861 78393323903110735 right intrinsic     central       MBON
#> 10 720575940611344078 80223254782485510  left intrinsic     central       MBON
#> 11 720575940634822751 81277549061520124  left intrinsic     central       MBON
#> 12 720575940617552340 79588561627810154 right intrinsic     central       MBON
#> 13 720575940624280328 81489685951698728  left intrinsic     central       MBON
#> 14 720575940617567206 78180912067538056 right intrinsic     central       MBON
#> 15 720575940643700640 80152611093049624  left intrinsic     central       MBON
#> 16 720575940614026193 80154466519131949  left intrinsic     central       MBON
#> 17 720575940623001480 80154535238527400  left intrinsic     central       MBON
#> 18 720575940608304220 81066305390179956  left intrinsic     central       MBON
#> 19 720575940614595218 78040449457502610 right intrinsic     central       MBON
#> 20 720575940636992368 80785654845309410  left intrinsic     central       MBON
#> 21 720575940627770628 79166830266640003 right intrinsic     central       MBON
#> 22 720575940637902938 78181599128215629 right intrinsic     central       MBON
#> 23 720575940638774606 78040449457461921 right intrinsic     central       MBON
#> 24 720575940639697827 80224766543977933  left intrinsic     central       MBON
#> 25 720575940626833021 78251349531485275 right intrinsic     central       MBON
#> 26 720575940618008859 77901086224032815 right intrinsic     central       MBON
#> 27 720575940635841143 78110818201436707 right intrinsic     central       MBON
#> 28 720575940618249797 80082723452177136  left intrinsic     central       MBON
#> 29 720575940628779563 77408917331762608 right intrinsic     central       MBON
#> 30 720575940632656173 77549311222677157 right intrinsic     central       MBON
#> 31 720575940623905719 80224972702135904  left intrinsic     central       MBON
#> 32 720575940628581379 78111986298850429 right intrinsic     central       MBON
#> 33 720575940617413957 77478942478556789 right intrinsic     central       MBON
#> 34 720575940617573922 77549311155597154 right intrinsic     central       MBON
#> 35 720575940613157329 78111986365505691 right intrinsic     central       MBON
#> 36 720575940624590316 79167173863853528  left intrinsic     central       MBON
#> 37 720575940645304430 79380479052758063 right intrinsic     central       MBON
#> 38 720575940623743415 77901154877073717 right intrinsic     central       MBON
#> 39 720575940612196850 81136880426746412  left intrinsic     central       MBON
#> 40 720575940622855629 78393117744840114 right intrinsic     central       MBON
#> 41 720575940630767959 78251486970442209 right intrinsic     central       MBON
#> 42 720575940635082347 76985537105536911 right intrinsic     central       MBON
#> 43 720575940628757547 78040449457463315 right intrinsic     central       MBON
#> 44 720575940627144069 79166692827138268  left intrinsic     central       MBON
#> 45 720575940622093436 81136880426744139  left intrinsic     central       MBON
#> 46 720575940622473548 78110818201651448 right intrinsic     central       MBON
#> 47 720575940626046154 77760348668670642 right intrinsic     central       MBON
#> 48 720575940647328900 81137979670303492  left intrinsic     central       MBON
#> 49 720575940625596720 78110818201681795 right intrinsic     central       MBON
#> 50 720575940639556467 81067610926141558  left intrinsic     central       MBON
#> 51 720575940612455338 81700242831237167  left intrinsic     central       MBON
#> 52 720575940617749538 80786135881765925  left intrinsic     central       MBON
#> 53 720575940617760257 81277549061455173  left intrinsic     central       MBON
#> 54 720575940634482783 81136605414678093  left intrinsic     central       MBON
#> 55 720575940644615716 78182286188838645 right intrinsic     central       MBON
#> 56 720575940626109315 81066305390093306  left intrinsic     central       MBON
#> 57 720575940623377802 78110818201621392 right intrinsic     central       MBON
#> 58 720575940626744921 81277549061481189  left intrinsic     central       MBON
#> 59 720575940624774548 81136880426764424  left intrinsic     central       MBON
#> 60 720575940622979277 79027192356412856 right intrinsic     central       MBON
#> 61 720575940614892182 81277617781167541  left intrinsic     central       MBON
#> 62 720575940626872094 78111917579300796 right intrinsic     central       MBON
#> 63 720575940643863496 78324741798855874 right intrinsic     central       MBON
#> 64 720575940630590572 81137979736901319  left intrinsic     central       MBON
#> 65 720575940606953858 81136880426735755  left intrinsic     central       MBON
#> 66 720575940630075703 79309972869657633 right intrinsic     central       MBON
#> 67 720575940652390134 78603949099622269 right intrinsic     central       MBON
#> 68 720575940632535756 79380204241413766 right intrinsic     central       MBON
#> 69 720575940643696288 79309904216692475 right intrinsic     central       MBON
#> 70 720575940623201833 78323023677138555 right intrinsic     central       MBON
#> 71 720575940635063135 80082242348953693  left intrinsic     central       MBON
#> 72 720575940616463477 77830305364995511 right intrinsic     central       MBON
#> 73 720575940629981440 78885355625276424 right intrinsic     central       MBON
#> 74 720575940623182847 78463211409646517 right intrinsic     central       MBON
#> 75 720575940626315010 79943222847625326  left intrinsic     central       MBON
#> 76 720575940632943277 80925980016963911  left intrinsic     central       MBON
#> 77 720575940624539284 78040449457553649 right intrinsic     central       MBON
#> 78 720575940620464321 80011530007496199 right intrinsic     central       MBON
#> 79 720575940628783363 80434292093847745  left intrinsic     central       MBON
#> 80 720575940607183026 79870998677653042 right intrinsic     central       MBON
#> 81 720575940614392999 79237542540872430 right intrinsic     central       MBON
#> 82 720575940617302365 78604086538417636 right intrinsic     central       MBON
#> 83 720575940622997453 81277549061591572  left intrinsic     central       MBON
#> 84 720575940624694503 79941230049538531 right intrinsic     central       MBON
#> 85 720575940617582822 79869899232522382  left intrinsic     central       MBON
#> 86 720575940610964946 81630904409213619  left intrinsic     central       MBON
#> 87 720575940607155890 80433124332533007  left intrinsic     central       MBON
#> 88 720575940623381956 80644779918075196  left intrinsic     central       MBON
#> 89 720575940630864847 81560466945525625  left intrinsic     central       MBON
#> 90 720575940638028607 81419592152467685  left intrinsic     central       MBON
#> 91 720575940621164720 80222636373797362 right intrinsic     central       MBON
#> 92 720575940623841463 78251349531447591 right intrinsic     central       MBON
#> 93 720575940631177803 78814711734490195 right intrinsic     central       MBON
#> 94 720575940624489964 79166761480079751  left intrinsic     central       MBON
#> 95 720575940609440324 79729161744437991 right intrinsic     central       MBON
#> 96 720575940623199684 78463280330557921 right intrinsic     central       MBON
#>      cell_type ito_lee_hemilineage hemibrain_type hemibrain_match vfb_id
#> 1       MBON02                <NA>         MBON02       424789697   <NA>
#> 2       MBON01                <NA>         MBON01       612371421   <NA>
#> 3       MBON32                <NA>         MBON32            <NA>   <NA>
#> 4       MBON10                EBa1         MBON10            <NA>   <NA>
#> 5       MBON30    putative_primary         MBON30            <NA>   <NA>
#> 6       MBON32    putative_primary         MBON32            <NA>   <NA>
#> 7       MBON06    putative_primary         MBON06            <NA>   <NA>
#> 8       MBON05                <NA>         MBON05            <NA>   <NA>
#> 9       MBON10                EBa1         MBON10       768555687   <NA>
#> 10      MBON20             primary         MBON20            <NA>   <NA>
#> 11      MBON16          DL1_dorsal         MBON16            <NA>   <NA>
#> 12      MBON02                <NA>         MBON02            <NA>   <NA>
#> 13      MBON33                <NA>         MBON33            <NA>   <NA>
#> 14      MBON23                <NA>         MBON23       423382015   <NA>
#> 15      MBON29         DL2_ventral         MBON29      5813052909   <NA>
#> 16      MBON14               FLAa2         MBON14       300972942   <NA>
#> 17      MBON12               FLAa2         MBON12            <NA>   <NA>
#> 18 MBON15-like                <NA>    MBON15-like            <NA>   <NA>
#> 19      MBON28          DL1_dorsal         MBON28       457196643   <NA>
#> 20      MBON31                <NA>         MBON31            <NA>   <NA>
#> 21      MBON12               FLAa2         MBON12       704466265   <NA>
#> 22      MBON35                <NA>         MBON35            <NA>   <NA>
#> 23      MBON17                <NA>         MBON17      5812981264   <NA>
#> 24      MBON14               FLAa2         MBON14      5813068729   <NA>
#> 25      MBON19          DL1_dorsal         MBON19            <NA>   <NA>
#> 26      MBON30    putative_primary         MBON30            <NA>   <NA>
#> 27      MBON15          DL1_dorsal         MBON15            <NA>   <NA>
#> 28      MBON27    putative_primary         MBON27      5813061538   <NA>
#> 29      MBON09                <NA>         MBON09            <NA>   <NA>
#> 30      MBON10                EBa1         MBON10            <NA>   <NA>
#> 31      MBON12               FLAa2         MBON12            <NA>   <NA>
#> 32      MBON10                EBa1         MBON10            <NA>   <NA>
#> 33      MBON10                EBa1         MBON10            <NA>   <NA>
#> 34      MBON10                EBa1         MBON10            <NA>   <NA>
#> 35      MBON10                EBa1         MBON10            <NA>   <NA>
#> 36      MBON03                <NA>         MBON03       612738462   <NA>
#> 37      MBON13               FLAa2         MBON13            <NA>   <NA>
#> 38        <NA>      AOTUv4_ventral           <NA>            <NA>   <NA>
#> 39 MBON15-like          DL1_dorsal    MBON15-like            <NA>   <NA>
#> 40      MBON10                EBa1         MBON10            <NA>   <NA>
#> 41      MBON15          DL1_dorsal         MBON15       673702721   <NA>
#> 42 MBON15-like          DL1_dorsal    MBON15-like       642664141   <NA>
#> 43 MBON15-like          DL1_dorsal    MBON15-like       579916831   <NA>
#> 44      MBON25              SMPad3  MBON25,MBON34            <NA>   <NA>
#> 45      MBON15          DL1_dorsal         MBON15            <NA>   <NA>
#> 46 MBON15-like          DL1_dorsal    MBON15-like       457175171   <NA>
#> 47      MBON09               AVM17         MBON09      1078693835   <NA>
#> 48      MBON10                EBa1         MBON10            <NA>   <NA>
#> 49      MBON17          DL1_dorsal    MBON17-like      5812981264   <NA>
#> 50      MBON10                EBa1         MBON10            <NA>   <NA>
#> 51      MBON15          DL1_dorsal         MBON15            <NA>   <NA>
#> 52      MBON11                <NA>         MBON11            <NA>   <NA>
#> 53      MBON17          DL1_dorsal    MBON17-like      1932553362   <NA>
#> 54      MBON19          DL1_dorsal         MBON19            <NA>   <NA>
#> 55      MBON31                <NA>         MBON31            <NA>   <NA>
#> 56      MBON19          DL1_dorsal         MBON19      1902178951   <NA>
#> 57      MBON16          DL1_dorsal         MBON16            <NA>   <NA>
#> 58      MBON17          DL1_dorsal         MBON17            <NA>   <NA>
#> 59      MBON24                <NA>         MBON24            <NA>   <NA>
#> 60      MBON27    putative_primary         MBON27      5813053891   <NA>
#> 61      MBON28          DL1_dorsal         MBON28            <NA>   <NA>
#> 62      MBON21                <NA>         MBON21            <NA>   <NA>
#> 63      MBON20                <NA>         MBON20            <NA>   <NA>
#> 64      MBON21                <NA>         MBON21       985813153   <NA>
#> 65      MBON23          DL1_dorsal         MBON23            <NA>   <NA>
#> 66      MBON12               FLAa2         MBON12            <NA>   <NA>
#> 67      MBON07                <NA>         MBON07            <NA>   <NA>
#> 68      MBON14               FLAa2         MBON14            <NA>   <NA>
#> 69      MBON14               FLAa2         MBON14            <NA>   <NA>
#> 70      MBON11                <NA>         MBON11            <NA>   <NA>
#> 71      MBON22                <NA>         MBON22       985131499   <NA>
#> 72      MBON22                <NA>         MBON22      5813042659   <NA>
#> 73      MBON26    putative_primary         MBON26            <NA>   <NA>
#> 74      MBON33                <NA>         MBON33            <NA>   <NA>
#> 75      MBON13               FLAa2         MBON13            <NA>   <NA>
#> 76      MBON35    putative_primary         MBON35            <NA>   <NA>
#> 77      MBON18                <NA>         MBON18      5813020828   <NA>
#> 78      MBON04    putative_primary         MBON04            <NA>   <NA>
#> 79      MBON07                <NA>         MBON07            <NA>   <NA>
#> 80      MBON01                <NA>         MBON01       673509195   <NA>
#> 81      MBON29         DL2_ventral         MBON29      5813058048   <NA>
#> 82      MBON07                <NA>         MBON07      1234386037   <NA>
#> 83      MBON18                <NA>         MBON18       457196444   <NA>
#> 84      MBON03                <NA>         MBON03       487925037   <NA>
#> 85      MBON04    putative_primary         MBON04            <NA>   <NA>
#> 86      MBON09                <NA>         MBON09            <NA>   <NA>
#> 87      MBON26    putative_primary         MBON26            <NA>   <NA>
#> 88      MBON07                <NA>         MBON07            <NA>   <NA>
#> 89      MBON09                <NA>         MBON09            <NA>   <NA>
#> 90      MBON06                <NA>         MBON06            <NA>   <NA>
#> 91      MBON05    putative_primary         MBON05            <NA>   <NA>
#> 92      MBON19          DL1_dorsal         MBON19       423774471   <NA>
#> 93      MBON24                <NA>         MBON24            <NA>   <NA>
#> 94      MBON25              SMPad3  MBON25,MBON34       612742248   <NA>
#> 95      MBON34              SMPad3  MBON25,MBON34            <NA>   <NA>
#> 96      MBON10                EBa1         MBON10            <NA>   <NA>
flytable_meta("type:MBON2%")
#>               root_id     supervoxel_id  side      flow super_class cell_class
#> 1  720575940632939181 79588424189358485  left intrinsic     central       <NA>
#> 2  720575940611344078 80223254782485510  left intrinsic     central       MBON
#> 3  720575940617567206 78180912067538056 right intrinsic     central       MBON
#> 4  720575940643700640 80152611093049624  left intrinsic     central       MBON
#> 5  720575940614595218 78040449457502610 right intrinsic     central       MBON
#> 6  720575940618249797 80082723452177136  left intrinsic     central       MBON
#> 7  720575940624659943 80011255129536552 right intrinsic     central       <NA>
#> 8  720575940627144069 79166692827138268  left intrinsic     central       MBON
#> 9  720575940624774548 81136880426764424  left intrinsic     central       MBON
#> 10 720575940622979277 79027192356412856 right intrinsic     central       MBON
#> 11 720575940614892182 81277617781167541  left intrinsic     central       MBON
#> 12 720575940626872094 78111917579300796 right intrinsic     central       MBON
#> 13 720575940643863496 78324741798855874 right intrinsic     central       MBON
#> 14 720575940630590572 81137979736901319  left intrinsic     central       MBON
#> 15 720575940606953858 81136880426735755  left intrinsic     central       MBON
#> 16 720575940635063135 80082242348953693  left intrinsic     central       MBON
#> 17 720575940616463477 77830305364995511 right intrinsic     central       MBON
#> 18 720575940629981440 78885355625276424 right intrinsic     central       MBON
#> 19 720575940614392999 79237542540872430 right intrinsic     central       MBON
#> 20 720575940607155890 80433124332533007  left intrinsic     central       MBON
#> 21 720575940631177803 78814711734490195 right intrinsic     central       MBON
#> 22 720575940624489964 79166761480079751  left intrinsic     central       MBON
#> 23 720575940609440324 79729161744437991 right intrinsic     central       MBON
#> 24 720575940623750325 79729161744426207 right intrinsic     central       <NA>
#>        cell_type ito_lee_hemilineage hemibrain_type hemibrain_match vfb_id
#> 1  MBON25,MBON34              SMPad3  MBON25,MBON34            <NA>   <NA>
#> 2         MBON20             primary         MBON20            <NA>   <NA>
#> 3         MBON23                <NA>         MBON23       423382015   <NA>
#> 4         MBON29         DL2_ventral         MBON29      5813052909   <NA>
#> 5         MBON28          DL1_dorsal         MBON28       457196643   <NA>
#> 6         MBON27    putative_primary         MBON27      5813061538   <NA>
#> 7  MBON25,MBON34              SMPad3  MBON25,MBON34       612422902   <NA>
#> 8         MBON25              SMPad3  MBON25,MBON34            <NA>   <NA>
#> 9         MBON24                <NA>         MBON24            <NA>   <NA>
#> 10        MBON27    putative_primary         MBON27      5813053891   <NA>
#> 11        MBON28          DL1_dorsal         MBON28            <NA>   <NA>
#> 12        MBON21                <NA>         MBON21            <NA>   <NA>
#> 13        MBON20                <NA>         MBON20            <NA>   <NA>
#> 14        MBON21                <NA>         MBON21       985813153   <NA>
#> 15        MBON23          DL1_dorsal         MBON23            <NA>   <NA>
#> 16        MBON22                <NA>         MBON22       985131499   <NA>
#> 17        MBON22                <NA>         MBON22      5813042659   <NA>
#> 18        MBON26    putative_primary         MBON26            <NA>   <NA>
#> 19        MBON29         DL2_ventral         MBON29      5813058048   <NA>
#> 20        MBON26    putative_primary         MBON26            <NA>   <NA>
#> 21        MBON24                <NA>         MBON24            <NA>   <NA>
#> 22        MBON25              SMPad3  MBON25,MBON34       612742248   <NA>
#> 23        MBON34              SMPad3  MBON25,MBON34            <NA>   <NA>
#> 24 MBON25,MBON34              SMPad3  MBON25,MBON34            <NA>   <NA>
# the / introduces a regex query (small performance penalty, more flexible)
flytable_meta("/type:MBON2[0-5]")
#>               root_id     supervoxel_id  side      flow super_class cell_class
#> 1  720575940632939181 79588424189358485  left intrinsic     central       <NA>
#> 2  720575940611344078 80223254782485510  left intrinsic     central       MBON
#> 3  720575940617567206 78180912067538056 right intrinsic     central       MBON
#> 4  720575940624659943 80011255129536552 right intrinsic     central       <NA>
#> 5  720575940627144069 79166692827138268  left intrinsic     central       MBON
#> 6  720575940624774548 81136880426764424  left intrinsic     central       MBON
#> 7  720575940626872094 78111917579300796 right intrinsic     central       MBON
#> 8  720575940643863496 78324741798855874 right intrinsic     central       MBON
#> 9  720575940630590572 81137979736901319  left intrinsic     central       MBON
#> 10 720575940606953858 81136880426735755  left intrinsic     central       MBON
#> 11 720575940635063135 80082242348953693  left intrinsic     central       MBON
#> 12 720575940616463477 77830305364995511 right intrinsic     central       MBON
#> 13 720575940631177803 78814711734490195 right intrinsic     central       MBON
#> 14 720575940624489964 79166761480079751  left intrinsic     central       MBON
#> 15 720575940623750325 79729161744426207 right intrinsic     central       <NA>
#>        cell_type ito_lee_hemilineage hemibrain_type hemibrain_match vfb_id
#> 1  MBON25,MBON34              SMPad3  MBON25,MBON34            <NA>   <NA>
#> 2         MBON20             primary         MBON20            <NA>   <NA>
#> 3         MBON23                <NA>         MBON23       423382015   <NA>
#> 4  MBON25,MBON34              SMPad3  MBON25,MBON34       612422902   <NA>
#> 5         MBON25              SMPad3  MBON25,MBON34            <NA>   <NA>
#> 6         MBON24                <NA>         MBON24            <NA>   <NA>
#> 7         MBON21                <NA>         MBON21            <NA>   <NA>
#> 8         MBON20                <NA>         MBON20            <NA>   <NA>
#> 9         MBON21                <NA>         MBON21       985813153   <NA>
#> 10        MBON23          DL1_dorsal         MBON23            <NA>   <NA>
#> 11        MBON22                <NA>         MBON22       985131499   <NA>
#> 12        MBON22                <NA>         MBON22      5813042659   <NA>
#> 13        MBON24                <NA>         MBON24            <NA>   <NA>
#> 14        MBON25              SMPad3  MBON25,MBON34       612742248   <NA>
#> 15 MBON25,MBON34              SMPad3  MBON25,MBON34            <NA>   <NA>
# }