Fetch flytable cell type information to a dataframe with flywire ids

add_celltype_info will add information to an existing dataframe.

flytable_meta will fetch a data.frame of metadata from flytable for a given set of identifiers.

Usage

add_celltype_info(
  x,
  idcol = NULL,
  version = NULL,
  suffix = NULL,
  table = c("both", "info", "optic"),
  ...
)

flytable_meta(
  ids = NULL,
  version = NULL,
  table = c("both", "info", "optic"),
  unique = FALSE,
  ...
)

Arguments

x

a data.frame containing root ids or a neuronlist ()

idcol

Optional character vector specifying the column containing ids of the neurons for which cell type information should be provided.

version

Optional numeric CAVE version (see flywire_cave_query). The special signalling value of TRUE uses the current default data dump as returned by flywire_connectome_data_version.

suffix

A character suffix for the new columns (default value of NULL implies no suffix).

table

Which cell type information tables to use (info for brain, optic for optic lobes or both).

...

additional arguments passed to flytable_cell_types

ids

Flywire identifiers/query in any form understood by flywire_ids

unique

Whether to ensure that rows contain only unique identifiers. Default FALSE. When TRUE duplicate rows will be returned with a warning.

Value

a data.frame with extra columns

Details

the root ids must be in a column called one of "pre_id", "post_id", "root_id", "post_pt_root_id", "pre_pt_root_id". If you do not have exactly one of these columns present then you must specify your preferred column with the idcol argument.

See also

Examples

# \donttest{
kcin=flywire_partner_summary("720575940626474889", partners = 'in',
  cleft.threshold = 50)
#> Warning: /home/runner/projects/JanFunke//flywire_synapses.db does not exist
#> Warning: /home/runner/projects/JanFunke//20191211_fafbv14_buhmann2019_li20190805_nt20201223.db does not exist
#> Fetching supervoxel ids for id: 720575940626474889
#> Finding synapses for supervoxels
#> Reading synapse data
#> Fetching root ids
kcin
#> # A tibble: 103 × 3
#>    pre_id             weight     n
#>    <chr>               <int> <int>
#>  1 720575940630547909     29     1
#>  2 720575940613583001     28     1
#>  3 720575940632698797     26     1
#>  4 720575940637628762     25     1
#>  5 720575940607687260     19     1
#>  6 720575940642086389     14     1
#>  7 720575940622673703      8     1
#>  8 720575940614137809      6     1
#>  9 720575940621178206      6     1
#> 10 720575940619665023      5     1
#> # ℹ 93 more rows
kcin2=add_celltype_info(kcin)
kcin2
#> # A tibble: 103 × 13
#>    pre_id          weight     n supervoxel_id side  flow  super_class cell_class
#>    <chr>            <int> <int> <chr>         <chr> <chr> <chr>       <chr>     
#>  1 72057594063054…     29     1 807870292352… right intr… central     ALPN      
#>  2 72057594061358…     28     1 813486739212… right intr… central     MBIN      
#>  3 72057594063269…     26     1 803649542764… right intr… central     ALPN      
#>  4 72057594063762…     25     1 802240106295… right intr… central     ALPN      
#>  5 72057594060768…     19     1 802240792818… right intr… central     ALPN      
#>  6 72057594064208…     14     1 804353917399… right intr… central     ALPN      
#>  7 72057594062267…      8     1 816302178184… right intr… central     Kenyon_Ce…
#>  8 72057594061413…      6     1 813494986893… right intr… central     NA        
#>  9 72057594062117…      6     1 812080052867… right intr… central     Kenyon_Ce…
#> 10 72057594061966…      5     1 814192490249… right intr… central     Kenyon_Ce…
#> # ℹ 93 more rows
#> # ℹ 5 more variables: cell_type <chr>, top_nt <chr>, ito_lee_hemilineage <chr>,
#> #   hemibrain_type <chr>, fbbt_id <chr>
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:nat’:
#> 
#>     intersect, setdiff, union
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
kcin2 %>%
  group_by(cell_type) %>%
  summarise(wt = sum(weight),n=n()) %>%
  arrange(desc(wt))
#> # A tibble: 20 × 3
#>    cell_type    wt     n
#>    <chr>     <int> <int>
#>  1 KCab        144    78
#>  2 VA1d_adPN    29     1
#>  3 APL          28     1
#>  4 DM2_lPN      26     2
#>  5 VM6_adPN     26     1
#>  6 VM2_adPN     19     1
#>  7 VA4_lPN      14     1
#>  8 LHPV12a1      9     2
#>  9 PAM11         4     2
#> 10 DPM           3     1
#> 11 PAM10         3     2
#> 12 PPL202        3     1
#> 13 PAM04         2     2
#> 14 PPL105        2     2
#> 15 KCapbp-m      1     1
#> 16 MBON02        1     1
#> 17 MBON06        1     1
#> 18 MBON07        1     1
#> 19 PPL106        1     1
#> 20 NA            1     1
kcin2 %>%
  count(cell_class, wt = weight)
#> # A tibble: 6 × 2
#>   cell_class      n
#>   <chr>       <int>
#> 1 ALPN          114
#> 2 DAN            15
#> 3 Kenyon_Cell   145
#> 4 MBIN           31
#> 5 MBON            3
#> 6 NA             10
# }

if (FALSE) {
# read neuronlist containing "dotprops" for some olfactory projection neurons
da2=read_l2dps('DA2')
# add cell type details to that
da2=add_celltype_info(da2)
}
# \donttest{
flytable_meta("class:MBON")
#>               root_id     supervoxel_id  side      flow super_class cell_class
#> 1  720575940624185095 79448854932107696  left intrinsic     central       MBON
#> 2  720575940624117245 80011942324010318 right intrinsic     central       MBON
#> 3  720575940638526278 80715286101144676 right intrinsic     central       MBON
#> 4  720575940629585602 81137979670302132 right intrinsic     central       MBON
#> 5  720575940637934308 81349085835581377 right intrinsic     central       MBON
#> 6  720575940632118343 78744342923662254  left intrinsic     central       MBON
#> 7  720575940609959637 78603948898007978  left intrinsic     central       MBON
#> 8  720575940638163428 78111986298788517  left intrinsic     central       MBON
#> 9  720575940630496374 81349223341739625 right intrinsic     central       MBON
#> 10 720575940642142861 78393323903110735  left intrinsic     central       MBON
#> 11 720575940611344078 80223254782485510 right intrinsic     central       MBON
#> 12 720575940634822751 81277549061520124 right intrinsic     central       MBON
#> 13 720575940617552340 79588561627810154 right intrinsic     central       MBON
#> 14 720575940624280328 81489685951698728 right intrinsic     central       MBON
#> 15 720575940617567206 78180912067538056  left intrinsic     central       MBON
#> 16 720575940643700640 80152611093049624 right intrinsic     central       MBON
#> 17 720575940614206866 81137979670297348 right intrinsic     central       MBON
#> 18 720575940614026193 80154466519131949 right intrinsic     central       MBON
#> 19 720575940633209647 78463486488950059  left intrinsic     central       MBON
#> 20 720575940623001480 80154535238527400 right intrinsic     central       MBON
#> 21 720575940608304220 81066305390179956 right intrinsic     central       MBON
#> 22 720575940614595218 78040449457502610  left intrinsic     central       MBON
#> 23 720575940636992368 80785654845309410 right intrinsic     central       MBON
#> 24 720575940627770628 79166830266640003  left intrinsic     central       MBON
#> 25 720575940637902938 78181599128215629  left intrinsic     central       MBON
#> 26 720575940638774606 78040449457461921  left intrinsic     central       MBON
#> 27 720575940639697827 80224766543977933 right intrinsic     central       MBON
#> 28 720575940626833021 78251349531485275  left intrinsic     central       MBON
#> 29 720575940618008859 77901086224032815  left intrinsic     central       MBON
#> 30 720575940635841143 78110818201436707  left intrinsic     central       MBON
#> 31 720575940618249797 80082723452177136 right intrinsic     central       MBON
#> 32 720575940629856515 77408917331762608  left intrinsic     central       MBON
#> 33 720575940623905719 80224972702135904 right intrinsic     central       MBON
#> 34 720575940629529722 78533992672102680  left intrinsic     central       MBON
#> 35 720575940624659943 80011255129536552  left intrinsic     central       MBON
#> 36 720575940624590316 79167173863853528 right intrinsic     central       MBON
#> 37 720575940645304430 79380479052758063  left intrinsic     central       MBON
#> 38 720575940612196850 81136880426746412 right intrinsic     central       MBON
#> 39 720575940630767959 78251486970442209  left intrinsic     central       MBON
#> 40 720575940635082347 76985537105536911  left intrinsic     central       MBON
#> 41 720575940628757547 78040449457463315  left intrinsic     central       MBON
#> 42 720575940627144069 79166692827138268 right intrinsic     central       MBON
#> 43 720575940610647416 81137979602919932 right intrinsic     central       MBON
#> 44 720575940622093436 81136880426744139 right intrinsic     central       MBON
#> 45 720575940624696810 78110818201651448  left intrinsic     central       MBON
#> 46 720575940629422086 77760348668670642  left intrinsic     central       MBON
#> 47 720575940647328900 81137979670303492 right intrinsic     central       MBON
#> 48 720575940650386553 78110818201681795  left intrinsic     central       MBON
#> 49 720575940639556467 81067610926141558 right intrinsic     central       MBON
#> 50 720575940612455338 81700242831237167 right intrinsic     central       MBON
#> 51 720575940617749538 80786135881765925 right intrinsic     central       MBON
#> 52 720575940617760257 81277549061455173 right intrinsic     central       MBON
#> 53 720575940619810389 79518055445188833 right intrinsic     central       MBON
#> 54 720575940634482783 81136605414678093 right intrinsic     central       MBON
#> 55 720575940644615716 78182286188838645  left intrinsic     central       MBON
#> 56 720575940626109315 81066305390093306 right intrinsic     central       MBON
#> 57 720575940623377802 78110818201621392  left intrinsic     central       MBON
#> 58 720575940626744921 81277549061481189 right intrinsic     central       MBON
#> 59 720575940623464316 81136880426764424 right intrinsic     central       MBON
#> 60 720575940622979277 79027192356412856  left intrinsic     central       MBON
#> 61 720575940614892182 81277617781167541 right intrinsic     central       MBON
#> 62 720575940620715224 78111917579300796  left intrinsic     central       MBON
#> 63 720575940643863496 78324741798855874  left intrinsic     central       MBON
#> 64 720575940621777391 81137979736901319 right intrinsic     central       MBON
#> 65 720575940606953858 81136880426735755 right intrinsic     central       MBON
#> 66 720575940630075703 79309972869657633  left intrinsic     central       MBON
#> 67 720575940652390134 78603949099622269  left intrinsic     central       MBON
#> 68 720575940632535756 79380204241413766  left intrinsic     central       MBON
#> 69 720575940643696288 79309904216692475  left intrinsic     central       MBON
#> 70 720575940623201833 78323023677138555  left intrinsic     central       MBON
#> 71 720575940635063135 80082242348953693 right intrinsic     central       MBON
#> 72 720575940616463477 77830305364995511  left intrinsic     central       MBON
#> 73 720575940629981440 78885355625276424  left intrinsic     central       MBON
#> 74 720575940623182847 78463211409646517  left intrinsic     central       MBON
#> 75 720575940626315010 79943222847625326 right intrinsic     central       MBON
#> 76 720575940632943277 80925980016963911 right intrinsic     central       MBON
#> 77 720575940624539284 78040449457553649  left intrinsic     central       MBON
#> 78 720575940620464321 80011530007496199  left intrinsic     central       MBON
#> 79 720575940628783363 80434292093847745 right intrinsic     central       MBON
#> 80 720575940643309197 79870998677653042  left intrinsic     central       MBON
#> 81 720575940614392999 79237542540872430  left intrinsic     central       MBON
#> 82 720575940617302365 78604086538417636  left intrinsic     central       MBON
#> 83 720575940622997453 81277549061591572 right intrinsic     central       MBON
#> 84 720575940624694503 79941230049538531  left intrinsic     central       MBON
#> 85 720575940628734376 79869899232522382 right intrinsic     central       MBON
#> 86 720575940610964946 81630904409213619 right intrinsic     central       MBON
#> 87 720575940607155890 80433124332533007 right intrinsic     central       MBON
#> 88 720575940623381956 80644779918075196 right intrinsic     central       MBON
#> 89 720575940630864847 81560466945525625 right intrinsic     central       MBON
#> 90 720575940638028607 81419592152467685 right intrinsic     central       MBON
#> 91 720575940621164720 80222636373797362  left intrinsic     central       MBON
#> 92 720575940623841463 78251349531447591  left intrinsic     central       MBON
#> 93 720575940631177803 78814711734490195  left intrinsic     central       MBON
#> 94 720575940624489964 79166761480079751 right intrinsic     central       MBON
#> 95 720575940616398303 79729161744437991  left intrinsic     central       MBON
#> 96 720575940623750325 79729161744426207  left intrinsic     central       MBON
#>        cell_type        top_nt ito_lee_hemilineage hemibrain_type       fbbt_id
#> 1         MBON02     glutamate    putative_primary         MBON02 FBbt_00111012
#> 2         MBON01     glutamate    putative_primary         MBON01 FBbt_00100234
#> 3         MBON32          gaba    putative_primary         MBON32 FBbt_00049852
#> 4         MBON10          gaba                EBa1         MBON10 FBbt_00111008
#> 5         MBON30     glutamate    putative_primary         MBON30 FBbt_00049850
#> 6         MBON10          gaba                EBa1         MBON10 FBbt_00111008
#> 7         MBON32          gaba    putative_primary         MBON32 FBbt_00049852
#> 8         MBON06     glutamate    putative_primary         MBON06 FBbt_00100242
#> 9         MBON05     glutamate    putative_primary         MBON05 FBbt_00111004
#> 10        MBON10     glutamate                EBa1         MBON10 FBbt_00111008
#> 11        MBON20          gaba    putative_primary         MBON20 FBbt_00111013
#> 12   MBON17-like acetylcholine          DL1_dorsal    MBON17-like          <NA>
#> 13        MBON02          gaba    putative_primary         MBON02 FBbt_00111012
#> 14        MBON33 acetylcholine    putative_primary         MBON33 FBbt_00049853
#> 15        MBON23 acetylcholine          DL1_dorsal         MBON23 FBbt_00049113
#> 16        MBON29 acetylcholine         DL2_ventral         MBON29 FBbt_00049849
#> 17        MBON10          gaba                EBa1         MBON10 FBbt_00111008
#> 18        MBON14 acetylcholine               FLAa2         MBON14 FBbt_00100238
#> 19        MBON10     glutamate                EBa1         MBON10 FBbt_00111008
#> 20        MBON12 acetylcholine               FLAa2         MBON12 FBbt_00111009
#> 21   MBON15-like acetylcholine          DL1_dorsal    MBON15-like          <NA>
#> 22        MBON28 acetylcholine          DL1_dorsal         MBON28 FBbt_00049848
#> 23        MBON31          gaba    putative_primary         MBON31 FBbt_00049851
#> 24        MBON12 acetylcholine               FLAa2         MBON12 FBbt_00111009
#> 25        MBON35 acetylcholine    putative_primary         MBON35 FBbt_00049855
#> 26        MBON17 acetylcholine          DL1_dorsal         MBON17 FBbt_00111064
#> 27        MBON14 acetylcholine               FLAa2         MBON14 FBbt_00100238
#> 28        MBON19 acetylcholine          DL1_dorsal         MBON19 FBbt_00111011
#> 29        MBON30     glutamate    putative_primary         MBON30 FBbt_00049850
#> 30        MBON15 acetylcholine          DL1_dorsal         MBON15 FBbt_00111010
#> 31        MBON27 acetylcholine    putative_primary         MBON27 FBbt_00049847
#> 32        MBON09          gaba    putative_primary         MBON09 FBbt_00111007
#> 33        MBON12 acetylcholine               FLAa2         MBON12 FBbt_00111009
#> 34        MBON10     glutamate                EBa1         MBON10 FBbt_00111008
#> 35 MBON25,MBON34     glutamate              SMPad3  MBON25,MBON34          <NA>
#> 36        MBON03     glutamate    putative_primary         MBON03 FBbt_00100232
#> 37        MBON13 acetylcholine               FLAa2         MBON13 FBbt_00100239
#> 38   MBON15-like acetylcholine          DL1_dorsal    MBON15-like          <NA>
#> 39        MBON15 acetylcholine          DL1_dorsal         MBON15 FBbt_00111010
#> 40   MBON15-like acetylcholine          DL1_dorsal    MBON15-like          <NA>
#> 41   MBON15-like acetylcholine          DL1_dorsal    MBON15-like          <NA>
#> 42 MBON25,MBON34     glutamate              SMPad3  MBON25,MBON34          <NA>
#> 43        MBON10          gaba                EBa1         MBON10 FBbt_00111008
#> 44        MBON15 acetylcholine          DL1_dorsal         MBON15 FBbt_00111010
#> 45   MBON15-like acetylcholine          DL1_dorsal    MBON15-like          <NA>
#> 46        MBON09          gaba    putative_primary         MBON09 FBbt_00111007
#> 47        MBON10          gaba                EBa1         MBON10 FBbt_00111008
#> 48   MBON17-like acetylcholine          DL1_dorsal    MBON17-like          <NA>
#> 49        MBON10          gaba                EBa1         MBON10 FBbt_00111008
#> 50        MBON15 acetylcholine          DL1_dorsal         MBON15 FBbt_00111010
#> 51        MBON11          gaba    putative_primary         MBON11 FBbt_00100246
#> 52        MBON17 acetylcholine          DL1_dorsal         MBON17 FBbt_00111064
#> 53 MBON25,MBON34     glutamate              SMPad3  MBON25,MBON34          <NA>
#> 54        MBON19 acetylcholine          DL1_dorsal         MBON19 FBbt_00111011
#> 55        MBON31          gaba    putative_primary         MBON31 FBbt_00049851
#> 56        MBON19 acetylcholine          DL1_dorsal         MBON19 FBbt_00111011
#> 57        MBON16 acetylcholine          DL1_dorsal         MBON16 FBbt_00111063
#> 58        MBON16 acetylcholine          DL1_dorsal         MBON16 FBbt_00111063
#> 59        MBON24 acetylcholine    putative_primary         MBON24 FBbt_00049844
#> 60        MBON27 acetylcholine    putative_primary         MBON27 FBbt_00049847
#> 61        MBON28 acetylcholine          DL1_dorsal         MBON28 FBbt_00049848
#> 62        MBON21 acetylcholine    putative_primary         MBON21 FBbt_00111046
#> 63        MBON20          gaba    putative_primary         MBON20 FBbt_00111013
#> 64        MBON21 acetylcholine    putative_primary         MBON21 FBbt_00111046
#> 65        MBON23 acetylcholine          DL1_dorsal         MBON23 FBbt_00049113
#> 66        MBON12 acetylcholine               FLAa2         MBON12 FBbt_00111009
#> 67        MBON07     glutamate    putative_primary         MBON07 FBbt_00111005
#> 68        MBON14 acetylcholine               FLAa2         MBON14 FBbt_00100238
#> 69        MBON14 acetylcholine               FLAa2         MBON14 FBbt_00100238
#> 70        MBON11          gaba    putative_primary         MBON11 FBbt_00100246
#> 71        MBON22 acetylcholine    putative_primary         MBON22 FBbt_00100240
#> 72        MBON22 acetylcholine    putative_primary         MBON22 FBbt_00100240
#> 73        MBON26 acetylcholine    putative_primary         MBON26 FBbt_00049846
#> 74        MBON33 acetylcholine    putative_primary         MBON33 FBbt_00049853
#> 75        MBON13 acetylcholine               FLAa2         MBON13 FBbt_00100239
#> 76        MBON35 acetylcholine    putative_primary         MBON35 FBbt_00049855
#> 77        MBON18 acetylcholine          DL1_dorsal         MBON18 FBbt_00110101
#> 78        MBON04     glutamate    putative_primary         MBON04 FBbt_00111014
#> 79        MBON07     glutamate    putative_primary         MBON07 FBbt_00111005
#> 80        MBON01     glutamate    putative_primary         MBON01 FBbt_00100234
#> 81        MBON29 acetylcholine         DL2_ventral         MBON29 FBbt_00049849
#> 82        MBON07     glutamate    putative_primary         MBON07 FBbt_00111005
#> 83        MBON18 acetylcholine          DL1_dorsal         MBON18 FBbt_00110101
#> 84        MBON03     glutamate    putative_primary         MBON03 FBbt_00100232
#> 85        MBON04     glutamate    putative_primary         MBON04 FBbt_00111014
#> 86        MBON09          gaba    putative_primary         MBON09 FBbt_00111007
#> 87        MBON26 acetylcholine    putative_primary         MBON26 FBbt_00049846
#> 88        MBON07     glutamate    putative_primary         MBON07 FBbt_00111005
#> 89        MBON09          gaba    putative_primary         MBON09 FBbt_00111007
#> 90        MBON06     glutamate    putative_primary         MBON06 FBbt_00100242
#> 91        MBON05 acetylcholine    putative_primary         MBON05 FBbt_00111004
#> 92        MBON19 acetylcholine          DL1_dorsal         MBON19 FBbt_00111011
#> 93        MBON24 acetylcholine    putative_primary         MBON24 FBbt_00049844
#> 94 MBON25,MBON34     glutamate              SMPad3  MBON25,MBON34          <NA>
#> 95 MBON25,MBON34     glutamate              SMPad3  MBON25,MBON34          <NA>
#> 96 MBON25,MBON34     glutamate              SMPad3  MBON25,MBON34          <NA>
flytable_meta("type:MBON2%")
#>               root_id     supervoxel_id  side      flow super_class cell_class
#> 1  720575940611344078 80223254782485510 right intrinsic     central       MBON
#> 2  720575940617567206 78180912067538056  left intrinsic     central       MBON
#> 3  720575940643700640 80152611093049624 right intrinsic     central       MBON
#> 4  720575940614595218 78040449457502610  left intrinsic     central       MBON
#> 5  720575940618249797 80082723452177136 right intrinsic     central       MBON
#> 6  720575940624659943 80011255129536552  left intrinsic     central       MBON
#> 7  720575940627144069 79166692827138268 right intrinsic     central       MBON
#> 8  720575940619810389 79518055445188833 right intrinsic     central       MBON
#> 9  720575940623464316 81136880426764424 right intrinsic     central       MBON
#> 10 720575940622979277 79027192356412856  left intrinsic     central       MBON
#> 11 720575940614892182 81277617781167541 right intrinsic     central       MBON
#> 12 720575940620715224 78111917579300796  left intrinsic     central       MBON
#> 13 720575940643863496 78324741798855874  left intrinsic     central       MBON
#> 14 720575940621777391 81137979736901319 right intrinsic     central       MBON
#> 15 720575940606953858 81136880426735755 right intrinsic     central       MBON
#> 16 720575940635063135 80082242348953693 right intrinsic     central       MBON
#> 17 720575940616463477 77830305364995511  left intrinsic     central       MBON
#> 18 720575940629981440 78885355625276424  left intrinsic     central       MBON
#> 19 720575940614392999 79237542540872430  left intrinsic     central       MBON
#> 20 720575940607155890 80433124332533007 right intrinsic     central       MBON
#> 21 720575940631177803 78814711734490195  left intrinsic     central       MBON
#> 22 720575940624489964 79166761480079751 right intrinsic     central       MBON
#> 23 720575940616398303 79729161744437991  left intrinsic     central       MBON
#> 24 720575940623750325 79729161744426207  left intrinsic     central       MBON
#>        cell_type        top_nt ito_lee_hemilineage hemibrain_type       fbbt_id
#> 1         MBON20          gaba    putative_primary         MBON20 FBbt_00111013
#> 2         MBON23 acetylcholine          DL1_dorsal         MBON23 FBbt_00049113
#> 3         MBON29 acetylcholine         DL2_ventral         MBON29 FBbt_00049849
#> 4         MBON28 acetylcholine          DL1_dorsal         MBON28 FBbt_00049848
#> 5         MBON27 acetylcholine    putative_primary         MBON27 FBbt_00049847
#> 6  MBON25,MBON34     glutamate              SMPad3  MBON25,MBON34          <NA>
#> 7  MBON25,MBON34     glutamate              SMPad3  MBON25,MBON34          <NA>
#> 8  MBON25,MBON34     glutamate              SMPad3  MBON25,MBON34          <NA>
#> 9         MBON24 acetylcholine    putative_primary         MBON24 FBbt_00049844
#> 10        MBON27 acetylcholine    putative_primary         MBON27 FBbt_00049847
#> 11        MBON28 acetylcholine          DL1_dorsal         MBON28 FBbt_00049848
#> 12        MBON21 acetylcholine    putative_primary         MBON21 FBbt_00111046
#> 13        MBON20          gaba    putative_primary         MBON20 FBbt_00111013
#> 14        MBON21 acetylcholine    putative_primary         MBON21 FBbt_00111046
#> 15        MBON23 acetylcholine          DL1_dorsal         MBON23 FBbt_00049113
#> 16        MBON22 acetylcholine    putative_primary         MBON22 FBbt_00100240
#> 17        MBON22 acetylcholine    putative_primary         MBON22 FBbt_00100240
#> 18        MBON26 acetylcholine    putative_primary         MBON26 FBbt_00049846
#> 19        MBON29 acetylcholine         DL2_ventral         MBON29 FBbt_00049849
#> 20        MBON26 acetylcholine    putative_primary         MBON26 FBbt_00049846
#> 21        MBON24 acetylcholine    putative_primary         MBON24 FBbt_00049844
#> 22 MBON25,MBON34     glutamate              SMPad3  MBON25,MBON34          <NA>
#> 23 MBON25,MBON34     glutamate              SMPad3  MBON25,MBON34          <NA>
#> 24 MBON25,MBON34     glutamate              SMPad3  MBON25,MBON34          <NA>
# the / introduces a regex query (small performance penalty, more flexible)
flytable_meta("/type:MBON2[0-5]")
#>               root_id     supervoxel_id  side      flow super_class cell_class
#> 1  720575940611344078 80223254782485510 right intrinsic     central       MBON
#> 2  720575940617567206 78180912067538056  left intrinsic     central       MBON
#> 3  720575940624659943 80011255129536552  left intrinsic     central       MBON
#> 4  720575940627144069 79166692827138268 right intrinsic     central       MBON
#> 5  720575940619810389 79518055445188833 right intrinsic     central       MBON
#> 6  720575940623464316 81136880426764424 right intrinsic     central       MBON
#> 7  720575940620715224 78111917579300796  left intrinsic     central       MBON
#> 8  720575940643863496 78324741798855874  left intrinsic     central       MBON
#> 9  720575940621777391 81137979736901319 right intrinsic     central       MBON
#> 10 720575940606953858 81136880426735755 right intrinsic     central       MBON
#> 11 720575940635063135 80082242348953693 right intrinsic     central       MBON
#> 12 720575940616463477 77830305364995511  left intrinsic     central       MBON
#> 13 720575940631177803 78814711734490195  left intrinsic     central       MBON
#> 14 720575940624489964 79166761480079751 right intrinsic     central       MBON
#> 15 720575940616398303 79729161744437991  left intrinsic     central       MBON
#> 16 720575940623750325 79729161744426207  left intrinsic     central       MBON
#>        cell_type        top_nt ito_lee_hemilineage hemibrain_type       fbbt_id
#> 1         MBON20          gaba    putative_primary         MBON20 FBbt_00111013
#> 2         MBON23 acetylcholine          DL1_dorsal         MBON23 FBbt_00049113
#> 3  MBON25,MBON34     glutamate              SMPad3  MBON25,MBON34          <NA>
#> 4  MBON25,MBON34     glutamate              SMPad3  MBON25,MBON34          <NA>
#> 5  MBON25,MBON34     glutamate              SMPad3  MBON25,MBON34          <NA>
#> 6         MBON24 acetylcholine    putative_primary         MBON24 FBbt_00049844
#> 7         MBON21 acetylcholine    putative_primary         MBON21 FBbt_00111046
#> 8         MBON20          gaba    putative_primary         MBON20 FBbt_00111013
#> 9         MBON21 acetylcholine    putative_primary         MBON21 FBbt_00111046
#> 10        MBON23 acetylcholine          DL1_dorsal         MBON23 FBbt_00049113
#> 11        MBON22 acetylcholine    putative_primary         MBON22 FBbt_00100240
#> 12        MBON22 acetylcholine    putative_primary         MBON22 FBbt_00100240
#> 13        MBON24 acetylcholine    putative_primary         MBON24 FBbt_00049844
#> 14 MBON25,MBON34     glutamate              SMPad3  MBON25,MBON34          <NA>
#> 15 MBON25,MBON34     glutamate              SMPad3  MBON25,MBON34          <NA>
#> 16 MBON25,MBON34     glutamate              SMPad3  MBON25,MBON34          <NA>
# }