fafbseg: Support functions for analysis of FAFB EM segmentation

Functions that read some of the raw data formats produced by Google electron microscopy segmentation tools as displayed by 'neuroglancer' and then allow basic 3D visualisation and analysis.

Package Options

fafbseg.skelziproot set to the location of a folder containing the zipped versions of the skeletonised segmentations. This will be used by read_segments, read_topn etc.

fafbseg.skeletonuri a brainmaps URI specifying a remote source used by read.neurons.brainmaps and brainmaps_skeleton to read neuronal skeletons.

fafbseg.sampleurl optionally set to a sample Neuroglancer URL that will modified to point to arbitrary locations by open_fafb_ngl.

fafbseg.zipdivisor see segmentid2zip.

See also

Examples

options()[grep("^fafbseg\\.", names(options()))]

#> $fafbseg.baseurl
#> [1] "https://neuroglancer-demo.appspot.com"
#> 
#> $fafbseg.brainmaps.meshName
#> [1] "mcws_quad1e6"
#> 
#> $fafbseg.brainmaps.volume
#> [1] "772153499790:fafb_v14:fafb_v14_16nm_v00c_split3xfill2"
#> 
#> $fafbseg.sampleurl
#> [1] "https://neuroglancer-demo.appspot.com/#!%7B%22layers%22:%7B%22fafb_v14_clahe%22:%7B%22source%22:%22brainmaps://772153499790:fafb_v14:fafb_v14_clahe%22,%22type%22:%22image%22%7D,%22fafb_v14_16nm_v00c_split3xfill2%22:%7B%22source%22:%22brainmaps://772153499790:fafb_v14:fafb_v14_16nm_v00c_split3xfill2%22,%22type%22:%22segmentation%22%7D%7D,%22navigation%22:%7B%22pose%22:%7B%22position%22:%7B%22voxelSize%22:[8,8,40],%22voxelCoordinates%22:[53323.1017,21251.8736,3282]%7D%7D,%22zoomFactor%22:8%7D,%22showAxisLines%22:false,%22perspectiveOrientation%22:[0.0214,0.4873,-0.505,-0.7121],%22perspectiveZoom%22:1024,%22showSlices%22:false%7D"
#> 
#> $fafbseg.skeletonuri
#> [1] "brainmaps://772153499790:fafb_v14:fafb_v14_16nm_v00c_split3xfill2_flatreseg2_skeletons32/teasar512_nnconn165_mc10000_prune10_thresh1000_sparse250"
#>