Find all merged segment ids for a given starting google segment id

Find all merged segment ids for a given starting google segment id

find_merged_segments(x)

Arguments

x
a segment id or any other input that can be interpreted by ngl_segments

Value

vector of segment ids (in ascending order)

Details

segment ids in ffn16reseg-ms1000_md0.02_c0.6_iou0.7 always match one raw segment id in fafb_v14_16nm_v00c_split3xfill2 but may map to a large agglomerated merge group.

Note that this functions depends on the segment_merge_groups.dt object from the fafbsegdata package.

Examples

find_merged_segments(7186840767)

#>  [1]  7186840767  7186844852  7188488573  7188488876  7506547847  7511521487
#>  [7]  7513174533  7523125878  7524769973  7528082928  7528082999  7528083090
#> [13]  7528087811  7844472994  7844473351  7866011642  7866015978  7869320680
#> [19]  8184049304  8184049546  8187380528  8193975632  8198954037  8515392737
#> [25]  8525318026  8525318518  8525318619  8525318881  8526948746  8535280040
#> [31]  8541870917  8543523319  8861569429  8866569630  8866569764  8869882415
#> [37]  8873165227  8873192134  8874821377  8876521709  8876543998  8878156810
#> [43]  8881456848  8883135153  9217782462  9219408058  9219408083  9221047258
#> [49]  9537450356  9539102248  9540758370  9540758400  9540758403  9554037005
#> [55]  9554041768  9555689874  9560632655  9562293635  9563945725  9875383416
#> [61]  9880358224  9885314546 10218295532 10233257420