Flytable database queries

flytable_query performs a SQL query against a flytable database. You can omit the base argument unless you have tables of the same name in different bases.

flytable_list_rows(
  table,
  base = NULL,
  view_name = NULL,
  order_by = NULL,
  desc = FALSE,
  start = 0L,
  limit = Inf,
  python = FALSE
)

flytable_query(
  sql,
  limit = 100000L,
  base = NULL,
  python = FALSE,
  convert = TRUE
)

Arguments

table

The name of a table inside your database

base

Character vector naming a seatable base (recommended) or a Base object returned by flytable_base (expert use).

view_name

An optional view which may limit the rows/columns displayed.

order_by

Optional name of columns to order results

desc

Whether to use descending order (default FALSE => ascending order)

start

Optional starting row

limit

An optional limit, which only applies if you do not specify a limit directly in the sql query. By default seatable limits SQL queries to 100 rows. We increase the limit to 100000 rows by default.

python

Whether to return a Python pandas DataFrame. The default of FALSE returns an R data.frame

sql

A SQL query string. See examples and seatable docs.

convert

Expert use only: Whether or not to allow the Python seatable module to process raw output from the database. This is is principally for debugging purposes. NB this imposes a requirement of seatable_api >=2.4.0.

Value

An R data.frame or Pandas DataFrame depending on the value of the python argument.

Details

Flytable uses programmatic access to the seatable API.

See also

Examples

# \donttest{
flytable_list_rows(table = "testfruit")
#>                        _id              _mtime              _ctime fruit_name
#> 1   H8BckTnXRL2PaVuLDcRPMA 2022-02-15 13:35:38 2021-12-17 13:36:17      apple
#> 2   SrBW3vFLRxafKRYHoPrGQQ 2022-02-15 13:35:38 2021-12-17 13:36:17     banana
#> 3   F1h-TZKpTcWiHujZ2xnF8Q 2022-02-15 13:35:38 2021-12-17 13:36:17 clementine
#> 4   XiRXr4GmQzq3SGZoAYiCRw 2022-02-15 13:35:38 2021-12-19 10:17:31       kiwi
#> 5   djnL1CoYQbmJ3PxnzJ6x2A 2022-02-15 13:35:38 2021-12-19 21:06:09       kiwi
#> 6   a1UXG6fJR1O0Tl8UjbAsiw 2022-02-15 13:35:38 2021-12-19 21:19:14       kiwi
#> 7   AzJoykOwTpezIPWFY7z9iA 2022-02-15 13:35:38 2021-12-29 12:52:05       kiwi
#> 8   BCDDhFjBTrGJztSms7Szsw 2022-02-15 13:35:38 2021-12-29 12:55:51       kiwi
#> 9   HUrk7IiISpCh9oGzh_9Owg 2022-02-15 13:35:38 2021-12-29 13:00:02       kiwi
#> 10  dugRpdkgQ6e4B46g3bhtRg 2022-02-15 13:35:38 2021-12-29 19:21:14       kiwi
#> 11  HoLyqPYVQdeFmr2LzLpNnw 2022-02-15 13:35:38 2021-12-29 19:25:24       kiwi
#> 12  LI9HBku-QHOFTNbVo4Juag 2022-02-15 13:35:38 2021-12-30 00:38:02       kiwi
#> 13  QxkH2piVT0yJIZvRLXbLUw 2022-02-15 13:35:38 2021-12-30 00:42:17       kiwi
#> 14  cyuW-E8VQs-idTXXOqz5CQ 2022-02-15 13:35:38 2021-12-30 00:50:28       kiwi
#> 15  HS6z29kzRvK9l3i8BO6oUQ 2022-02-15 13:35:38 2021-12-30 16:06:25       kiwi
#> 16  cS9g37jFQ06k11ySMQ7N9g 2022-02-15 13:35:38 2021-12-30 16:10:50       kiwi
#> 17  LeayZwlnTSOyT0sxG5N9YQ 2022-02-15 13:35:38 2021-12-30 16:14:01       kiwi
#> 18  X70fVFMBSquw-VWahpSMPg 2022-02-15 13:35:38 2021-12-30 16:43:25       kiwi
#> 19  JP2MyiPHT2yV7xK_q1fhbA 2022-02-15 13:35:38 2021-12-30 16:46:20       kiwi
#> 20  Ayd2jg3rRdmRLx01WrEAVw 2022-02-15 13:35:38 2021-12-30 16:47:10       kiwi
#> 21  LOqm56qQQ9KU5cbiMp93hg 2022-02-15 13:35:38 2021-12-30 16:47:36       kiwi
#> 22  SyeQt_ecRcGbUwQRZogRgA 2022-02-15 13:35:38 2021-12-30 16:50:13       kiwi
#> 23  Pl0llbdtS1CLpcuc426NXg 2022-02-15 13:35:38 2021-12-30 16:51:39       kiwi
#> 24  b8i2GZfDQ8C_pLK1VWL50w 2022-02-15 13:35:38 2021-12-30 16:57:41       kiwi
#> 25  aCl3lZSMRiS0oz7kXwoZqQ 2022-02-15 13:35:38 2021-12-30 17:00:41       kiwi
#> 26  KUsOwD93TWOQlNwPAWAdAA 2022-02-15 13:35:38 2021-12-30 17:01:43       kiwi
#> 27  boyjo9HfQn6MCdgzKHy4jA 2022-02-15 13:35:38 2022-01-05 13:28:10       kiwi
#> 28  Wx20-ZkiTWeQ_9cSJsHoXg 2022-02-15 13:35:38 2022-01-05 17:53:56       kiwi
#> 29  QknE1XgKSW2uwkk_MNODAA 2022-02-15 13:35:38 2022-01-05 18:08:59       kiwi
#> 30  KusYu2ooTSGj4QuKr1abLw 2022-02-15 13:35:38 2022-01-06 23:42:45       kiwi
#> 31  PfPJ-zRLRY2_4WVzf-4ycA 2022-02-15 13:35:38 2022-01-06 23:49:04       kiwi
#> 32  RwyOs2LGSIebNwdCd09rQA 2022-02-15 13:35:38 2022-01-06 23:49:59       kiwi
#> 33  YoWiwZ_JTC6kTK_1gov_QA 2022-02-15 13:35:38 2022-01-06 23:53:13       kiwi
#> 34  I0duq8v9R5-0n345cMoC0w 2022-02-15 13:35:38 2022-01-07 00:03:32       kiwi
#> 35  JvGmt9BpRnOMuyOi7j5N_g 2022-02-15 13:35:38 2022-01-07 00:03:35       kiwi
#> 36  Dd-8i5cXSl6u6dyoLVTymg 2022-02-15 13:35:38 2022-01-07 09:17:56       kiwi
#> 37  ZKEGXA-lTsCyyER3Y07_iA 2022-02-15 13:35:38 2022-01-07 09:22:38       kiwi
#> 38  CylulFqQSj6wuYi1IxAoPQ 2022-02-15 13:35:38 2022-01-07 10:15:07       kiwi
#> 39  MkCVhMmpTSq1AhmINYRxUA 2022-02-15 13:35:38 2022-01-07 10:19:50       kiwi
#> 40  IzEFFHW1Tm6aR9BLU7JkCQ 2022-02-15 13:35:38 2022-01-07 10:26:51       kiwi
#> 41  DUzIPNTjTc-NuRBa2fu9mw 2022-02-15 13:35:38 2022-01-13 10:46:36       kiwi
#> 42  Ys2_3fyiQKiQ1qtCC_HMNw 2022-02-15 13:35:38 2022-01-13 10:56:15       kiwi
#> 43  bL6t_iDbT0OjYl2YCGTE7Q 2022-02-15 13:35:38 2022-01-13 10:57:07       kiwi
#> 44  Yaq7iSivTvm3i5F2wTFU1Q 2022-02-15 13:35:38 2022-01-13 11:01:14       kiwi
#> 45  eP1o-7OGTtm-mAIaeVUAoQ 2022-02-15 13:35:38 2022-01-13 11:01:41       kiwi
#> 46  FR6t-JEcRhyY9AEhjI7NrA 2022-02-15 13:35:38 2022-01-13 11:10:09       kiwi
#> 47  eA80dN5jQrOXvg2UvG06eQ 2022-02-15 13:35:38 2022-01-13 19:59:09       kiwi
#> 48  ZCLGAXeJR6S1uu-ZdUYFOw 2022-02-15 13:35:38 2022-01-13 20:03:56       kiwi
#> 49  TBCsFQ6QRf2cVtdA5d6qUA 2022-02-15 13:35:38 2022-01-13 20:13:59       kiwi
#> 50  FBVfbHULQ7Wp_JyP8Xe_Jg 2022-02-15 13:35:38 2022-01-15 22:37:19       kiwi
#> 51  P-wiCZ_-QHq4Nf4pK8LBIA 2022-02-15 13:35:38 2022-01-16 00:52:45       kiwi
#> 52  Rtv9eZanR9Gs0KhgtymgNg 2022-02-15 13:35:38 2022-01-16 01:29:07       kiwi
#> 53  UdUl9w0jTVOqDvBXTu77Kw 2022-02-15 13:35:38 2022-01-16 01:40:28       kiwi
#> 54  RdHHk0pnSLyeyVo2BxeF5w 2022-02-15 13:35:38 2022-01-16 01:45:32       kiwi
#> 55  IrWKOJWQRRi3PON9oGAYUA 2022-02-15 13:35:38 2022-01-16 01:57:57       kiwi
#> 56  I3f9OTPLSPOOwBtGzbtiKA 2022-02-15 13:35:38 2022-01-16 09:29:05       kiwi
#> 57  POy7nJgzSQa_xhdpzUHydg 2022-02-15 13:35:38 2022-01-16 09:33:34       kiwi
#> 58  ZbJKRYo2SGmVQdIliw79ng 2022-02-15 13:35:38 2022-01-16 12:21:06       kiwi
#> 59  E6zj9ZxaSyeowfgDpY2UKw 2022-02-15 13:35:38 2022-01-16 16:22:53       kiwi
#> 60  fxJw51f8SjmBEJ551o5ZmQ 2022-02-15 13:35:38 2022-01-16 16:27:59       kiwi
#> 61  VR6kgtqxTZObuneVx-cOwg 2022-02-15 13:35:38 2022-01-16 16:36:55       kiwi
#> 62  bGFKR02RT-2f02J7-wMOag 2022-02-15 13:35:38 2022-01-16 16:59:35       kiwi
#> 63  e71Y8iLrQ-q3qUPhAuNRIw 2022-02-15 13:35:38 2022-01-16 17:43:22       kiwi
#> 64  ISlaXp5ZSV23a_gHS8kU3w 2022-02-15 13:35:38 2022-01-16 17:48:21       kiwi
#> 65  HWHmpvfOSIKyw_hr_wjzjQ 2022-02-15 13:35:38 2022-01-16 19:11:42       kiwi
#> 66  PmesHme7Qnyg_OEve1nuLA 2022-02-15 13:35:38 2022-01-16 19:16:20       kiwi
#> 67  TLxj72TKR5C34YwsCTxggA 2022-02-15 13:35:38 2022-01-16 19:22:22       kiwi
#> 68  JRs0cgDsR4WsjgICugq9OQ 2022-02-15 13:35:38 2022-01-16 20:01:23       kiwi
#> 69  MWH3m1hHQFil39DqoeIJcQ 2022-02-15 13:35:38 2022-01-16 20:06:52       kiwi
#> 70  P9F8rNXYQKyUNYG6ZRzsCA 2022-02-15 13:35:38 2022-01-16 20:12:05       kiwi
#> 71  Aus7VN-vTZGgTVrI4ojR9Q 2022-02-15 13:35:38 2022-01-16 21:36:41       kiwi
#> 72  EDi2LNmAQLSlH89-qEPrJg 2022-02-15 13:35:38 2022-01-16 21:39:37       kiwi
#> 73  KG4-_JxbRs-AWhNiP01qEw 2022-02-15 13:35:38 2022-01-16 21:41:31       kiwi
#> 74  QPlyJ3iaQUmiCHJ827RPbA 2022-02-15 13:35:38 2022-01-17 10:06:14       kiwi
#> 75  Sm5SkKARTOOp3AFKZjuQ5A 2022-02-15 13:35:38 2022-01-17 10:11:38       kiwi
#> 76  LgvLx2VVT4yw5l03SR6-9g 2022-02-15 13:35:38 2022-01-17 10:31:33       kiwi
#> 77  QgZn62lQTiegMjB-T97v8w 2022-02-15 13:35:38 2022-01-18 18:24:30       kiwi
#> 78  Nm7_SQBHRpmvFj-F1cb_ew 2022-02-15 13:35:38 2022-01-18 18:29:59       kiwi
#> 79  IVOgfW_4T8C3ZTQE3cP9nA 2022-02-15 13:35:38 2022-01-18 18:39:34       kiwi
#> 80  LnVHPTtZRCmRyoEa4FVmew 2022-02-15 13:35:38 2022-01-21 08:24:21       kiwi
#> 81  UzfCGg-wRd2WTXem_zRE3A 2022-02-15 13:35:38 2022-01-21 08:30:03       kiwi
#> 82  Ev-gKb2WRCSAT47CP1_SDQ 2022-02-15 13:35:38 2022-01-21 09:10:54       kiwi
#> 83  JWp-Y9QUS8G2YZvJOopgXw 2022-02-15 13:35:38 2022-01-21 09:16:30       kiwi
#> 84  FZSe6rpzTKGW3A01Myu5AA 2022-02-15 13:35:38 2022-01-21 09:26:09       kiwi
#> 85  UZQypYzgQ66MBvYtKAc1-A 2022-02-15 13:35:38 2022-01-21 22:01:26       kiwi
#> 86  AlzVhpSmQk-1PCLAeT-gKw 2022-02-15 13:35:38 2022-01-21 22:34:12       kiwi
#> 87  E1aMIlihRYCniWYyt8gBww 2022-02-15 13:35:38 2022-01-21 22:39:44       kiwi
#> 88  eE4NYlwlRsayQUeeUCaOIQ 2022-02-15 13:35:38 2022-01-21 22:56:12       kiwi
#> 89  dhz_xtLUTMaMkWthGQYzaQ 2022-02-15 13:35:38 2022-01-27 16:24:54       kiwi
#> 90  N0PQ0x2QSI6NydRtfUrkPg 2022-02-15 13:35:38 2022-01-27 16:30:51       kiwi
#> 91  NGb5046FSFuRTfXjo_2xGw 2022-02-15 13:35:38 2022-01-27 16:38:17       kiwi
#> 92  ZhbU5CAHRzanDRUd51h6qg 2022-02-15 13:35:38 2022-02-07 10:07:07       kiwi
#> 93  WF-X4oLiQUa2z5dZ9KviGg 2022-02-15 13:35:38 2022-02-09 18:48:42       kiwi
#> 94  SQ5iMjreTdKM_H2ME7JAKg 2022-02-15 13:35:38 2022-02-10 03:49:39       kiwi
#> 95  dcUrtfxBR9aH42k9x4Nl4w 2022-02-15 13:35:38 2022-02-10 04:00:40       kiwi
#> 96  dI0HmgkzSj26YgHouGOggA 2022-02-15 13:35:38 2022-02-10 10:34:48       kiwi
#> 97  SB_O-AAJTvCvAiqWIrxAlA 2022-02-15 13:35:38 2022-02-10 10:40:16       kiwi
#> 98  b4vNCIj0R5a4zbG_PxLIFw 2022-02-15 13:35:38 2022-02-10 20:02:02       kiwi
#> 99  Jd6gQ0D9S1ew1LyhiOzheA 2022-02-15 13:35:38 2022-02-10 20:07:16       kiwi
#> 100 Sl26iC3pSCuG_W_jxk2tNw 2022-02-15 13:35:38 2022-02-10 20:16:44       kiwi
#> 101 WdHHmoJaS-WMGvEeZbQplg 2022-02-15 13:35:38 2022-02-13 19:12:46       kiwi
#> 102 VDs7pdJfS76cPDa3piu_hA 2022-02-15 13:35:38 2022-02-13 19:22:26       kiwi
#> 103 VDvXuJgmRh6AyjNtxo0PNg 2022-02-15 13:35:38 2022-02-13 21:08:27       kiwi
#> 104 NkeSKiFWSXuA6yeG_Lyqrw 2022-02-15 13:35:38 2022-02-14 14:11:03       kiwi
#> 105 V6wbn7xfQxWszT9w4_hLRQ 2022-02-15 13:35:38 2022-02-14 18:03:44       kiwi
#> 106 MDKKr2YaRlSyy7vlHq5ELw 2022-02-15 13:35:38 2022-02-14 22:02:36       kiwi
#> 107 L_LKJwoKTLmIH4lqp8hK8g 2022-02-15 13:35:38 2022-02-14 22:04:21       kiwi
#> 108 I6u7VTWqQpe3fZCQIq7qKA 2022-02-15 13:35:38 2022-02-14 22:08:10       kiwi
#> 109 M5YBoo6_TbapT2oOCvl3jg 2022-02-15 13:35:38 2022-02-15 04:55:27       kiwi
#> 110 TL35EHbiTxWRj8daQU09YQ 2022-02-15 13:35:38 2022-02-15 04:59:39       kiwi
#> 111 aJ824pzFSSOLltQZeEg_7g 2022-02-15 13:35:38 2022-02-15 05:00:53       kiwi
#> 112 E0L4TZIaTl25duOz1ZNiGg 2022-02-15 13:35:38 2022-02-15 05:05:56       kiwi
#> 113 QZi1WlHhR_KLhf7m2wUrcg 2022-02-15 13:35:38 2022-02-15 05:08:06       kiwi
#> 114 d0qsCXH7Q5qJBz0oouu-fw 2022-02-15 13:35:38 2022-02-15 11:06:10       kiwi
#> 115 HCbxGzFaRO6sp0Oef1nF2Q 2022-02-15 13:35:38 2022-02-15 11:06:13       kiwi
#> 116 NRKN4TUBQwmEXkmsITkDQQ 2022-02-15 13:35:38 2022-02-15 11:11:36       kiwi
#> 117 Tx23AkxkTqO03-ZxC5kD1Q 2022-02-15 13:35:38 2022-02-15 11:12:05       kiwi
#> 118 JRtUH_oHTjah7VChpdwT1g 2022-02-15 13:35:38 2022-02-15 11:14:50       kiwi
#> 119 GaFGtyr4RBqTbmNx_0n8OA 2022-02-15 13:35:38 2022-02-15 11:18:16       kiwi
#>     nid              person       last_modified date_nominute
#> 1     1               Alice 2022-02-15 13:35:38    2022-01-06
#> 2     2                 Bob 2022-02-15 13:35:38          <NA>
#> 3     3               Clara 2022-02-15 13:35:38          <NA>
#> 4     6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 5     6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 6     6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 7     6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 8     6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 9     6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 10    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 11    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 12    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 13    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 14    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 15    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 16    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 17    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 18    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 19    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 20    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 21    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 22    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 23    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 24    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 25    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 26    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 27    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 28    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 29    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 30    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 31    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 32    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 33    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 34    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 35    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 36    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 37    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 38    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 39    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 40    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 41    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 42    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 43    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 44    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 45    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 46    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 47    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 48    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 49    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 50    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 51    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 52    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 53    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 54    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 55    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 56    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 57    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 58    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 59    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 60    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 61    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 62    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 63    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 64    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 65    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 66    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 67    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 68    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 69    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 70    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 71    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 72    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 73    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 74    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 75    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 76    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 77    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 78    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 79    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 80    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 81    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 82    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 83    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 84    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 85    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 86    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 87    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 88    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 89    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 90    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 91    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 92    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 93    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 94    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 95    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 96    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 97    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 98    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 99    6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 100   6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 101   6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 102   6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 103   6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 104   6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 105   6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 106   6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 107   6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 108   6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 109   6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 110   6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 111   6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 112   6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 113   6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 114   6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 115   6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 116   6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 117   6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 118   6 Frederick the Great 2022-02-15 13:35:38          <NA>
#> 119   6 Frederick the Great 2022-02-15 13:35:38          <NA>
#>            date_wminute                                        user
#> 1   2022-01-12 09:30:00 8adf4f5dd661449fa6cc1f5a0b1815c0@auth.local
#> 2   2022-01-03 07:56:00 c7efb8019da54923a9b04d4a74f0fde8@auth.local
#> 3   2021-08-05 08:30:00                                         NaN
#> 4                  <NA>                                         NaN
#> 5                  <NA>                                         NaN
#> 6                  <NA>                                         NaN
#> 7                  <NA>                                         NaN
#> 8                  <NA>                                         NaN
#> 9                  <NA>                                         NaN
#> 10                 <NA>                                         NaN
#> 11                 <NA>                                         NaN
#> 12                 <NA>                                         NaN
#> 13                 <NA>                                         NaN
#> 14                 <NA>                                         NaN
#> 15                 <NA>                                         NaN
#> 16                 <NA>                                         NaN
#> 17                 <NA>                                         NaN
#> 18                 <NA>                                         NaN
#> 19                 <NA>                                         NaN
#> 20                 <NA>                                         NaN
#> 21                 <NA>                                         NaN
#> 22                 <NA>                                         NaN
#> 23                 <NA>                                         NaN
#> 24                 <NA>                                         NaN
#> 25                 <NA>                                         NaN
#> 26                 <NA>                                         NaN
#> 27                 <NA>                                         NaN
#> 28                 <NA>                                         NaN
#> 29                 <NA>                                         NaN
#> 30                 <NA>                                         NaN
#> 31                 <NA>                                         NaN
#> 32                 <NA>                                         NaN
#> 33                 <NA>                                         NaN
#> 34                 <NA>                                         NaN
#> 35                 <NA>                                         NaN
#> 36                 <NA>                                         NaN
#> 37                 <NA>                                         NaN
#> 38                 <NA>                                         NaN
#> 39                 <NA>                                         NaN
#> 40                 <NA>                                         NaN
#> 41                 <NA>                                         NaN
#> 42                 <NA>                                         NaN
#> 43                 <NA>                                         NaN
#> 44                 <NA>                                         NaN
#> 45                 <NA>                                         NaN
#> 46                 <NA>                                         NaN
#> 47                 <NA>                                         NaN
#> 48                 <NA>                                         NaN
#> 49                 <NA>                                         NaN
#> 50                 <NA>                                         NaN
#> 51                 <NA>                                         NaN
#> 52                 <NA>                                         NaN
#> 53                 <NA>                                         NaN
#> 54                 <NA>                                         NaN
#> 55                 <NA>                                         NaN
#> 56                 <NA>                                         NaN
#> 57                 <NA>                                         NaN
#> 58                 <NA>                                         NaN
#> 59                 <NA>                                         NaN
#> 60                 <NA>                                         NaN
#> 61                 <NA>                                         NaN
#> 62                 <NA>                                         NaN
#> 63                 <NA>                                         NaN
#> 64                 <NA>                                         NaN
#> 65                 <NA>                                         NaN
#> 66                 <NA>                                         NaN
#> 67                 <NA>                                         NaN
#> 68                 <NA>                                         NaN
#> 69                 <NA>                                         NaN
#> 70                 <NA>                                         NaN
#> 71                 <NA>                                         NaN
#> 72                 <NA>                                         NaN
#> 73                 <NA>                                         NaN
#> 74                 <NA>                                         NaN
#> 75                 <NA>                                         NaN
#> 76                 <NA>                                         NaN
#> 77                 <NA>                                         NaN
#> 78                 <NA>                                         NaN
#> 79                 <NA>                                         NaN
#> 80                 <NA>                                         NaN
#> 81                 <NA>                                         NaN
#> 82                 <NA>                                         NaN
#> 83                 <NA>                                         NaN
#> 84                 <NA>                                         NaN
#> 85                 <NA>                                         NaN
#> 86                 <NA>                                         NaN
#> 87                 <NA>                                         NaN
#> 88                 <NA>                                         NaN
#> 89                 <NA>                                         NaN
#> 90                 <NA>                                         NaN
#> 91                 <NA>                                         NaN
#> 92                 <NA>                                         NaN
#> 93                 <NA>                                         NaN
#> 94                 <NA>                                         NaN
#> 95                 <NA>                                         NaN
#> 96                 <NA>                                         NaN
#> 97                 <NA>                                         NaN
#> 98                 <NA>                                         NaN
#> 99                 <NA>                                         NaN
#> 100                <NA>                                         NaN
#> 101                <NA>                                         NaN
#> 102                <NA>                                         NaN
#> 103                <NA>                                         NaN
#> 104                <NA>                                         NaN
#> 105                <NA>                                         NaN
#> 106                <NA>                                         NaN
#> 107                <NA>                                         NaN
#> 108                <NA>                                         NaN
#> 109                <NA>                                         NaN
#> 110                <NA>                                         NaN
#> 111                <NA>                                         NaN
#> 112                <NA>                                         NaN
#> 113                <NA>                                         NaN
#> 114                <NA>                                         NaN
#> 115                <NA>                                         NaN
#> 116                <NA>                                         NaN
#> 117                <NA>                                         NaN
#> 118                <NA>                                         NaN
#> 119                <NA>                                         NaN
# }
# \donttest{
flytable_query("SELECT person, fruit_name FROM testfruit WHERE person!='Bob'")
#>                  person fruit_name
#> 1                 Alice      apple
#> 2                 Clara clementine
#> 3   Frederick the Great       kiwi
#> 4   Frederick the Great       kiwi
#> 5   Frederick the Great       kiwi
#> 6   Frederick the Great       kiwi
#> 7   Frederick the Great       kiwi
#> 8   Frederick the Great       kiwi
#> 9   Frederick the Great       kiwi
#> 10  Frederick the Great       kiwi
#> 11  Frederick the Great       kiwi
#> 12  Frederick the Great       kiwi
#> 13  Frederick the Great       kiwi
#> 14  Frederick the Great       kiwi
#> 15  Frederick the Great       kiwi
#> 16  Frederick the Great       kiwi
#> 17  Frederick the Great       kiwi
#> 18  Frederick the Great       kiwi
#> 19  Frederick the Great       kiwi
#> 20  Frederick the Great       kiwi
#> 21  Frederick the Great       kiwi
#> 22  Frederick the Great       kiwi
#> 23  Frederick the Great       kiwi
#> 24  Frederick the Great       kiwi
#> 25  Frederick the Great       kiwi
#> 26  Frederick the Great       kiwi
#> 27  Frederick the Great       kiwi
#> 28  Frederick the Great       kiwi
#> 29  Frederick the Great       kiwi
#> 30  Frederick the Great       kiwi
#> 31  Frederick the Great       kiwi
#> 32  Frederick the Great       kiwi
#> 33  Frederick the Great       kiwi
#> 34  Frederick the Great       kiwi
#> 35  Frederick the Great       kiwi
#> 36  Frederick the Great       kiwi
#> 37  Frederick the Great       kiwi
#> 38  Frederick the Great       kiwi
#> 39  Frederick the Great       kiwi
#> 40  Frederick the Great       kiwi
#> 41  Frederick the Great       kiwi
#> 42  Frederick the Great       kiwi
#> 43  Frederick the Great       kiwi
#> 44  Frederick the Great       kiwi
#> 45  Frederick the Great       kiwi
#> 46  Frederick the Great       kiwi
#> 47  Frederick the Great       kiwi
#> 48  Frederick the Great       kiwi
#> 49  Frederick the Great       kiwi
#> 50  Frederick the Great       kiwi
#> 51  Frederick the Great       kiwi
#> 52  Frederick the Great       kiwi
#> 53  Frederick the Great       kiwi
#> 54  Frederick the Great       kiwi
#> 55  Frederick the Great       kiwi
#> 56  Frederick the Great       kiwi
#> 57  Frederick the Great       kiwi
#> 58  Frederick the Great       kiwi
#> 59  Frederick the Great       kiwi
#> 60  Frederick the Great       kiwi
#> 61  Frederick the Great       kiwi
#> 62  Frederick the Great       kiwi
#> 63  Frederick the Great       kiwi
#> 64  Frederick the Great       kiwi
#> 65  Frederick the Great       kiwi
#> 66  Frederick the Great       kiwi
#> 67  Frederick the Great       kiwi
#> 68  Frederick the Great       kiwi
#> 69  Frederick the Great       kiwi
#> 70  Frederick the Great       kiwi
#> 71  Frederick the Great       kiwi
#> 72  Frederick the Great       kiwi
#> 73  Frederick the Great       kiwi
#> 74  Frederick the Great       kiwi
#> 75  Frederick the Great       kiwi
#> 76  Frederick the Great       kiwi
#> 77  Frederick the Great       kiwi
#> 78  Frederick the Great       kiwi
#> 79  Frederick the Great       kiwi
#> 80  Frederick the Great       kiwi
#> 81  Frederick the Great       kiwi
#> 82  Frederick the Great       kiwi
#> 83  Frederick the Great       kiwi
#> 84  Frederick the Great       kiwi
#> 85  Frederick the Great       kiwi
#> 86  Frederick the Great       kiwi
#> 87  Frederick the Great       kiwi
#> 88  Frederick the Great       kiwi
#> 89  Frederick the Great       kiwi
#> 90  Frederick the Great       kiwi
#> 91  Frederick the Great       kiwi
#> 92  Frederick the Great       kiwi
#> 93  Frederick the Great       kiwi
#> 94  Frederick the Great       kiwi
#> 95  Frederick the Great       kiwi
#> 96  Frederick the Great       kiwi
#> 97  Frederick the Great       kiwi
#> 98  Frederick the Great       kiwi
#> 99  Frederick the Great       kiwi
#> 100 Frederick the Great       kiwi
#> 101 Frederick the Great       kiwi
#> 102 Frederick the Great       kiwi
#> 103 Frederick the Great       kiwi
#> 104 Frederick the Great       kiwi
#> 105 Frederick the Great       kiwi
#> 106 Frederick the Great       kiwi
#> 107 Frederick the Great       kiwi
#> 108 Frederick the Great       kiwi
#> 109 Frederick the Great       kiwi
#> 110 Frederick the Great       kiwi
#> 111 Frederick the Great       kiwi
#> 112 Frederick the Great       kiwi
#> 113 Frederick the Great       kiwi
#> 114 Frederick the Great       kiwi
#> 115 Frederick the Great       kiwi
#> 116 Frederick the Great       kiwi
#> 117 Frederick the Great       kiwi
#> 118 Frederick the Great       kiwi
# }
if (FALSE) {
flytable_query(paste("SELECT root_id, supervoxel_id FROM info limit 5"))
}