Flytable database queries

flytable_query performs a SQL query against a flytable database. You can omit the base argument unless you have tables of the same name in different bases.

flytable_list_rows(
  table,
  base = NULL,
  view_name = NULL,
  order_by = NULL,
  desc = FALSE,
  start = NULL,
  limit = Inf,
  python = FALSE
)

flytable_query(sql, limit = 100000L, base = NULL, python = FALSE)

Arguments

table
The name of a table inside your database

base
Character vector naming a seatable base (recommended) or a Base object returned by flytable_base (expert use).

view_name
An optional view which may limit the rows/columns displayed.

order_by
Optional name of columns to order results

desc
Whether to use descending order (default FALSE => ascending order)

start
Optional starting row

limit
An optional limit, which only applies if you do not specify a limit directly in the sql query. By default seatable limits SQL queries to 100 rows. We increase the limit to 100000 rows by default.

python
Whether to return a Python pandas DataFrame. The default of FALSE returns an R data.frame

sql
A SQL query string. See examples and seatable docs.

Value

An R data.frame or Pandas DataFrame depending on the value of the python argument.

Details

Flytable uses programmatic access to the seatable API.

See also

Examples

# \donttest{
flytable_list_rows(table = "testfruit")

#>                      _id                        _mtime
#> 1 CaBjlFo5R5ivWDhYbo_GGg 2021-10-29T20:01:56.611+00:00
#> 2 AqAT6V0xRU2Il6X_2WhFpg 2021-10-29T20:01:56.611+00:00
#> 3 QlhNQScFQhOa4cSjcC0cxA 2021-10-29T20:01:56.611+00:00
#>                          _ctime fruit_name person nid
#> 1 2021-10-15T17:50:19.038+00:00     banana  Alice   1
#> 2 2021-10-15T17:50:26.233+00:00      mango    Bob   2
#> 3 2021-10-16T19:39:35.348+00:00    rhubarb Claire   3

# }
# \donttest{
flytable_query("SELECT person, fruit_name FROM testfruit WHERE person!='Bob'")

#> # A tibble: 2 × 3
#>   fruit_name `_id`                  person
#>   <chr>      <chr>                  <chr> 
#> 1 banana     CaBjlFo5R5ivWDhYbo_GGg Alice 
#> 2 rhubarb    QlhNQScFQhOa4cSjcC0cxA Claire

# }
if (FALSE) {
flytable_query("select FLYWIREsvid, hemibrain_match FROM fafb_hemilineages_survey WHERE hemibrain_match!='' limit 5", base="hemilineages")
}