Flytable database queries

flytable_query performs a SQL query against a flytable database. You can omit the base argument unless you have tables of the same name in different bases.

flytable_list_rows(
  table,
  base = NULL,
  view_name = NULL,
  order_by = NULL,
  desc = FALSE,
  start = 0L,
  limit = Inf,
  python = FALSE,
  chunksize = NULL
)

flytable_query(
  sql,
  limit = 100000L,
  base = NULL,
  python = FALSE,
  convert = TRUE
)

Arguments

table

The name of a table inside your database

base

Character vector naming a seatable base (recommended) or a Base object returned by flytable_base (expert use).

view_name

An optional view which may limit the rows/columns displayed.

order_by

Optional name of columns to order results

desc

Whether to use descending order (default FALSE => ascending order)

start

Optional starting row

limit

An optional limit, which only applies if you do not specify a limit directly in the sql query. By default seatable limits SQL queries to 100 rows. We increase the limit to 100000 rows by default.

python

Whether to return a Python pandas DataFrame. The default of FALSE returns an R data.frame

chunksize

Optional The maximum number of rows to request in one web request. For advanced use only as the default value of NULL will fetch as many as possible.

sql

A SQL query string. See examples and seatable docs.

convert

Expert use only: Whether or not to allow the Python seatable module to process raw output from the database. This is is principally for debugging purposes. NB this imposes a requirement of seatable_api >=2.4.0.

Value

An R data.frame or Pandas DataFrame depending on the value of the python argument.

a data.frame of results. There should be 0 rows if no rows matched query.

Details

Flytable uses programmatic access to the seatable API.

See also

Examples

# \donttest{
flytable_list_rows(table = "testfruit")
#>                       _id              _mtime              _ctime fruit_name
#> 1  H8BckTnXRL2PaVuLDcRPMA 2022-05-12 16:58:07 2021-12-17 13:36:17      apple
#> 2  SrBW3vFLRxafKRYHoPrGQQ 2022-05-12 16:58:08 2021-12-17 13:36:17     banana
#> 3  F1h-TZKpTcWiHujZ2xnF8Q 2022-05-12 16:58:08 2021-12-17 13:36:17 clementine
#> 4  KZrMyoJKTeS_o9Pgynu7YQ 2022-09-14 09:54:33 2022-05-11 00:34:27       kiwi
#> 5  cI4fqlOUQK2A1jZKZvFpdg 2022-05-12 16:58:09 2022-05-11 00:39:59       kiwi
#> 6  AOUcHHZ0RTC-auONZURx4g 2022-05-12 16:58:10 2022-05-11 00:40:57       kiwi
#> 7  dxs7n1NpRM-hWoGJxPZpKA 2022-05-12 16:58:10 2022-05-11 00:41:17       kiwi
#> 8  J-JrluKcRzmYjgKCtMEAKw 2022-05-12 16:58:11 2022-05-11 00:41:25       kiwi
#> 9  aMmi4iegRNCdpY39-iAfdg 2022-05-12 16:58:11 2022-05-11 01:00:47       kiwi
#> 10 EH-l0qAyQyyJO8UpNYSj8w 2022-05-12 16:58:12 2022-05-11 01:07:10       kiwi
#> 11 E6uAS8ehQyid0sSj1bD-HA 2022-05-12 16:58:12 2022-05-11 20:52:00       kiwi
#> 12 P83tySHzT-iFfbwvlhJU7A 2022-05-12 16:58:13 2022-05-11 21:18:45       kiwi
#> 13 GBCn8fkaTnidSg8Nh7CLkQ 2022-05-12 16:58:13 2022-05-11 21:25:28       kiwi
#> 14 DHSK4OmSRYiNHFn6yHWEjQ 2022-05-12 16:58:13 2022-05-11 21:29:40       kiwi
#> 15 CVL59Pt2Ta6BWXXmEmLFSw 2022-05-12 16:58:14 2022-05-11 21:34:07       kiwi
#> 16 DHq2ty1nQn-LolwKbR5D9A 2022-05-12 16:58:14 2022-05-11 21:34:22       kiwi
#> 17 Y4aoKGjIQU6WskMCll9LtA 2022-05-12 16:58:15 2022-05-12 16:56:52       kiwi
#> 18 L94KQt88S6-cPBQSSk8IDg 2022-05-12 16:58:15 2022-05-12 16:58:15       kiwi
#> 19 JgCx-eJfQI22f1ynv_G4aA 2022-05-12 18:16:18 2022-05-12 18:16:18       kiwi
#> 20 HWP0B4R2TVinR51ZEiCh1Q 2022-05-12 18:33:47 2022-05-12 18:33:47       kiwi
#> 21 frinTnwQS_yT-p-6bUdUjg 2022-05-12 18:49:36 2022-05-12 18:49:36       kiwi
#> 22 Q79Qb0tMSi-NT__KJOXdnw 2022-05-12 19:03:47 2022-05-12 19:03:47       kiwi
#> 23 e63mt6sxSCSHK-AqkyE7rQ 2022-05-12 20:22:02 2022-05-12 20:22:02       kiwi
#> 24 e3AH4HFxTHiIILhWAgVfnQ 2022-05-12 20:41:57 2022-05-12 20:41:57       kiwi
#> 25 LIbNUlKpQhq3pHuqs_GG2A 2022-05-13 13:10:55 2022-05-13 13:10:55       kiwi
#> 26 Jox91CBQQIq8zv58pjE-gA 2022-05-13 13:11:49 2022-05-13 13:11:49       kiwi
#> 27 Mm7xFfvlQ6Ohb3FFA7-y3g 2022-05-13 13:12:24 2022-05-13 13:12:24       kiwi
#> 28 SY5dA23CSfCqFnePxJP1Ng 2022-05-13 13:13:31 2022-05-13 13:13:31       kiwi
#> 29 c-9ONdxnTsyIWgMx8_Qo5w 2022-05-13 14:15:29 2022-05-13 14:15:29       kiwi
#> 30 NvptrkSgS36T-Q6TBbyioA 2022-05-13 14:20:30 2022-05-13 14:20:30       kiwi
#> 31 PXX2Ay_XRwSOkVrJqBftsA 2022-05-13 14:21:21 2022-05-13 14:21:21       kiwi
#> 32 ET9-W5hxRdOoeEr54TgfZw 2022-05-13 14:26:16 2022-05-13 14:26:16       kiwi
#> 33 GBy5d_8jTYyV1RUEwpAwfQ 2022-05-13 14:32:10 2022-05-13 14:32:10       kiwi
#> 34 UvG5kHD7ST6m4yj5-xk2QA 2022-05-22 07:43:19 2022-05-22 07:43:19       kiwi
#> 35 SI2RldvDSsmadBqYrCqaqQ 2022-05-22 07:47:43 2022-05-22 07:47:43       kiwi
#> 36 MNPFKn5zRRKMHq66LQNWiA 2022-05-22 07:51:06 2022-05-22 07:51:06       kiwi
#> 37 OTsRNBFrS3mtawfKiREmrw 2022-05-29 06:31:28 2022-05-29 06:31:28       kiwi
#> 38 Nv8oP_hUSvyHgws5DQ0ZGA 2022-05-29 10:30:15 2022-05-29 10:30:15       kiwi
#> 39 OrEYIA_bQDCx6v-MkTuDYw 2022-05-29 10:34:54 2022-05-29 10:34:54       kiwi
#> 40 BhzO_0xqQsi41eCEKsAjGQ 2022-05-29 10:38:00 2022-05-29 10:38:00       kiwi
#> 41 BNeHPQNnSyuMvAEd3Kh6MQ 2022-05-29 20:33:27 2022-05-29 20:33:27       kiwi
#> 42 aVoVAUuWTFOtoUhQvOd_bg 2022-05-29 20:40:41 2022-05-29 20:40:41       kiwi
#> 43 BYauV5OwTa2tOk9JWW1N_A 2022-05-29 20:42:10 2022-05-29 20:42:10       kiwi
#> 44 DR3lfQrkQmu_tUcwlzx6Aw 2022-05-29 20:45:57 2022-05-29 20:45:57       kiwi
#> 45 JVksPUCFSFCD4pF__s5E2A 2022-06-06 14:40:12 2022-06-06 14:40:12       kiwi
#> 46 T4IY3mh6Sniovww6Yg3OWQ 2022-06-06 14:44:59 2022-06-06 14:44:59       kiwi
#> 47 TQqr1_nAQv27Qm3p1GXIMQ 2022-06-06 14:50:54 2022-06-06 14:50:54       kiwi
#> 48 c87JLpHXS5CRtj1YtXpL0g 2022-07-01 20:06:17 2022-07-01 20:06:17       kiwi
#> 49 AirzCY7yRS26fFBU_F6j5g 2022-07-01 20:33:14 2022-07-01 20:33:14       kiwi
#> 50 fRamo6FETfqaProIZuRbaA 2022-07-01 20:41:09 2022-07-01 20:41:09       kiwi
#> 51 Uy0XHWAhQjqJ2QtM-rDj3A 2022-07-01 20:42:29 2022-07-01 20:42:29       kiwi
#> 52 EcabRY2tSZyxjT9dztNVhQ 2022-07-01 21:11:03 2022-07-01 21:11:03       kiwi
#> 53 VdYIA9bAQaWnTafJOR7-IA 2022-07-01 21:16:00 2022-07-01 21:16:00       kiwi
#> 54 YiTEtjEnT3OlWI8RJ0GbHA 2022-07-01 21:25:19 2022-07-01 21:25:19       kiwi
#> 55 Gw_qH6Z7RIeR1QHF-PnQfw 2022-08-01 19:57:01 2022-08-01 19:57:01       kiwi
#> 56 Za09YgQyT0y4y7ACVChnFA 2022-08-01 20:02:23 2022-08-01 20:02:23       kiwi
#> 57 OqE9c2o2RYO-jGSfI7hlPg 2022-08-01 20:11:08 2022-08-01 20:11:08       kiwi
#> 58 V__EDBYxRDaqLNoCGJnQ8w 2022-08-01 20:48:34 2022-08-01 20:48:34       kiwi
#> 59 VPVed66YRFu7BH6deAd_Hw 2022-08-01 20:55:12 2022-08-01 20:55:12       kiwi
#> 60 QO0OBDKNSl2mbxJX5vEjKg 2022-08-01 21:00:19 2022-08-01 21:00:19       kiwi
#> 61 J9juoAsSQXuzzBI3TtuL6w 2022-08-28 18:36:52 2022-08-28 18:36:52       kiwi
#> 62 MjCZO6wETZ29wzZuAWoSlQ 2022-08-28 18:42:31 2022-08-28 18:42:31       kiwi
#> 63 IxvyCT3UQXOwGo2CUgjOuA 2022-08-28 19:44:37 2022-08-28 19:44:37       kiwi
#> 64 OqW1gtFTRtq6MO06Sm_lCw 2022-08-31 19:52:42 2022-08-31 19:52:42       kiwi
#> 65 JCAX6dZ5TvSycDTdhgjCsw 2022-08-31 20:00:52 2022-08-31 20:00:52       kiwi
#> 66 a6G0CwCFRIaBdsN0ZKqrJg 2022-08-31 21:35:54 2022-08-31 21:35:54       kiwi
#> 67 JMfv_ifOSQ6QsHg_tfjpVw 2022-08-31 21:41:49 2022-08-31 21:41:49       kiwi
#> 68 LZp9F2A8RhGlWRJpVO5daA 2022-08-31 21:51:44 2022-08-31 21:51:44       kiwi
#> 69 GeE_rs_bQfCTKEVKgTb9jQ 2022-09-12 11:42:51 2022-09-12 11:42:51       kiwi
#> 70 AxVvTwGmQAafJYRpOJqx8g 2022-09-12 11:59:40 2022-09-12 11:59:40       kiwi
#> 71 B5pXXJ8wTCO4mzbbesSIVA 2022-09-12 12:21:12 2022-09-12 12:21:12       kiwi
#> 72 d_AkJSgzSRuZEqPapTahLA 2022-09-12 12:39:22 2022-09-12 12:39:22       kiwi
#> 73 Ampj7vLgTYiP-w0p1LrHOA 2022-09-12 13:59:26 2022-09-12 13:59:26       kiwi
#> 74 W3VuY_tXRpKwV644qtqh1A 2022-09-12 14:05:01 2022-09-12 14:05:01       kiwi
#> 75 XoUJBlWuTP-E_m_Mkwtnbg 2022-09-13 09:01:13 2022-09-13 09:01:13       kiwi
#> 76 WBbegVpbSCa_paGdR718pw 2022-09-13 09:07:08 2022-09-13 09:07:08       kiwi
#> 77 Xu_lF9gIRBGZ1VY0bnzSqg 2022-09-13 09:09:27 2022-09-13 09:09:27       kiwi
#> 78 AaQTau4rQMC0vJMti63XqA 2022-09-14 09:31:22 2022-09-14 09:31:22       kiwi
#> 79 NGvXEDVGTKusZMQFb7X7bA 2022-09-14 09:36:24 2022-09-14 09:36:24       kiwi
#> 80 IdF2TmPgT3mKDmxzZ6SSXw 2022-09-14 09:54:34 2022-09-14 09:54:34       kiwi
#>    nid              person       last_modified date_nominute
#> 1    1               Alice 2022-05-12 16:58:07    2022-01-06
#> 2    2                 Bob 2022-05-12 16:58:08    2022-01-03
#> 3    3               Clara 2022-05-12 16:58:08    2021-08-05
#> 4    6 Frederick the Great 2022-09-14 09:54:33          <NA>
#> 5    6 Frederick the Great 2022-05-12 16:58:09          <NA>
#> 6    6 Frederick the Great 2022-05-12 16:58:10          <NA>
#> 7    6 Frederick the Great 2022-05-12 16:58:10          <NA>
#> 8    6 Frederick the Great 2022-05-12 16:58:11          <NA>
#> 9    6 Frederick the Great 2022-05-12 16:58:11          <NA>
#> 10   6 Frederick the Great 2022-05-12 16:58:12          <NA>
#> 11   6 Frederick the Great 2022-05-12 16:58:12          <NA>
#> 12   6 Frederick the Great 2022-05-12 16:58:13          <NA>
#> 13   6 Frederick the Great 2022-05-12 16:58:13          <NA>
#> 14   6 Frederick the Great 2022-05-12 16:58:13          <NA>
#> 15   6 Frederick the Great 2022-05-12 16:58:14          <NA>
#> 16   6 Frederick the Great 2022-05-12 16:58:14          <NA>
#> 17   6 Frederick the Great 2022-05-12 16:58:15          <NA>
#> 18   6 Frederick the Great 2022-05-12 16:58:15          <NA>
#> 19   6 Frederick the Great 2022-05-12 18:16:18          <NA>
#> 20   6 Frederick the Great 2022-05-12 18:33:47          <NA>
#> 21   6 Frederick the Great 2022-05-12 18:49:36          <NA>
#> 22   6 Frederick the Great 2022-05-12 19:03:47          <NA>
#> 23   6 Frederick the Great 2022-05-12 20:22:02          <NA>
#> 24   6 Frederick the Great 2022-05-12 20:41:57          <NA>
#> 25   6 Frederick the Great 2022-05-13 13:10:55          <NA>
#> 26   6 Frederick the Great 2022-05-13 13:11:49          <NA>
#> 27   6 Frederick the Great 2022-05-13 13:12:24          <NA>
#> 28   6 Frederick the Great 2022-05-13 13:13:31          <NA>
#> 29   6 Frederick the Great 2022-05-13 14:15:29          <NA>
#> 30   6 Frederick the Great 2022-05-13 14:20:30          <NA>
#> 31   6 Frederick the Great 2022-05-13 14:21:21          <NA>
#> 32   6 Frederick the Great 2022-05-13 14:26:16          <NA>
#> 33   6 Frederick the Great 2022-05-13 14:32:10          <NA>
#> 34   6 Frederick the Great 2022-05-22 07:43:19          <NA>
#> 35   6 Frederick the Great 2022-05-22 07:47:43          <NA>
#> 36   6 Frederick the Great 2022-05-22 07:51:06          <NA>
#> 37   6 Frederick the Great 2022-05-29 06:31:28          <NA>
#> 38   6 Frederick the Great 2022-05-29 10:30:15          <NA>
#> 39   6 Frederick the Great 2022-05-29 10:34:54          <NA>
#> 40   6 Frederick the Great 2022-05-29 10:38:00          <NA>
#> 41   6 Frederick the Great 2022-05-29 20:33:27          <NA>
#> 42   6 Frederick the Great 2022-05-29 20:40:41          <NA>
#> 43   6 Frederick the Great 2022-05-29 20:42:10          <NA>
#> 44   6 Frederick the Great 2022-05-29 20:45:57          <NA>
#> 45   6 Frederick the Great 2022-06-06 14:40:12          <NA>
#> 46   6 Frederick the Great 2022-06-06 14:44:59          <NA>
#> 47   6 Frederick the Great 2022-06-06 14:50:54          <NA>
#> 48   6 Frederick the Great 2022-07-01 20:06:17          <NA>
#> 49   6 Frederick the Great 2022-07-01 20:33:14          <NA>
#> 50   6 Frederick the Great 2022-07-01 20:41:09          <NA>
#> 51   6 Frederick the Great 2022-07-01 20:42:29          <NA>
#> 52   6 Frederick the Great 2022-07-01 21:11:03          <NA>
#> 53   6 Frederick the Great 2022-07-01 21:16:00          <NA>
#> 54   6 Frederick the Great 2022-07-01 21:25:19          <NA>
#> 55   6 Frederick the Great 2022-08-01 19:57:01          <NA>
#> 56   6 Frederick the Great 2022-08-01 20:02:23          <NA>
#> 57   6 Frederick the Great 2022-08-01 20:11:08          <NA>
#> 58   6 Frederick the Great 2022-08-01 20:48:34          <NA>
#> 59   6 Frederick the Great 2022-08-01 20:55:12          <NA>
#> 60   6 Frederick the Great 2022-08-01 21:00:19          <NA>
#> 61   6 Frederick the Great 2022-08-28 18:36:52          <NA>
#> 62   6 Frederick the Great 2022-08-28 18:42:31          <NA>
#> 63   6 Frederick the Great 2022-08-28 19:44:37          <NA>
#> 64   6 Frederick the Great 2022-08-31 19:52:42          <NA>
#> 65   6 Frederick the Great 2022-08-31 20:00:52          <NA>
#> 66   6 Frederick the Great 2022-08-31 21:35:54          <NA>
#> 67   6 Frederick the Great 2022-08-31 21:41:49          <NA>
#> 68   6 Frederick the Great 2022-08-31 21:51:44          <NA>
#> 69   6 Frederick the Great 2022-09-12 11:42:51          <NA>
#> 70   6 Frederick the Great 2022-09-12 11:59:40          <NA>
#> 71   6 Frederick the Great 2022-09-12 12:21:12          <NA>
#> 72   6 Frederick the Great 2022-09-12 12:39:22          <NA>
#> 73   6 Frederick the Great 2022-09-12 13:59:26          <NA>
#> 74   6 Frederick the Great 2022-09-12 14:05:01          <NA>
#> 75   6 Frederick the Great 2022-09-13 09:01:13          <NA>
#> 76   6 Frederick the Great 2022-09-13 09:07:08          <NA>
#> 77   6 Frederick the Great 2022-09-13 09:09:27          <NA>
#> 78   6 Frederick the Great 2022-09-14 09:31:22          <NA>
#> 79   6 Frederick the Great 2022-09-14 09:36:24          <NA>
#> 80   6 Frederick the Great 2022-09-14 09:54:34          <NA>
#>           date_wminute                                        user  camid
#> 1  2022-01-12 09:30:00 8adf4f5dd661449fa6cc1f5a0b1815c0@auth.local 100001
#> 2  2022-01-03 07:56:00 c7efb8019da54923a9b04d4a74f0fde8@auth.local 100002
#> 3  2021-08-05 08:30:00 c7efb8019da54923a9b04d4a74f0fde8@auth.local 100003
#> 4                 <NA>                                         NaN 100189
#> 5                 <NA>                                         NaN 100190
#> 6                 <NA>                                         NaN 100191
#> 7                 <NA>                                         NaN 100192
#> 8                 <NA>                                         NaN 100193
#> 9                 <NA>                                         NaN 100194
#> 10                <NA>                                         NaN 100195
#> 11                <NA>                                         NaN 100196
#> 12                <NA>                                         NaN 100197
#> 13                <NA>                                         NaN 100198
#> 14                <NA>                                         NaN 100199
#> 15                <NA>                                         NaN 100200
#> 16                <NA>                                         NaN 100201
#> 17                <NA>                                         NaN 100202
#> 18                <NA>                                         NaN 100203
#> 19                <NA>                                         NaN 100204
#> 20                <NA>                                         NaN 100205
#> 21                <NA>                                         NaN 100206
#> 22                <NA>                                         NaN 100207
#> 23                <NA>                                         NaN 100208
#> 24                <NA>                                         NaN 100209
#> 25                <NA>                                         NaN 100210
#> 26                <NA>                                         NaN 100211
#> 27                <NA>                                         NaN 100212
#> 28                <NA>                                         NaN 100213
#> 29                <NA>                                         NaN 100214
#> 30                <NA>                                         NaN 100215
#> 31                <NA>                                         NaN 100216
#> 32                <NA>                                         NaN 100217
#> 33                <NA>                                         NaN 100218
#> 34                <NA>                                         NaN 100219
#> 35                <NA>                                         NaN 100220
#> 36                <NA>                                         NaN 100221
#> 37                <NA>                                         NaN 100222
#> 38                <NA>                                         NaN 100223
#> 39                <NA>                                         NaN 100224
#> 40                <NA>                                         NaN 100225
#> 41                <NA>                                         NaN 100226
#> 42                <NA>                                         NaN 100227
#> 43                <NA>                                         NaN 100228
#> 44                <NA>                                         NaN 100229
#> 45                <NA>                                         NaN 100230
#> 46                <NA>                                         NaN 100231
#> 47                <NA>                                         NaN 100232
#> 48                <NA>                                         NaN 100233
#> 49                <NA>                                         NaN 100234
#> 50                <NA>                                         NaN 100235
#> 51                <NA>                                         NaN 100236
#> 52                <NA>                                         NaN 100237
#> 53                <NA>                                         NaN 100238
#> 54                <NA>                                         NaN 100239
#> 55                <NA>                                         NaN 100240
#> 56                <NA>                                         NaN 100241
#> 57                <NA>                                         NaN 100242
#> 58                <NA>                                         NaN 100243
#> 59                <NA>                                         NaN 100244
#> 60                <NA>                                         NaN 100245
#> 61                <NA>                                         NaN 100246
#> 62                <NA>                                         NaN 100247
#> 63                <NA>                                         NaN 100248
#> 64                <NA>                                         NaN 100249
#> 65                <NA>                                         NaN 100250
#> 66                <NA>                                         NaN 100251
#> 67                <NA>                                         NaN 100252
#> 68                <NA>                                         NaN 100253
#> 69                <NA>                                         NaN 100254
#> 70                <NA>                                         NaN 100255
#> 71                <NA>                                         NaN 100256
#> 72                <NA>                                         NaN 100257
#> 73                <NA>                                         NaN 100258
#> 74                <NA>                                         NaN 100259
#> 75                <NA>                                         NaN 100260
#> 76                <NA>                                         NaN 100261
#> 77                <NA>                                         NaN 100262
#> 78                <NA>                                         NaN 100263
#> 79                <NA>                                         NaN 100264
#> 80                <NA>                                         NaN 100265
# }
# \donttest{
flytable_query("SELECT person, fruit_name FROM testfruit WHERE person!='Bob'")
#>                 person fruit_name
#> 1                Alice      apple
#> 2                Clara clementine
#> 3  Frederick the Great       kiwi
#> 4  Frederick the Great       kiwi
#> 5  Frederick the Great       kiwi
#> 6  Frederick the Great       kiwi
#> 7  Frederick the Great       kiwi
#> 8  Frederick the Great       kiwi
#> 9  Frederick the Great       kiwi
#> 10 Frederick the Great       kiwi
#> 11 Frederick the Great       kiwi
#> 12 Frederick the Great       kiwi
#> 13 Frederick the Great       kiwi
#> 14 Frederick the Great       kiwi
#> 15 Frederick the Great       kiwi
#> 16 Frederick the Great       kiwi
#> 17 Frederick the Great       kiwi
#> 18 Frederick the Great       kiwi
#> 19 Frederick the Great       kiwi
#> 20 Frederick the Great       kiwi
#> 21 Frederick the Great       kiwi
#> 22 Frederick the Great       kiwi
#> 23 Frederick the Great       kiwi
#> 24 Frederick the Great       kiwi
#> 25 Frederick the Great       kiwi
#> 26 Frederick the Great       kiwi
#> 27 Frederick the Great       kiwi
#> 28 Frederick the Great       kiwi
#> 29 Frederick the Great       kiwi
#> 30 Frederick the Great       kiwi
#> 31 Frederick the Great       kiwi
#> 32 Frederick the Great       kiwi
#> 33 Frederick the Great       kiwi
#> 34 Frederick the Great       kiwi
#> 35 Frederick the Great       kiwi
#> 36 Frederick the Great       kiwi
#> 37 Frederick the Great       kiwi
#> 38 Frederick the Great       kiwi
#> 39 Frederick the Great       kiwi
#> 40 Frederick the Great       kiwi
#> 41 Frederick the Great       kiwi
#> 42 Frederick the Great       kiwi
#> 43 Frederick the Great       kiwi
#> 44 Frederick the Great       kiwi
#> 45 Frederick the Great       kiwi
#> 46 Frederick the Great       kiwi
#> 47 Frederick the Great       kiwi
#> 48 Frederick the Great       kiwi
#> 49 Frederick the Great       kiwi
#> 50 Frederick the Great       kiwi
#> 51 Frederick the Great       kiwi
#> 52 Frederick the Great       kiwi
#> 53 Frederick the Great       kiwi
#> 54 Frederick the Great       kiwi
#> 55 Frederick the Great       kiwi
#> 56 Frederick the Great       kiwi
#> 57 Frederick the Great       kiwi
#> 58 Frederick the Great       kiwi
#> 59 Frederick the Great       kiwi
#> 60 Frederick the Great       kiwi
#> 61 Frederick the Great       kiwi
#> 62 Frederick the Great       kiwi
#> 63 Frederick the Great       kiwi
#> 64 Frederick the Great       kiwi
#> 65 Frederick the Great       kiwi
#> 66 Frederick the Great       kiwi
#> 67 Frederick the Great       kiwi
#> 68 Frederick the Great       kiwi
#> 69 Frederick the Great       kiwi
#> 70 Frederick the Great       kiwi
#> 71 Frederick the Great       kiwi
#> 72 Frederick the Great       kiwi
#> 73 Frederick the Great       kiwi
#> 74 Frederick the Great       kiwi
#> 75 Frederick the Great       kiwi
#> 76 Frederick the Great       kiwi
#> 77 Frederick the Great       kiwi
#> 78 Frederick the Great       kiwi
#> 79 Frederick the Great       kiwi
# }
if (FALSE) {
flytable_query(paste("SELECT root_id, supervoxel_id FROM info limit 5"))
}