Fetch (memoised) flywire cell type information from flytable

Fetch (memoised) flywire cell type information from flytable

flytable_cell_types(
  pattern = NULL,
  version = NULL,
  timestamp = NULL,
  target = c("type", "cell_type", "hemibrain_type", "cell_class"),
  transfer_hemibrain_type = c("extra", "none", "all"),
  cache = TRUE
)

Arguments

pattern

Optional character vector specifying a pattern that cell types must match in a SQL LIKE statement executed by flytable_query. The suffix _L or _R can be used to restricted to neurons annotated to the L or R hemisphere. See examples.

version

An optional CAVE materialisation version number. See details and examples.

timestamp

An optional timestamp as a string or POSIXct, interpreted as UTC when no timezone is specified.

target

A character vector specifying which flytable columns pattern should match. The special value of type means either cell_type or hemibrain_type should match.

transfer_hemibrain_type

Whether to transfer the hemibrain_type column into the cell_type (default TRUE, see details)

cache

Whether to cache the results for 5m (default TRUE since this is a little expensive)

Value

The original data.frame left joined to appropriate rows from flytable.

Details

when transfer_hemibrain_type=TRUE, hemibrain_type values will be transferred into the cell_type column if cell_type is empty.

See also

Examples

# \donttest{
flytable_cell_types("MBON%")
#>               root_id     supervoxel_id  side cell_class cell_type
#> 1  720575940630697303 79448854932107696 right       MBON    MBON02
#> 2  720575940619719509 80011942324010318  left       MBON    MBON01
#> 3  720575940614718870 80715286101144676  left       MBON    MBON32
#> 4  720575940627138386 81137979670302132  left       MBON    MBON10
#> 5  720575940620843585 81349085835581377  left       MBON    MBON30
#> 6  720575940625991268 78603948898007978 right       MBON    MBON32
#> 7  720575940619259352 78111986298788517 right       MBON    MBON06
#> 8  720575940624752398 81349223341739625  left       MBON    MBON05
#> 9  720575940612772579 78393323903110735 right       MBON    MBON10
#> 10 720575940632773037 80223254782485510  left       MBON    MBON20
#> 11 720575940634822751 81277549061520124  left       MBON    MBON28
#> 12 720575940628764495 79588561627810154 right       MBON    MBON02
#> 13 720575940622796813 81489685951698728  left       MBON    MBON33
#> 14 720575940616298635 78180912067538056 right       MBON    MBON23
#> 15 720575940629790967 80152611093049624  left       MBON    MBON29
#> 16 720575940624100587 80154466519131949  left       MBON    MBON14
#> 17 720575940615955286 82897129756231517  left       MBON    MBON15
#> 18 720575940624752568 80154535238527400  left       MBON    MBON12
#> 19 720575940634657899 81066305390179956  left       MBON    MBON15
#> 20 720575940608410889 78040449457502610 right       MBON    MBON28
#> 21 720575940639694168 80785654845309410  left       MBON    MBON31
#> 22 720575940624753336 79166830266640003 right       MBON    MBON12
#> 23 720575940621032052 78181599128215629 right       MBON    MBON35
#> 24 720575940638774606 78040449457461921 right       MBON    MBON17
#> 25 720575940647213956 80224766543977933  left       MBON    MBON14
#> 26 720575940625168528 78251349531485275 right       MBON    MBON19
#> 27 720575940639693144 77901086224032815 right       MBON    MBON30
#> 28 720575940613290477 80082723452177136  left       MBON    MBON27
#> 29 720575940620362945 77408917331762608 right       MBON    MBON09
#> 30 720575940632656173 77549311222677157 right       MBON    MBON10
#> 31 720575940636391566 80224972702135904  left       MBON    MBON12
#> 32 720575940628581379 78111986298850429 right       MBON    MBON10
#> 33 720575940617413957 77478942478556789 right       MBON    MBON10
#> 34 720575940617573922 77549311155597154 right       MBON    MBON10
#> 35 720575940613157329 78111986365505691 right       MBON    MBON10
#> 36 720575940620379119 79167173863853528  left       MBON    MBON03
#> 37 720575940627091210 79380479052758063 right       MBON    MBON13
#> 38 720575940623743415 77901154877073717 right       MBON    MBON09
#> 39 720575940631634489 81136880426746412  left       MBON    MBON15
#> 40 720575940638910141 78393117744840114 right       MBON    MBON10
#> 41 720575940625850826 78251486970442209 right       MBON    MBON15
#> 42 720575940630373113 76985537105536911 right       MBON    MBON15
#> 43 720575940622281430 78040449457463315 right       MBON    MBON15
#> 44 720575940637831487 79166692827138268 right       MBON    MBON25
#> 45 720575940604592300 81136880426744139  left       MBON    MBON15
#> 46 720575940613666001 78110818201651448 right       MBON    MBON15
#> 47 720575940617869083 77760348668670642 right       MBON    MBON09
#> 48 720575940620810561 81137979670303492  left       MBON    MBON10
#> 49 720575940636311694 78110818201681795 right       MBON    MBON17
#> 50 720575940628569603 81067610926141558  left       MBON    MBON10
#> 51 720575940620845633 80786135881765925  left       MBON    MBON11
#> 52 720575940608324105 81277549061455173  left       MBON    MBON17
#> 53 720575940634482783 81136605414678093  left       MBON    MBON19
#> 54 720575940630387820 78182286188838645 right       MBON    MBON31
#> 55 720575940626109315 81066305390093306  left       MBON    MBON19
#> 56 720575940627584516 78110818201621392 right       MBON    MBON16
#> 57 720575940625971300 81277549061481189  left       MBON    MBON17
#> 58 720575940619254744 81136880426764424  left       MBON    MBON24
#> 59 720575940622268374 79027192356412856 right       MBON    MBON27
#> 60 720575940620377583 81277617781167541  left       MBON    MBON16
#> 61 720575940615186079 77899986914253830 right       MBON    MBON15
#> 62 720575940612636694 78111917579300796 right       MBON    MBON21
#> 63 720575940620252075 81348124031010814  left       MBON    MBON15
#> 64 720575940632028753 78324741798855874 right       MBON    MBON20
#> 65 720575940625083981 81137979736901319  left       MBON    MBON21
#> 66 720575940617176925 81136880426735755  left       MBON    MBON23
#> 67 720575940617125081 79309972869657633 right       MBON    MBON12
#> 68 720575940612637974 78603949099622269 right       MBON    MBON07
#> 69 720575940654692001 79380204241413766 right       MBON    MBON14
#> 70 720575940625994596 79309904216692475 right       MBON    MBON14
#> 71 720575940628587779 78323023677138555 right       MBON    MBON11
#> 72 720575940604498400 80082242348953693  left       MBON    MBON22
#> 73 720575940619271128 77830305364995511 right       MBON    MBON22
#> 74 720575940619720277 78885355625276424 right       MBON    MBON26
#> 75 720575940631031055 78463211409646517 right       MBON    MBON33
#> 76 720575940620841793 79943222847625326  left       MBON    MBON13
#> 77 720575940640470992 80925980016963911  left       MBON    MBON35
#> 78 720575940624744462 78040449457553649 right       MBON    MBON18
#> 79 720575940631940236 80011530007496199 right       MBON    MBON04
#> 80 720575940628938907 80434292093847745  left       MBON    MBON07
#> 81 720575940623204548 79870998677653042 right       MBON    MBON01
#> 82 720575940630390265 79237542540872430 right       MBON    MBON29
#> 83 720575940630691415 78604086538417636 right       MBON    MBON07
#> 84 720575940627159122 81277549061591572  left       MBON    MBON18
#> 85 720575940614720662 79941230049538531 right       MBON    MBON03
#> 86 720575940634969707 79869899232522382  left       MBON    MBON04
#> 87 720575940621035380 81630904409213619  left       MBON    MBON09
#> 88 720575940635407352 80433124332533007  left       MBON    MBON26
#> 89 720575940617624833 80644779918075196  left       MBON    MBON07
#> 90 720575940632219300 81560466945525625  left       MBON    MBON09
#> 91 720575940644417828 81419592152467685  left       MBON    MBON06
#> 92 720575940623088937 80222636373797362 right       MBON    MBON05
#> 93 720575940632393164 78251349531447591 right       MBON    MBON19
#> 94 720575940621137334 78814711734490195 right       MBON    MBON24
#> 95 720575940635746671 79166761480079751  left       MBON    MBON25
#> 96 720575940623199684 78463280330557921 right       MBON    MBON10
#>    hemibrain_type hemibrain_match vfb_id
#> 1          MBON02       424789697   <NA>
#> 2          MBON01       612371421   <NA>
#> 3          MBON32            <NA>   <NA>
#> 4                            <NA>   <NA>
#> 5          MBON30            <NA>   <NA>
#> 6          MBON32            <NA>   <NA>
#> 7          MBON06            <NA>   <NA>
#> 8          MBON05            <NA>   <NA>
#> 9                       768555687   <NA>
#> 10         MBON20            <NA>   <NA>
#> 11                           <NA>   <NA>
#> 12         MBON02            <NA>   <NA>
#> 13         MBON33            <NA>   <NA>
#> 14         MBON23       423382015   <NA>
#> 15         MBON29      5813052909   <NA>
#> 16                      300972942   <NA>
#> 17                           <NA>   <NA>
#> 18                           <NA>   <NA>
#> 19                           <NA>   <NA>
#> 20         MBON28       457196643   <NA>
#> 21         MBON31            <NA>   <NA>
#> 22                      704466265   <NA>
#> 23         MBON35            <NA>   <NA>
#> 24                     5812981264   <NA>
#> 25                     5813068729   <NA>
#> 26                           <NA>   <NA>
#> 27         MBON30            <NA>   <NA>
#> 28         MBON27      5813061538   <NA>
#> 29                           <NA>   <NA>
#> 30                           <NA>   <NA>
#> 31                           <NA>   <NA>
#> 32                           <NA>   <NA>
#> 33                           <NA>   <NA>
#> 34                           <NA>   <NA>
#> 35                           <NA>   <NA>
#> 36         MBON03       612738462   <NA>
#> 37         MBON13            <NA>   <NA>
#> 38                     1078693835   <NA>
#> 39                           <NA>   <NA>
#> 40                           <NA>   <NA>
#> 41                      673702721   <NA>
#> 42                      642664141   <NA>
#> 43                      579916831   <NA>
#> 44                           <NA>   <NA>
#> 45                           <NA>   <NA>
#> 46                      457175171   <NA>
#> 47                     1078693835   <NA>
#> 48                           <NA>   <NA>
#> 49                     5812981264   <NA>
#> 50                           <NA>   <NA>
#> 51         MBON11            <NA>   <NA>
#> 52                     1932553362   <NA>
#> 53                           <NA>   <NA>
#> 54         MBON31            <NA>   <NA>
#> 55                     1902178951   <NA>
#> 56                           <NA>   <NA>
#> 57                           <NA>   <NA>
#> 58         MBON24            <NA>   <NA>
#> 59         MBON27      5813053891   <NA>
#> 60         MBON28            <NA>   <NA>
#> 61        LHPV4m1       518899665   <NA>
#> 62         MBON21            <NA>   <NA>
#> 63        LHPV4m1       580935628   <NA>
#> 64         MBON20            <NA>   <NA>
#> 65         MBON21       985813153   <NA>
#> 66         MBON23            <NA>   <NA>
#> 67                           <NA>   <NA>
#> 68                           <NA>   <NA>
#> 69                           <NA>   <NA>
#> 70                           <NA>   <NA>
#> 71         MBON11            <NA>   <NA>
#> 72         MBON22       985131499   <NA>
#> 73         MBON22      5813042659   <NA>
#> 74         MBON26            <NA>   <NA>
#> 75         MBON33            <NA>   <NA>
#> 76         MBON13            <NA>   <NA>
#> 77         MBON35            <NA>   <NA>
#> 78         MBON18      5813020828   <NA>
#> 79         MBON04            <NA>   <NA>
#> 80                           <NA>   <NA>
#> 81         MBON01       673509195   <NA>
#> 82         MBON29      5813058048   <NA>
#> 83                     1234386037   <NA>
#> 84         MBON18       457196444   <NA>
#> 85         MBON03       487925037   <NA>
#> 86         MBON04            <NA>   <NA>
#> 87                           <NA>   <NA>
#> 88         MBON26            <NA>   <NA>
#> 89                           <NA>   <NA>
#> 90                           <NA>   <NA>
#> 91         MBON06            <NA>   <NA>
#> 92         MBON05            <NA>   <NA>
#> 93                      423774471   <NA>
#> 94         MBON24            <NA>   <NA>
#> 95                      612742248   <NA>
#> 96                           <NA>   <NA>
flytable_cell_types("MBON%", version=450)
#> Updating 84 ids
#>               root_id     supervoxel_id  side cell_class cell_type
#> 1  720575940620475757 79448854932107696 right       MBON    MBON02
#> 2  720575940635988581 80011942324010318  left       MBON    MBON01
#> 3  720575940616047221 80715286101144676  left       MBON    MBON32
#> 4  720575940638780494 81137979670302132  left       MBON    MBON10
#> 5  720575940610014510 81349085835581377  left       MBON    MBON30
#> 6  720575940620520769 78603948898007978 right       MBON    MBON32
#> 7  720575940621245680 78111986298788517 right       MBON    MBON06
#> 8  720575940631910308 81349223341739625  left       MBON    MBON05
#> 9  720575940612772579 78393323903110735 right       MBON    MBON10
#> 10 720575940616240011 80223254782485510  left       MBON    MBON20
#> 11 720575940624260839 81277549061520124  left       MBON    MBON28
#> 12 720575940625496010 79588561627810154 right       MBON    MBON02
#> 13 720575940621704361 81489685951698728  left       MBON    MBON33
#> 14 720575940616298635 78180912067538056 right       MBON    MBON23
#> 15 720575940621187056 80152611093049624  left       MBON    MBON29
#> 16 720575940621208820 80154466519131949  left       MBON    MBON14
#> 17 720575940615955286 82897129756231517  left       MBON    MBON15
#> 18 720575940639603584 80154535238527400  left       MBON    MBON12
#> 19 720575940634657899 81066305390179956  left       MBON    MBON15
#> 20 720575940630796363 78040449457502610 right       MBON    MBON28
#> 21 720575940621196784 80785654845309410  left       MBON    MBON31
#> 22 720575940617807429 79166830266640003 right       MBON    MBON12
#> 23 720575940623046035 78181599128215629 right       MBON    MBON35
#> 24 720575940638774606 78040449457461921 right       MBON    MBON17
#> 25 720575940643518408 80224766543977933  left       MBON    MBON14
#> 26 720575940625168528 78251349531485275 right       MBON    MBON19
#> 27 720575940615443468 77901086224032815 right       MBON    MBON30
#> 28 720575940613049325 80082723452177136  left       MBON    MBON27
#> 29 720575940627152712 77408917331762608 right       MBON    MBON09
#> 30 720575940632009517 77549311222677157 right       MBON    MBON10
#> 31 720575940623615561 80224972702135904  left       MBON    MBON12
#> 32 720575940634844152 78111986298850429 right       MBON    MBON10
#> 33 720575940617413957 77478942478556789 right       MBON    MBON10
#> 34 720575940616008482 77549311155597154 right       MBON    MBON10
#> 35 720575940613157329 78111986365505691 right       MBON    MBON10
#> 36 720575940628356611 79167173863853528  left       MBON    MBON03
#> 37 720575940629103486 79380479052758063 right       MBON    MBON13
#> 38 720575940616937940 77901154877073717 right       MBON    MBON09
#> 39 720575940626510729 81136880426746412  left       MBON    MBON15
#> 40 720575940626935208 78393117744840114 right       MBON    MBON10
#> 41 720575940631573165 78251486970442209 right       MBON    MBON15
#> 42 720575940627235215 76985537105536911 right       MBON    MBON15
#> 43 720575940625623165 78040449457463315 right       MBON    MBON15
#> 44 720575940641532557 79166692827138268 right       MBON    MBON25
#> 45 720575940604592300 81136880426744139  left       MBON    MBON15
#> 46 720575940613666001 78110818201651448 right       MBON    MBON15
#> 47 720575940620015275 77760348668670642 right       MBON    MBON09
#> 48 720575940619587975 81137979670303492  left       MBON    MBON10
#> 49 720575940636538480 78110818201681795 right       MBON    MBON17
#> 50 720575940629528064 81067610926141558  left       MBON    MBON10
#> 51 720575940611534193 80786135881765925  left       MBON    MBON11
#> 52 720575940624149908 81277549061455173  left       MBON    MBON17
#> 53 720575940634482783 81136605414678093  left       MBON    MBON19
#> 54 720575940628680039 78182286188838645 right       MBON    MBON31
#> 55 720575940626109315 81066305390093306  left       MBON    MBON19
#> 56 720575940644139812 78110818201621392 right       MBON    MBON16
#> 57 720575940654302881 81277549061481189  left       MBON    MBON17
#> 58 720575940626786949 81136880426764424  left       MBON    MBON24
#> 59 720575940611536241 79027192356412856 right       MBON    MBON27
#> 60 720575940623018730 81277617781167541  left       MBON    MBON16
#> 61 720575940644732526 77899986914253830 right       MBON    MBON15
#> 62 720575940635488623 78111917579300796 right       MBON    MBON21
#> 63 720575940622944511 81348124031010814  left       MBON    MBON15
#> 64 720575940616286603 78324741798855874 right       MBON    MBON20
#> 65 720575940621503265 81137979736901319  left       MBON    MBON21
#> 66 720575940631189304 81136880426735755  left       MBON    MBON23
#> 67 720575940617330945 79309972869657633 right       MBON    MBON12
#> 68 720575940628523419 78603949099622269 right       MBON    MBON07
#> 69 720575940610543826 79380204241413766 right       MBON    MBON14
#> 70 720575940616968633 79309904216692475 right       MBON    MBON14
#> 71 720575940623762923 78323023677138555 right       MBON    MBON11
#> 72 720575940626901637 80082242348953693  left       MBON    MBON22
#> 73 720575940623211498 77830305364995511 right       MBON    MBON22
#> 74 720575940625816420 78885355625276424 right       MBON    MBON26
#> 75 720575940619919531 78463211409646517 right       MBON    MBON33
#> 76 720575940614532970 79943222847625326  left       MBON    MBON13
#> 77 720575940643356064 80925980016963911  left       MBON    MBON35
#> 78 720575940649242233 78040449457553649 right       MBON    MBON18
#> 79 720575940619620487 80011530007496199 right       MBON    MBON04
#> 80 720575940624835661 80434292093847745  left       MBON    MBON07
#> 81 720575940625538110 79870998677653042 right       MBON    MBON01
#> 82 720575940626706825 79237542540872430 right       MBON    MBON29
#> 83 720575940605195878 78604086538417636 right       MBON    MBON07
#> 84 720575940631910052 81277549061591572  left       MBON    MBON18
#> 85 720575940608426197 79941230049538531 right       MBON    MBON03
#> 86 720575940625222288 79869899232522382  left       MBON    MBON04
#> 87 720575940620070278 81630904409213619  left       MBON    MBON09
#> 88 720575940633688220 80433124332533007  left       MBON    MBON26
#> 89 720575940613357096 80644779918075196  left       MBON    MBON07
#> 90 720575940608355029 81560466945525625  left       MBON    MBON09
#> 91 720575940624105196 81419592152467685  left       MBON    MBON06
#> 92 720575940613666769 80222636373797362 right       MBON    MBON05
#> 93 720575940605182822 78251349531447591 right       MBON    MBON19
#> 94 720575940635161268 78814711734490195 right       MBON    MBON24
#> 95 720575940631601292 79166761480079751  left       MBON    MBON25
#> 96 720575940614573930 78463280330557921 right       MBON    MBON10
#>    hemibrain_type hemibrain_match vfb_id
#> 1          MBON02       424789697   <NA>
#> 2          MBON01       612371421   <NA>
#> 3          MBON32            <NA>   <NA>
#> 4                            <NA>   <NA>
#> 5          MBON30            <NA>   <NA>
#> 6          MBON32            <NA>   <NA>
#> 7          MBON06            <NA>   <NA>
#> 8          MBON05            <NA>   <NA>
#> 9                       768555687   <NA>
#> 10         MBON20            <NA>   <NA>
#> 11                           <NA>   <NA>
#> 12         MBON02            <NA>   <NA>
#> 13         MBON33            <NA>   <NA>
#> 14         MBON23       423382015   <NA>
#> 15         MBON29      5813052909   <NA>
#> 16                      300972942   <NA>
#> 17                           <NA>   <NA>
#> 18                           <NA>   <NA>
#> 19                           <NA>   <NA>
#> 20         MBON28       457196643   <NA>
#> 21         MBON31            <NA>   <NA>
#> 22                      704466265   <NA>
#> 23         MBON35            <NA>   <NA>
#> 24                     5812981264   <NA>
#> 25                     5813068729   <NA>
#> 26                           <NA>   <NA>
#> 27         MBON30            <NA>   <NA>
#> 28         MBON27      5813061538   <NA>
#> 29                           <NA>   <NA>
#> 30                           <NA>   <NA>
#> 31                           <NA>   <NA>
#> 32                           <NA>   <NA>
#> 33                           <NA>   <NA>
#> 34                           <NA>   <NA>
#> 35                           <NA>   <NA>
#> 36         MBON03       612738462   <NA>
#> 37         MBON13            <NA>   <NA>
#> 38                     1078693835   <NA>
#> 39                           <NA>   <NA>
#> 40                           <NA>   <NA>
#> 41                      673702721   <NA>
#> 42                      642664141   <NA>
#> 43                      579916831   <NA>
#> 44                           <NA>   <NA>
#> 45                           <NA>   <NA>
#> 46                      457175171   <NA>
#> 47                     1078693835   <NA>
#> 48                           <NA>   <NA>
#> 49                     5812981264   <NA>
#> 50                           <NA>   <NA>
#> 51         MBON11            <NA>   <NA>
#> 52                     1932553362   <NA>
#> 53                           <NA>   <NA>
#> 54         MBON31            <NA>   <NA>
#> 55                     1902178951   <NA>
#> 56                           <NA>   <NA>
#> 57                           <NA>   <NA>
#> 58         MBON24            <NA>   <NA>
#> 59         MBON27      5813053891   <NA>
#> 60         MBON28            <NA>   <NA>
#> 61        LHPV4m1       518899665   <NA>
#> 62         MBON21            <NA>   <NA>
#> 63        LHPV4m1       580935628   <NA>
#> 64         MBON20            <NA>   <NA>
#> 65         MBON21       985813153   <NA>
#> 66         MBON23            <NA>   <NA>
#> 67                           <NA>   <NA>
#> 68                           <NA>   <NA>
#> 69                           <NA>   <NA>
#> 70                           <NA>   <NA>
#> 71         MBON11            <NA>   <NA>
#> 72         MBON22       985131499   <NA>
#> 73         MBON22      5813042659   <NA>
#> 74         MBON26            <NA>   <NA>
#> 75         MBON33            <NA>   <NA>
#> 76         MBON13            <NA>   <NA>
#> 77         MBON35            <NA>   <NA>
#> 78         MBON18      5813020828   <NA>
#> 79         MBON04            <NA>   <NA>
#> 80                           <NA>   <NA>
#> 81         MBON01       673509195   <NA>
#> 82         MBON29      5813058048   <NA>
#> 83                     1234386037   <NA>
#> 84         MBON18       457196444   <NA>
#> 85         MBON03       487925037   <NA>
#> 86         MBON04            <NA>   <NA>
#> 87                           <NA>   <NA>
#> 88         MBON26            <NA>   <NA>
#> 89                           <NA>   <NA>
#> 90                           <NA>   <NA>
#> 91         MBON06            <NA>   <NA>
#> 92         MBON05            <NA>   <NA>
#> 93                      423774471   <NA>
#> 94         MBON24            <NA>   <NA>
#> 95                      612742248   <NA>
#> 96                           <NA>   <NA>
# two characters
flytable_cell_types("MBON__")
#>               root_id     supervoxel_id  side cell_class cell_type
#> 1  720575940630697303 79448854932107696 right       MBON    MBON02
#> 2  720575940619719509 80011942324010318  left       MBON    MBON01
#> 3  720575940614718870 80715286101144676  left       MBON    MBON32
#> 4  720575940627138386 81137979670302132  left       MBON    MBON10
#> 5  720575940620843585 81349085835581377  left       MBON    MBON30
#> 6  720575940625991268 78603948898007978 right       MBON    MBON32
#> 7  720575940619259352 78111986298788517 right       MBON    MBON06
#> 8  720575940624752398 81349223341739625  left       MBON    MBON05
#> 9  720575940612772579 78393323903110735 right       MBON    MBON10
#> 10 720575940632773037 80223254782485510  left       MBON    MBON20
#> 11 720575940634822751 81277549061520124  left       MBON    MBON28
#> 12 720575940628764495 79588561627810154 right       MBON    MBON02
#> 13 720575940622796813 81489685951698728  left       MBON    MBON33
#> 14 720575940616298635 78180912067538056 right       MBON    MBON23
#> 15 720575940629790967 80152611093049624  left       MBON    MBON29
#> 16 720575940624100587 80154466519131949  left       MBON    MBON14
#> 17 720575940615955286 82897129756231517  left       MBON    MBON15
#> 18 720575940624752568 80154535238527400  left       MBON    MBON12
#> 19 720575940634657899 81066305390179956  left       MBON    MBON15
#> 20 720575940608410889 78040449457502610 right       MBON    MBON28
#> 21 720575940639694168 80785654845309410  left       MBON    MBON31
#> 22 720575940624753336 79166830266640003 right       MBON    MBON12
#> 23 720575940621032052 78181599128215629 right       MBON    MBON35
#> 24 720575940638774606 78040449457461921 right       MBON    MBON17
#> 25 720575940647213956 80224766543977933  left       MBON    MBON14
#> 26 720575940625168528 78251349531485275 right       MBON    MBON19
#> 27 720575940639693144 77901086224032815 right       MBON    MBON30
#> 28 720575940613290477 80082723452177136  left       MBON    MBON27
#> 29 720575940620362945 77408917331762608 right       MBON    MBON09
#> 30 720575940632656173 77549311222677157 right       MBON    MBON10
#> 31 720575940636391566 80224972702135904  left       MBON    MBON12
#> 32 720575940628581379 78111986298850429 right       MBON    MBON10
#> 33 720575940617413957 77478942478556789 right       MBON    MBON10
#> 34 720575940617573922 77549311155597154 right       MBON    MBON10
#> 35 720575940613157329 78111986365505691 right       MBON    MBON10
#> 36 720575940620379119 79167173863853528  left       MBON    MBON03
#> 37 720575940627091210 79380479052758063 right       MBON    MBON13
#> 38 720575940623743415 77901154877073717 right       MBON    MBON09
#> 39 720575940631634489 81136880426746412  left       MBON    MBON15
#> 40 720575940638910141 78393117744840114 right       MBON    MBON10
#> 41 720575940625850826 78251486970442209 right       MBON    MBON15
#> 42 720575940630373113 76985537105536911 right       MBON    MBON15
#> 43 720575940622281430 78040449457463315 right       MBON    MBON15
#> 44 720575940637831487 79166692827138268 right       MBON    MBON25
#> 45 720575940604592300 81136880426744139  left       MBON    MBON15
#> 46 720575940613666001 78110818201651448 right       MBON    MBON15
#> 47 720575940617869083 77760348668670642 right       MBON    MBON09
#> 48 720575940620810561 81137979670303492  left       MBON    MBON10
#> 49 720575940636311694 78110818201681795 right       MBON    MBON17
#> 50 720575940628569603 81067610926141558  left       MBON    MBON10
#> 51 720575940620845633 80786135881765925  left       MBON    MBON11
#> 52 720575940608324105 81277549061455173  left       MBON    MBON17
#> 53 720575940634482783 81136605414678093  left       MBON    MBON19
#> 54 720575940630387820 78182286188838645 right       MBON    MBON31
#> 55 720575940626109315 81066305390093306  left       MBON    MBON19
#> 56 720575940627584516 78110818201621392 right       MBON    MBON16
#> 57 720575940625971300 81277549061481189  left       MBON    MBON17
#> 58 720575940619254744 81136880426764424  left       MBON    MBON24
#> 59 720575940622268374 79027192356412856 right       MBON    MBON27
#> 60 720575940620377583 81277617781167541  left       MBON    MBON16
#> 61 720575940615186079 77899986914253830 right       MBON    MBON15
#> 62 720575940612636694 78111917579300796 right       MBON    MBON21
#> 63 720575940620252075 81348124031010814  left       MBON    MBON15
#> 64 720575940632028753 78324741798855874 right       MBON    MBON20
#> 65 720575940625083981 81137979736901319  left       MBON    MBON21
#> 66 720575940617176925 81136880426735755  left       MBON    MBON23
#> 67 720575940617125081 79309972869657633 right       MBON    MBON12
#> 68 720575940612637974 78603949099622269 right       MBON    MBON07
#> 69 720575940654692001 79380204241413766 right       MBON    MBON14
#> 70 720575940625994596 79309904216692475 right       MBON    MBON14
#> 71 720575940628587779 78323023677138555 right       MBON    MBON11
#> 72 720575940604498400 80082242348953693  left       MBON    MBON22
#> 73 720575940619271128 77830305364995511 right       MBON    MBON22
#> 74 720575940619720277 78885355625276424 right       MBON    MBON26
#> 75 720575940631031055 78463211409646517 right       MBON    MBON33
#> 76 720575940620841793 79943222847625326  left       MBON    MBON13
#> 77 720575940640470992 80925980016963911  left       MBON    MBON35
#> 78 720575940624744462 78040449457553649 right       MBON    MBON18
#> 79 720575940631940236 80011530007496199 right       MBON    MBON04
#> 80 720575940628938907 80434292093847745  left       MBON    MBON07
#> 81 720575940623204548 79870998677653042 right       MBON    MBON01
#> 82 720575940630390265 79237542540872430 right       MBON    MBON29
#> 83 720575940630691415 78604086538417636 right       MBON    MBON07
#> 84 720575940627159122 81277549061591572  left       MBON    MBON18
#> 85 720575940614720662 79941230049538531 right       MBON    MBON03
#> 86 720575940634969707 79869899232522382  left       MBON    MBON04
#> 87 720575940621035380 81630904409213619  left       MBON    MBON09
#> 88 720575940635407352 80433124332533007  left       MBON    MBON26
#> 89 720575940617624833 80644779918075196  left       MBON    MBON07
#> 90 720575940632219300 81560466945525625  left       MBON    MBON09
#> 91 720575940644417828 81419592152467685  left       MBON    MBON06
#> 92 720575940623088937 80222636373797362 right       MBON    MBON05
#> 93 720575940632393164 78251349531447591 right       MBON    MBON19
#> 94 720575940621137334 78814711734490195 right       MBON    MBON24
#> 95 720575940635746671 79166761480079751  left       MBON    MBON25
#> 96 720575940623199684 78463280330557921 right       MBON    MBON10
#>    hemibrain_type hemibrain_match vfb_id
#> 1          MBON02       424789697   <NA>
#> 2          MBON01       612371421   <NA>
#> 3          MBON32            <NA>   <NA>
#> 4                            <NA>   <NA>
#> 5          MBON30            <NA>   <NA>
#> 6          MBON32            <NA>   <NA>
#> 7          MBON06            <NA>   <NA>
#> 8          MBON05            <NA>   <NA>
#> 9                       768555687   <NA>
#> 10         MBON20            <NA>   <NA>
#> 11                           <NA>   <NA>
#> 12         MBON02            <NA>   <NA>
#> 13         MBON33            <NA>   <NA>
#> 14         MBON23       423382015   <NA>
#> 15         MBON29      5813052909   <NA>
#> 16                      300972942   <NA>
#> 17                           <NA>   <NA>
#> 18                           <NA>   <NA>
#> 19                           <NA>   <NA>
#> 20         MBON28       457196643   <NA>
#> 21         MBON31            <NA>   <NA>
#> 22                      704466265   <NA>
#> 23         MBON35            <NA>   <NA>
#> 24                     5812981264   <NA>
#> 25                     5813068729   <NA>
#> 26                           <NA>   <NA>
#> 27         MBON30            <NA>   <NA>
#> 28         MBON27      5813061538   <NA>
#> 29                           <NA>   <NA>
#> 30                           <NA>   <NA>
#> 31                           <NA>   <NA>
#> 32                           <NA>   <NA>
#> 33                           <NA>   <NA>
#> 34                           <NA>   <NA>
#> 35                           <NA>   <NA>
#> 36         MBON03       612738462   <NA>
#> 37         MBON13            <NA>   <NA>
#> 38                     1078693835   <NA>
#> 39                           <NA>   <NA>
#> 40                           <NA>   <NA>
#> 41                      673702721   <NA>
#> 42                      642664141   <NA>
#> 43                      579916831   <NA>
#> 44                           <NA>   <NA>
#> 45                           <NA>   <NA>
#> 46                      457175171   <NA>
#> 47                     1078693835   <NA>
#> 48                           <NA>   <NA>
#> 49                     5812981264   <NA>
#> 50                           <NA>   <NA>
#> 51         MBON11            <NA>   <NA>
#> 52                     1932553362   <NA>
#> 53                           <NA>   <NA>
#> 54         MBON31            <NA>   <NA>
#> 55                     1902178951   <NA>
#> 56                           <NA>   <NA>
#> 57                           <NA>   <NA>
#> 58         MBON24            <NA>   <NA>
#> 59         MBON27      5813053891   <NA>
#> 60         MBON28            <NA>   <NA>
#> 61        LHPV4m1       518899665   <NA>
#> 62         MBON21            <NA>   <NA>
#> 63        LHPV4m1       580935628   <NA>
#> 64         MBON20            <NA>   <NA>
#> 65         MBON21       985813153   <NA>
#> 66         MBON23            <NA>   <NA>
#> 67                           <NA>   <NA>
#> 68                           <NA>   <NA>
#> 69                           <NA>   <NA>
#> 70                           <NA>   <NA>
#> 71         MBON11            <NA>   <NA>
#> 72         MBON22       985131499   <NA>
#> 73         MBON22      5813042659   <NA>
#> 74         MBON26            <NA>   <NA>
#> 75         MBON33            <NA>   <NA>
#> 76         MBON13            <NA>   <NA>
#> 77         MBON35            <NA>   <NA>
#> 78         MBON18      5813020828   <NA>
#> 79         MBON04            <NA>   <NA>
#> 80                           <NA>   <NA>
#> 81         MBON01       673509195   <NA>
#> 82         MBON29      5813058048   <NA>
#> 83                     1234386037   <NA>
#> 84         MBON18       457196444   <NA>
#> 85         MBON03       487925037   <NA>
#> 86         MBON04            <NA>   <NA>
#> 87                           <NA>   <NA>
#> 88         MBON26            <NA>   <NA>
#> 89                           <NA>   <NA>
#> 90                           <NA>   <NA>
#> 91         MBON06            <NA>   <NA>
#> 92         MBON05            <NA>   <NA>
#> 93                      423774471   <NA>
#> 94         MBON24            <NA>   <NA>
#> 95                      612742248   <NA>
#> 96                           <NA>   <NA>
# at least one character
flytable_cell_types("MBON_%")
#>               root_id     supervoxel_id  side cell_class cell_type
#> 1  720575940630697303 79448854932107696 right       MBON    MBON02
#> 2  720575940619719509 80011942324010318  left       MBON    MBON01
#> 3  720575940614718870 80715286101144676  left       MBON    MBON32
#> 4  720575940627138386 81137979670302132  left       MBON    MBON10
#> 5  720575940620843585 81349085835581377  left       MBON    MBON30
#> 6  720575940625991268 78603948898007978 right       MBON    MBON32
#> 7  720575940619259352 78111986298788517 right       MBON    MBON06
#> 8  720575940624752398 81349223341739625  left       MBON    MBON05
#> 9  720575940612772579 78393323903110735 right       MBON    MBON10
#> 10 720575940632773037 80223254782485510  left       MBON    MBON20
#> 11 720575940634822751 81277549061520124  left       MBON    MBON28
#> 12 720575940628764495 79588561627810154 right       MBON    MBON02
#> 13 720575940622796813 81489685951698728  left       MBON    MBON33
#> 14 720575940616298635 78180912067538056 right       MBON    MBON23
#> 15 720575940629790967 80152611093049624  left       MBON    MBON29
#> 16 720575940624100587 80154466519131949  left       MBON    MBON14
#> 17 720575940615955286 82897129756231517  left       MBON    MBON15
#> 18 720575940624752568 80154535238527400  left       MBON    MBON12
#> 19 720575940634657899 81066305390179956  left       MBON    MBON15
#> 20 720575940608410889 78040449457502610 right       MBON    MBON28
#> 21 720575940639694168 80785654845309410  left       MBON    MBON31
#> 22 720575940624753336 79166830266640003 right       MBON    MBON12
#> 23 720575940621032052 78181599128215629 right       MBON    MBON35
#> 24 720575940638774606 78040449457461921 right       MBON    MBON17
#> 25 720575940647213956 80224766543977933  left       MBON    MBON14
#> 26 720575940625168528 78251349531485275 right       MBON    MBON19
#> 27 720575940639693144 77901086224032815 right       MBON    MBON30
#> 28 720575940613290477 80082723452177136  left       MBON    MBON27
#> 29 720575940620362945 77408917331762608 right       MBON    MBON09
#> 30 720575940632656173 77549311222677157 right       MBON    MBON10
#> 31 720575940636391566 80224972702135904  left       MBON    MBON12
#> 32 720575940628581379 78111986298850429 right       MBON    MBON10
#> 33 720575940617413957 77478942478556789 right       MBON    MBON10
#> 34 720575940617573922 77549311155597154 right       MBON    MBON10
#> 35 720575940613157329 78111986365505691 right       MBON    MBON10
#> 36 720575940620379119 79167173863853528  left       MBON    MBON03
#> 37 720575940627091210 79380479052758063 right       MBON    MBON13
#> 38 720575940623743415 77901154877073717 right       MBON    MBON09
#> 39 720575940631634489 81136880426746412  left       MBON    MBON15
#> 40 720575940638910141 78393117744840114 right       MBON    MBON10
#> 41 720575940625850826 78251486970442209 right       MBON    MBON15
#> 42 720575940630373113 76985537105536911 right       MBON    MBON15
#> 43 720575940622281430 78040449457463315 right       MBON    MBON15
#> 44 720575940637831487 79166692827138268 right       MBON    MBON25
#> 45 720575940604592300 81136880426744139  left       MBON    MBON15
#> 46 720575940613666001 78110818201651448 right       MBON    MBON15
#> 47 720575940617869083 77760348668670642 right       MBON    MBON09
#> 48 720575940620810561 81137979670303492  left       MBON    MBON10
#> 49 720575940636311694 78110818201681795 right       MBON    MBON17
#> 50 720575940628569603 81067610926141558  left       MBON    MBON10
#> 51 720575940620845633 80786135881765925  left       MBON    MBON11
#> 52 720575940608324105 81277549061455173  left       MBON    MBON17
#> 53 720575940634482783 81136605414678093  left       MBON    MBON19
#> 54 720575940630387820 78182286188838645 right       MBON    MBON31
#> 55 720575940626109315 81066305390093306  left       MBON    MBON19
#> 56 720575940627584516 78110818201621392 right       MBON    MBON16
#> 57 720575940625971300 81277549061481189  left       MBON    MBON17
#> 58 720575940619254744 81136880426764424  left       MBON    MBON24
#> 59 720575940622268374 79027192356412856 right       MBON    MBON27
#> 60 720575940620377583 81277617781167541  left       MBON    MBON16
#> 61 720575940615186079 77899986914253830 right       MBON    MBON15
#> 62 720575940612636694 78111917579300796 right       MBON    MBON21
#> 63 720575940620252075 81348124031010814  left       MBON    MBON15
#> 64 720575940632028753 78324741798855874 right       MBON    MBON20
#> 65 720575940625083981 81137979736901319  left       MBON    MBON21
#> 66 720575940617176925 81136880426735755  left       MBON    MBON23
#> 67 720575940617125081 79309972869657633 right       MBON    MBON12
#> 68 720575940612637974 78603949099622269 right       MBON    MBON07
#> 69 720575940654692001 79380204241413766 right       MBON    MBON14
#> 70 720575940625994596 79309904216692475 right       MBON    MBON14
#> 71 720575940628587779 78323023677138555 right       MBON    MBON11
#> 72 720575940604498400 80082242348953693  left       MBON    MBON22
#> 73 720575940619271128 77830305364995511 right       MBON    MBON22
#> 74 720575940619720277 78885355625276424 right       MBON    MBON26
#> 75 720575940631031055 78463211409646517 right       MBON    MBON33
#> 76 720575940620841793 79943222847625326  left       MBON    MBON13
#> 77 720575940640470992 80925980016963911  left       MBON    MBON35
#> 78 720575940624744462 78040449457553649 right       MBON    MBON18
#> 79 720575940631940236 80011530007496199 right       MBON    MBON04
#> 80 720575940628938907 80434292093847745  left       MBON    MBON07
#> 81 720575940623204548 79870998677653042 right       MBON    MBON01
#> 82 720575940630390265 79237542540872430 right       MBON    MBON29
#> 83 720575940630691415 78604086538417636 right       MBON    MBON07
#> 84 720575940627159122 81277549061591572  left       MBON    MBON18
#> 85 720575940614720662 79941230049538531 right       MBON    MBON03
#> 86 720575940634969707 79869899232522382  left       MBON    MBON04
#> 87 720575940621035380 81630904409213619  left       MBON    MBON09
#> 88 720575940635407352 80433124332533007  left       MBON    MBON26
#> 89 720575940617624833 80644779918075196  left       MBON    MBON07
#> 90 720575940632219300 81560466945525625  left       MBON    MBON09
#> 91 720575940644417828 81419592152467685  left       MBON    MBON06
#> 92 720575940623088937 80222636373797362 right       MBON    MBON05
#> 93 720575940632393164 78251349531447591 right       MBON    MBON19
#> 94 720575940621137334 78814711734490195 right       MBON    MBON24
#> 95 720575940635746671 79166761480079751  left       MBON    MBON25
#> 96 720575940623199684 78463280330557921 right       MBON    MBON10
#>    hemibrain_type hemibrain_match vfb_id
#> 1          MBON02       424789697   <NA>
#> 2          MBON01       612371421   <NA>
#> 3          MBON32            <NA>   <NA>
#> 4                            <NA>   <NA>
#> 5          MBON30            <NA>   <NA>
#> 6          MBON32            <NA>   <NA>
#> 7          MBON06            <NA>   <NA>
#> 8          MBON05            <NA>   <NA>
#> 9                       768555687   <NA>
#> 10         MBON20            <NA>   <NA>
#> 11                           <NA>   <NA>
#> 12         MBON02            <NA>   <NA>
#> 13         MBON33            <NA>   <NA>
#> 14         MBON23       423382015   <NA>
#> 15         MBON29      5813052909   <NA>
#> 16                      300972942   <NA>
#> 17                           <NA>   <NA>
#> 18                           <NA>   <NA>
#> 19                           <NA>   <NA>
#> 20         MBON28       457196643   <NA>
#> 21         MBON31            <NA>   <NA>
#> 22                      704466265   <NA>
#> 23         MBON35            <NA>   <NA>
#> 24                     5812981264   <NA>
#> 25                     5813068729   <NA>
#> 26                           <NA>   <NA>
#> 27         MBON30            <NA>   <NA>
#> 28         MBON27      5813061538   <NA>
#> 29                           <NA>   <NA>
#> 30                           <NA>   <NA>
#> 31                           <NA>   <NA>
#> 32                           <NA>   <NA>
#> 33                           <NA>   <NA>
#> 34                           <NA>   <NA>
#> 35                           <NA>   <NA>
#> 36         MBON03       612738462   <NA>
#> 37         MBON13            <NA>   <NA>
#> 38                     1078693835   <NA>
#> 39                           <NA>   <NA>
#> 40                           <NA>   <NA>
#> 41                      673702721   <NA>
#> 42                      642664141   <NA>
#> 43                      579916831   <NA>
#> 44                           <NA>   <NA>
#> 45                           <NA>   <NA>
#> 46                      457175171   <NA>
#> 47                     1078693835   <NA>
#> 48                           <NA>   <NA>
#> 49                     5812981264   <NA>
#> 50                           <NA>   <NA>
#> 51         MBON11            <NA>   <NA>
#> 52                     1932553362   <NA>
#> 53                           <NA>   <NA>
#> 54         MBON31            <NA>   <NA>
#> 55                     1902178951   <NA>
#> 56                           <NA>   <NA>
#> 57                           <NA>   <NA>
#> 58         MBON24            <NA>   <NA>
#> 59         MBON27      5813053891   <NA>
#> 60         MBON28            <NA>   <NA>
#> 61        LHPV4m1       518899665   <NA>
#> 62         MBON21            <NA>   <NA>
#> 63        LHPV4m1       580935628   <NA>
#> 64         MBON20            <NA>   <NA>
#> 65         MBON21       985813153   <NA>
#> 66         MBON23            <NA>   <NA>
#> 67                           <NA>   <NA>
#> 68                           <NA>   <NA>
#> 69                           <NA>   <NA>
#> 70                           <NA>   <NA>
#> 71         MBON11            <NA>   <NA>
#> 72         MBON22       985131499   <NA>
#> 73         MBON22      5813042659   <NA>
#> 74         MBON26            <NA>   <NA>
#> 75         MBON33            <NA>   <NA>
#> 76         MBON13            <NA>   <NA>
#> 77         MBON35            <NA>   <NA>
#> 78         MBON18      5813020828   <NA>
#> 79         MBON04            <NA>   <NA>
#> 80                           <NA>   <NA>
#> 81         MBON01       673509195   <NA>
#> 82         MBON29      5813058048   <NA>
#> 83                     1234386037   <NA>
#> 84         MBON18       457196444   <NA>
#> 85         MBON03       487925037   <NA>
#> 86         MBON04            <NA>   <NA>
#> 87                           <NA>   <NA>
#> 88         MBON26            <NA>   <NA>
#> 89                           <NA>   <NA>
#> 90                           <NA>   <NA>
#> 91         MBON06            <NA>   <NA>
#> 92         MBON05            <NA>   <NA>
#> 93                      423774471   <NA>
#> 94         MBON24            <NA>   <NA>
#> 95                      612742248   <NA>
#> 96                           <NA>   <NA>
# range
flytable_cell_types("MBON2[0-5]")
#>               root_id     supervoxel_id  side cell_class cell_type
#> 1  720575940632773037 80223254782485510  left       MBON    MBON20
#> 2  720575940616298635 78180912067538056 right       MBON    MBON23
#> 3  720575940637831487 79166692827138268 right       MBON    MBON25
#> 4  720575940619254744 81136880426764424  left       MBON    MBON24
#> 5  720575940612636694 78111917579300796 right       MBON    MBON21
#> 6  720575940632028753 78324741798855874 right       MBON    MBON20
#> 7  720575940625083981 81137979736901319  left       MBON    MBON21
#> 8  720575940617176925 81136880426735755  left       MBON    MBON23
#> 9  720575940604498400 80082242348953693  left       MBON    MBON22
#> 10 720575940619271128 77830305364995511 right       MBON    MBON22
#> 11 720575940621137334 78814711734490195 right       MBON    MBON24
#> 12 720575940635746671 79166761480079751  left       MBON    MBON25
#>    hemibrain_type hemibrain_match vfb_id
#> 1          MBON20            <NA>   <NA>
#> 2          MBON23       423382015   <NA>
#> 3                            <NA>   <NA>
#> 4          MBON24            <NA>   <NA>
#> 5          MBON21            <NA>   <NA>
#> 6          MBON20            <NA>   <NA>
#> 7          MBON21       985813153   <NA>
#> 8          MBON23            <NA>   <NA>
#> 9          MBON22       985131499   <NA>
#> 10         MBON22      5813042659   <NA>
#> 11         MBON24            <NA>   <NA>
#> 12                      612742248   <NA>

# include side specification
flytable_cell_types("DA2_lPN_R")
#>              root_id     supervoxel_id  side cell_class cell_type
#> 1 720575940611849187 78816429587620663 right       ALPN   DA2_lPN
#> 2 720575940622364184 78675692099364165 right       ALPN   DA2_lPN
#> 3 720575940622734835 78746060843497614 right       ALPN   DA2_lPN
#> 4 720575940626937617 78816429587622223 right       ALPN   DA2_lPN
#> 5 720575940638719104 78746060843486449 right       ALPN   DA2_lPN
#>   hemibrain_type hemibrain_match vfb_id
#> 1                     1796817841   <NA>
#> 2                     1827516355   <NA>
#> 3                      818983130   <NA>
#> 4                     1827516355   <NA>
#> 5                     1797505019   <NA>
# only the RHS MBON20
flytable_cell_types("MBON20_R")
#>              root_id     supervoxel_id  side cell_class cell_type
#> 1 720575940632028753 78324741798855874 right       MBON    MBON20
#>   hemibrain_type hemibrain_match vfb_id
#> 1         MBON20            <NA>   <NA>
# all RHS cells with class MBON
flytable_cell_types("MBON_R", target="cell_class")
#>               root_id     supervoxel_id  side cell_class cell_type
#> 1  720575940630697303 79448854932107696 right       MBON    MBON02
#> 2  720575940625991268 78603948898007978 right       MBON    MBON32
#> 3  720575940619259352 78111986298788517 right       MBON    MBON06
#> 4  720575940612772579 78393323903110735 right       MBON    MBON10
#> 5  720575940628764495 79588561627810154 right       MBON    MBON02
#> 6  720575940616298635 78180912067538056 right       MBON    MBON23
#> 7  720575940608410889 78040449457502610 right       MBON    MBON28
#> 8  720575940624753336 79166830266640003 right       MBON    MBON12
#> 9  720575940621032052 78181599128215629 right       MBON    MBON35
#> 10 720575940638774606 78040449457461921 right       MBON    MBON17
#> 11 720575940625168528 78251349531485275 right       MBON    MBON19
#> 12 720575940639693144 77901086224032815 right       MBON    MBON30
#> 13 720575940620362945 77408917331762608 right       MBON    MBON09
#> 14 720575940632656173 77549311222677157 right       MBON    MBON10
#> 15 720575940628581379 78111986298850429 right       MBON    MBON10
#> 16 720575940617413957 77478942478556789 right       MBON    MBON10
#> 17 720575940617573922 77549311155597154 right       MBON    MBON10
#> 18 720575940613157329 78111986365505691 right       MBON    MBON10
#> 19 720575940627091210 79380479052758063 right       MBON    MBON13
#> 20 720575940623743415 77901154877073717 right       MBON    MBON09
#> 21 720575940638910141 78393117744840114 right       MBON    MBON10
#> 22 720575940625850826 78251486970442209 right       MBON    MBON15
#> 23 720575940630373113 76985537105536911 right       MBON    MBON15
#> 24 720575940622281430 78040449457463315 right       MBON    MBON15
#> 25 720575940637831487 79166692827138268 right       MBON    MBON25
#> 26 720575940613666001 78110818201651448 right       MBON    MBON15
#> 27 720575940617869083 77760348668670642 right       MBON    MBON09
#> 28 720575940636311694 78110818201681795 right       MBON    MBON17
#> 29 720575940630387820 78182286188838645 right       MBON    MBON31
#> 30 720575940627584516 78110818201621392 right       MBON    MBON16
#> 31 720575940622268374 79027192356412856 right       MBON    MBON27
#> 32 720575940615186079 77899986914253830 right       MBON    MBON15
#> 33 720575940612636694 78111917579300796 right       MBON    MBON21
#> 34 720575940632028753 78324741798855874 right       MBON    MBON20
#> 35 720575940617125081 79309972869657633 right       MBON    MBON12
#> 36 720575940612637974 78603949099622269 right       MBON    MBON07
#> 37 720575940654692001 79380204241413766 right       MBON    MBON14
#> 38 720575940625994596 79309904216692475 right       MBON    MBON14
#> 39 720575940628587779 78323023677138555 right       MBON    MBON11
#> 40 720575940619271128 77830305364995511 right       MBON    MBON22
#> 41 720575940619720277 78885355625276424 right       MBON    MBON26
#> 42 720575940631031055 78463211409646517 right       MBON    MBON33
#> 43 720575940624744462 78040449457553649 right       MBON    MBON18
#> 44 720575940631940236 80011530007496199 right       MBON    MBON04
#> 45 720575940623204548 79870998677653042 right       MBON    MBON01
#> 46 720575940630390265 79237542540872430 right       MBON    MBON29
#> 47 720575940630691415 78604086538417636 right       MBON    MBON07
#> 48 720575940614720662 79941230049538531 right       MBON    MBON03
#> 49 720575940623088937 80222636373797362 right       MBON    MBON05
#> 50 720575940632393164 78251349531447591 right       MBON    MBON19
#> 51 720575940621137334 78814711734490195 right       MBON    MBON24
#> 52 720575940623199684 78463280330557921 right       MBON    MBON10
#>    hemibrain_type hemibrain_match vfb_id
#> 1          MBON02       424789697   <NA>
#> 2          MBON32            <NA>   <NA>
#> 3          MBON06            <NA>   <NA>
#> 4                       768555687   <NA>
#> 5          MBON02            <NA>   <NA>
#> 6          MBON23       423382015   <NA>
#> 7          MBON28       457196643   <NA>
#> 8                       704466265   <NA>
#> 9          MBON35            <NA>   <NA>
#> 10                     5812981264   <NA>
#> 11                           <NA>   <NA>
#> 12         MBON30            <NA>   <NA>
#> 13                           <NA>   <NA>
#> 14                           <NA>   <NA>
#> 15                           <NA>   <NA>
#> 16                           <NA>   <NA>
#> 17                           <NA>   <NA>
#> 18                           <NA>   <NA>
#> 19         MBON13            <NA>   <NA>
#> 20                     1078693835   <NA>
#> 21                           <NA>   <NA>
#> 22                      673702721   <NA>
#> 23                      642664141   <NA>
#> 24                      579916831   <NA>
#> 25                           <NA>   <NA>
#> 26                      457175171   <NA>
#> 27                     1078693835   <NA>
#> 28                     5812981264   <NA>
#> 29         MBON31            <NA>   <NA>
#> 30                           <NA>   <NA>
#> 31         MBON27      5813053891   <NA>
#> 32        LHPV4m1       518899665   <NA>
#> 33         MBON21            <NA>   <NA>
#> 34         MBON20            <NA>   <NA>
#> 35                           <NA>   <NA>
#> 36                           <NA>   <NA>
#> 37                           <NA>   <NA>
#> 38                           <NA>   <NA>
#> 39         MBON11            <NA>   <NA>
#> 40         MBON22      5813042659   <NA>
#> 41         MBON26            <NA>   <NA>
#> 42         MBON33            <NA>   <NA>
#> 43         MBON18      5813020828   <NA>
#> 44         MBON04            <NA>   <NA>
#> 45         MBON01       673509195   <NA>
#> 46         MBON29      5813058048   <NA>
#> 47                     1234386037   <NA>
#> 48         MBON03       487925037   <NA>
#> 49         MBON05            <NA>   <NA>
#> 50                      423774471   <NA>
#> 51         MBON24            <NA>   <NA>
#> 52                           <NA>   <NA>
# }