Low level functions to access the flytable metadata service

flytable_login uses your flytable user name and email to log into the service.

flytable_set_token will obtain and store a permanent seatable user-level API token.

flytable_base returns a base object (equivalent to a mysql database) which allows you to access one or more tables, logging in to the service if necessary. The returned base object give you full access to the Python Base API allowing a range of row/column manipulations.

flytable_alltables lists all tables across all flytables bases.

flytable_login(
  url = "https://flytable.mrc-lmb.cam.ac.uk/",
  token = Sys.getenv("FLYTABLE_TOKEN", unset = NA_character_)
)

flytable_set_token(user, pwd, url = "https://flytable.mrc-lmb.cam.ac.uk/")

flytable_base(
  table = NULL,
  base_name = NULL,
  workspace_id = NULL,
  url = "https://flytable.mrc-lmb.cam.ac.uk/",
  cached = FALSE
)

flytable_alltables(ac = NULL, cached = TRUE)

Arguments

url
Optional URL to the server

token
normally retrieved from FLYTABLE_TOKEN environment variable.

user, pwd
flytable user and password used by flytable_set_token to obtain a token

table
Character vector specifying a table foe which you want a base object.

base_name
Character vector specifying the base

workspace_id
A numeric id specifying the workspace. Advanced use only since we can normally figure this out from base_name.

cached
Whether to use a cached version of the response if available. Set to FALSE if you know tables have been added or renamed during your session.

ac
Optional account object returned by flytables_login

Value

For flytable_login, a Python Account object from the seatable api as wrapped by reticulate.

For flytable_base, a Python Base object from the seatable api as wrapped by reticulate.

A data.frame containing the base_name, workspace_id, table name and table _id.

Details

Besides initial setup (next paragraph), you should not need to use these lower level functions directly. Instead we recommend higher level functions such as flytable_query.

In order to start using flytable, you must get an API token. There doesn't seem to be a convenient way to do this from the seatable web interface but you can get one by calling flytable_set_token with your flytable user and password. This should be a once only step. Thereafter you should have a FLYTABLE_TOKEN environment variable set in your .Renviron file.

flytable_base will use your flytable API token to log into the service.

See also

Examples

if (FALSE) {
flytable_login()
}
if (FALSE) {
flytable_set_token(user='xxx@gmail.com', pwd='yyy')
}
if (FALSE) {
hemilineages=flytable_base(base_name='hemilineages')
# equivalent, but simpler since you only have to remember the table name
hemilineages=flytable_base('fafb_hemilineages_survey')
}

# \donttest{
flytable_alltables()

#>      base_name workspace_id                     name  _id
#> 1 hemilineages            5 fafb_hemilineages_survey 0000
#> 2 hemilineages            5                testfruit diIw
#> 3 hemilineages            5                column_ss d4aS
#> 4   LH_library            5                LH_bodies dvGc

# }