Update rows in a flytable database

Update rows in a flytable database

flytable_update_rows(df, table, base = NULL, chunksize = 1000L, ...)

Arguments

df
A data.frame containing the data to upload including an _id column that can identify each row in the remote table.

table
Character vector naming a table

base
Character vector naming a seatable base (recommended) or a Base object returned by flytable_base (expert use).

chunksize
To split large requests into smaller ones with max this many rows.

...
Additional arguments passed to pbsapply which might include cl=2 to specify a number of parallel jobs to run.

Value

Logical indicating success, invisibly (failures will normally cause premature termination with errors written to the console).

Details

seatable automatically maintains a unique id for each row in a _id column. This is returned by flytable_query and friends. If you modify data and then want to update again, you need to keep the column containing this row _id.

The chunksize argument is required because it seems that there is a maximum of 1000 rows per update action.

See also

Examples

# \donttest{
fruit=flytable_list_rows('testfruit')
flytable_update_rows(table='testfruit', fruit[c(1,4:6)])

#> [1] TRUE

# }