Low level access to the Flywire CAVE annotation system

Low level access to the Flywire CAVE annotation system

Usage

flywire_cave_client(
  datastack_name = getOption("fafbseg.cave.datastack_name", "flywire_fafb_production")
)

Arguments

datastack_name

defaults to "flywire_fafb_production". See https://global.daf-apis.com/info/ for other options.

Value

The caveclient.frameworkclient.CAVEclientFull Python module wrapped by reticulate.

Details

This depends on installation of the Python caveclient library. See flywire_cave_query for more details.

Examples

# \donttest{
fac <- flywire_cave_client()
fac$annotation$get_tables()
#> Error in py_call_impl(callable, call_args$unnamed, call_args$named): <...truncated...>n    aligned_volume_names = infoclient.get_aligned_volumes()\\n", "  File \\"/usr/local/lib/python3.9/site-packages/caveclient/infoservice.py\\", line 187, in get_aligned_volumes\\n    return handle_response(response)\\n", "  File \\"/usr/local/lib/python3.9/site-packages/caveclient/base.py\\", line 81, in handle_response\\n    _raise_for_status(response, log_warning=log_warning)\\n", "  File \\"/usr/local/lib/python3.9/site-packages/caveclient/base.py\\", line 72, in _raise_for_status\\n    raise requests.HTTPError(http_error_msg, response=r)\\n", "requests.exceptions.HTTPError: 502 Server Error: Bad Gateway for url: https://global.daf-apis.com/info/api/v2/aligned_volume content:b\'<html>\\\\r\\\\n<head><title>502 Bad Gateway</title></head>\\\\r\\\\n<body>\\\\r\\\\n<center><h1>502 Bad Gateway</h1></center>\\\\r\\\\n<hr><center>nginx/1.23.2</center>\\\\r\\\\n</body>\\\\r\\\\n</html>\\\\r\\\\n\'\\n"]}\n'
#> Run `reticulate::py_last_error()` for details.
fac$annotation$get_table_metadata('nuclei_v1')
#> Error in py_call_impl(callable, call_args$unnamed, call_args$named): <...truncated...>n    aligned_volume_names = infoclient.get_aligned_volumes()\\n", "  File \\"/usr/local/lib/python3.9/site-packages/caveclient/infoservice.py\\", line 187, in get_aligned_volumes\\n    return handle_response(response)\\n", "  File \\"/usr/local/lib/python3.9/site-packages/caveclient/base.py\\", line 81, in handle_response\\n    _raise_for_status(response, log_warning=log_warning)\\n", "  File \\"/usr/local/lib/python3.9/site-packages/caveclient/base.py\\", line 72, in _raise_for_status\\n    raise requests.HTTPError(http_error_msg, response=r)\\n", "requests.exceptions.HTTPError: 502 Server Error: Bad Gateway for url: https://global.daf-apis.com/info/api/v2/aligned_volume content:b\'<html>\\\\r\\\\n<head><title>502 Bad Gateway</title></head>\\\\r\\\\n<body>\\\\r\\\\n<center><h1>502 Bad Gateway</h1></center>\\\\r\\\\n<hr><center>nginx/1.23.2</center>\\\\r\\\\n</body>\\\\r\\\\n</html>\\\\r\\\\n\'\\n"]}\n'
#> Run `reticulate::py_last_error()` for details.

# annotation tables that have been materialised in order to map XYZ
# points onto current root ids (and immutable supervoxel ids)
# typically the same as fac$annotation$get_tables()
fac$materialize$get_tables()
#>  [1] "proofread_neurons"                   
#>  [2] "fly_synapses_neuropil_v2"            
#>  [3] "synaptic_clefts"                     
#>  [4] "hierarchical_neuron_annotations"     
#>  [5] "fly_synapses_neuropil_v6"            
#>  [6] "neuron_information_v2"               
#>  [7] "cambridge_celltypes_v2"              
#>  [8] "fly_synapses_neuropil_v5"            
#>  [9] "neuron_information_vtest"            
#> [10] "proofreading_status_public_v1"       
#> [11] "synapses_nt_v1"                      
#> [12] "fly_synapses_neuropil"               
#> [13] "proofreading_review_public_vtest"    
#> [14] "braincircuits_annotation_test"       
#> [15] "valid_synapses_nt_vtest"             
#> [16] "nuclei_v1"                           
#> [17] "valid_synapses_nt_v2"                
#> [18] "valid_synapses_nt_v1"                
#> [19] "fly_synapses_neuropil_v4"            
#> [20] "hemilineages"                        
#> [21] "neuron_representative_coordinates_v1"
#> [22] "cambridge_celltypes"                 
#> [23] "fly_synapses_neuropil_v3"            
#> [24] "proofreading_review_public_v1"       

info=fac$info$get_datastack_info()
# the default synapse table for the dataset
info$synapse_table
#> [1] "synapses_nt_v1"
# }