Query the FlyWire CAVE annotation system

Query the FlyWire CAVE annotation system

flywire_cave_query(
  table,
  datastack_name = getOption("fafbseg.cave.datastack_name", "flywire_fafb_production"),
  version = NULL,
  timestamp = NULL,
  live = is.null(version) && is.null(timestamp),
  filter_in_dict = NULL,
  filter_out_dict = NULL,
  ...
)

Arguments

table

The name of the table to query

datastack_name

defaults to "flywire_fafb_production". See https://global.daf-apis.com/info/ for other options.

version

An optional CAVE materialisation version number. See details and examples.

timestamp

An optional timestamp as a string or POSIXct, interpreted as UTC when no timezone is specified.

live

Whether to use live query mode, which updates any root ids to their current value.

filter_in_dict, filter_out_dict

Optional arguments consisting of key value lists that restrict the returned rows (keeping only matches or filtering out matches). See examples and CAVE documentation for details.

...

Additional arguments to the query method. See examples and details.

Value

A tibble. Note that xyzmatrix can be used on single columns containing XYZ locations.

Details

CAVE (Connectome Annotation Versioning Engine) provides a shared infrastructure for a number of connectomics projects involving Sebastian Seung's group at Princeton and collaborators at the Allen Institute. There is both a backend system running on their servers and a Python client for end users.

You can find out more at https://caveclient.readthedocs.io/ as well as looking at the Python notebooks on the github repo https://github.com/seung-lab/CAVEclient.

The annotation system shares authentication infrastructure with the rest of the FlyWire API (see flywire_set_token).

CAVE has a concept of table snapshots identified by an integer materialization version number. In some cases you may wish to query a table at this defined version number so that you can avoid root_ids changing during an analysis. Your calls will also be faster since no root id updates are required.

Note however that materialisation versions expire at which point the corresponding version of the database is expunged. However it is still possible to find the timestamp for an expired materialisation version. flywire_cave_query does this automatically using flywire_timestamp. In these circumstances queries will again be slower (quite possibly slower than the live query) since all root ids must be recalculated to match the timestamp.

See also

Examples

# \donttest{
# note use of limit to restrict the number of rows (must be integer)
n10=flywire_cave_query(table = 'nuclei_v1', limit=10L)
head(as.data.frame(n10))
#>        id valid    volume  pt_supervoxel_id         pt_root_id
#> 1 7393349     t 26.141245 82827379285852979 720575940626838909
#> 2 7416439     t 11.523400 82827998029690530 720575940627484553
#> 3 7415038     t 32.895959 83038623024880664 720575940626046919
#> 4 7415013     t 53.711176 83038760463398837 720575940630311383
#> 5 7415848     t  9.280717 83038554439606237 720575940633370649
#> 6 7415851     t 19.018342 83038691811464073 720575940638709326
#>              pt_position      bb_start_position        bb_end_position
#> 1  709888, 227744, 57160  708032, 226144, 54760  711904, 229440, 59280
#> 2 710592, 263392, 129800 708800, 262112, 128200 712064, 264832, 131080
#> 3  722528, 234656, 77000  720768, 232896, 74480  724224, 236576, 79360
#> 4  722912, 244032, 65200  720480, 242144, 62480  725600, 246208, 67760
#> 5 721984, 229792, 119560 720832, 228320, 118240 723232, 231264, 120840
#> 6  722432, 239520, 98440  720864, 237888, 96200 724256, 241056, 100760
# }
if (FALSE) {
nuclei_v1=flywire_cave_query(table = 'nuclei_v1')
points3d(xyzmatrix(nuclei_v1$pt_position))

library(elmr)
# calculate signed distance to FAFB surface
# NB this surface is not a perfect fit in the optic lobes
nuclei_v1$d=pointsinside(xyzmatrix(nuclei_v1$pt_position), FAFB.surf,
  rval = 'dist')
points3d(xyzmatrix(nuclei_v1$pt_position),
  col=matlab::jet.colors(20)[cut(nuclei_v1$d,20)])
plot3d(FAFB)
}
# Example of a query on a table
if (FALSE) {
# the Princeton (mostly) and Cambridge groups have tagged some bodies as
# not a neuron - these are often glia.
nans=flywire_cave_query('neuron_information_v2',
  filter_in_dict = list(tag='not a neuron'))
nrow(nans)
table(nans$user_id)
}
if (FALSE) {
psp_351=flywire_cave_query(table = 'proofreading_status_public_v1',
  version=351)
# get the last listed materialisation version
fcc=flywire_cave_client()
lastv=tail(fcc$materialize$get_versions(), n=1)
# pull that
psp_last=flywire_cave_query(table = 'proofreading_status_public_v1',
  version=lastv)
}