Summarise edits for one or more FlyWire segment ids

Summarise edits for one or more FlyWire segment ids

flywire_change_log(
  x,
  root_ids = FALSE,
  filtered = TRUE,
  tz = "UTC",
  OmitFailures = TRUE,
  ...
)

Arguments

x
One or more flywire segment ids (character vector strongly recommended but any format accepted by ngl_segments will work).

root_ids
Whether to look up the root ids before/after each change. Default FALSE. NB this is quite resource intensive so please do not flood the server with requests of this sort.

filtered
Whether to filter out edits unlikely to relate to the current state of the neuron (default TRUE, see details).

tz
Time zone for edit timestamps. Defaults to "UTC" i.e. Universal Time, Coordinated. Set to "" for your current timezone. See as.POSIXct for more details.

OmitFailures
Whether to omit neurons for which there is an API timeout or error. The default value (TRUE) will skip over errors, while NA) will result in a hard stop on error. See nlapply for more details.

...
Additional arguments passed to flywire_fetch

Value

A data frame with values itemize

In addition when filtered=FALSE, in_neuron is_relevant

In addition when root_ids=TRUE, before_root_ids after_root_ids, as space separated strings.

Details

When filtered=TRUE the following are removed

edits that are outside of the given root id - meaning these edits are in the history of the cell but the affected part of the arbour was since cut off.

edits that were since undone by an inverse operation (often temporary splits made while reviewing a neuron.)

Examples

# \donttest{
flywire_change_log("720575940619010932")

#>   operation_id           timestamp user_id is_merge        user_name
#> 0       194549 2020-09-03 19:33:23     397    FALSE Marlon Blanquart
#> 1       194550 2020-09-03 19:37:26     397     TRUE Marlon Blanquart
#> 2       194551 2020-09-03 19:38:12     397    FALSE Marlon Blanquart
#> 3       194552 2020-09-03 19:40:33     397     TRUE Marlon Blanquart
#> 4       194582 2020-09-03 20:05:41     397     TRUE Marlon Blanquart
#> 5       194584 2020-09-03 20:06:51     397    FALSE Marlon Blanquart

flywire_change_log("720575940619010932", root_ids = TRUE)

#>   operation_id           timestamp user_id is_merge
#> 0       194549 2020-09-03 19:33:23     397    FALSE
#> 1       194550 2020-09-03 19:37:26     397     TRUE
#> 2       194551 2020-09-03 19:38:12     397    FALSE
#> 3       194552 2020-09-03 19:40:33     397     TRUE
#> 4       194582 2020-09-03 20:05:41     397     TRUE
#> 5       194584 2020-09-03 20:06:51     397    FALSE
#>                         before_root_ids                        after_root_ids
#> 0                    720575940615250973 720575940611598545 720575940625952497
#> 1 720575940622305363 720575940625952497                    720575940603341158
#> 2                    720575940603341158 720575940618539430 720575940620629528
#> 3 720575940620629528 720575940637849454                    720575940625315922
#> 4 720575940625315922 720575940635457646                    720575940610162858
#> 5                    720575940610162858 720575940618698480 720575940619010932
#>          user_name
#> 0 Marlon Blanquart
#> 1 Marlon Blanquart
#> 2 Marlon Blanquart
#> 3 Marlon Blanquart
#> 4 Marlon Blanquart
#> 5 Marlon Blanquart

flywire_change_log("720575940619010932", filtered = FALSE)

#>   operation_id           timestamp user_id is_merge in_neuron is_relevant
#> 0       194549 2020-09-03 19:33:23     397    FALSE      TRUE        TRUE
#> 1       194550 2020-09-03 19:37:26     397     TRUE      TRUE        TRUE
#> 2       194551 2020-09-03 19:38:12     397    FALSE      TRUE        TRUE
#> 3       194552 2020-09-03 19:40:33     397     TRUE      TRUE        TRUE
#> 4       194582 2020-09-03 20:05:41     397     TRUE      TRUE        TRUE
#> 5       194584 2020-09-03 20:06:51     397    FALSE      TRUE        TRUE
#>          user_name
#> 0 Marlon Blanquart
#> 1 Marlon Blanquart
#> 2 Marlon Blanquart
#> 3 Marlon Blanquart
#> 4 Marlon Blanquart
#> 5 Marlon Blanquart

# with a flywire URL
u="https://ngl.flywire.ai/?json_url=https://globalv1.flywire-daf.com/nglstate/5409525645443072"
flywire_change_log(u)

#>                         id operation_id           timestamp user_id is_merge
#> 10      720575940621039145         1710 2019-08-03 23:17:28       4    FALSE
#> 11      720575940621039145         1715 2019-08-03 23:21:22       4    FALSE
#> 12      720575940621039145         1716 2019-08-03 23:22:16       4    FALSE
#> 31      720575940621039145         1735 2019-08-03 23:31:40       4    FALSE
#> 32      720575940621039145         1736 2019-08-03 23:31:57       4    FALSE
#> 33      720575940621039145         1737 2019-08-03 23:32:08       4    FALSE
#> 34      720575940621039145         1738 2019-08-03 23:32:20       4    FALSE
#> 35      720575940621039145         1739 2019-08-03 23:32:30       4    FALSE
#> 38      720575940621039145         1742 2019-08-03 23:33:22       4    FALSE
#> 40      720575940621039145         1744 2019-08-03 23:33:56       4    FALSE
#> 44...11 720575940621039145         1748 2019-08-03 23:34:34       4    FALSE
#> 46      720575940621039145         1750 2019-08-03 23:34:48       4    FALSE
#> 48      720575940621039145         1752 2019-08-03 23:35:15       4    FALSE
#> 49      720575940621039145         1753 2019-08-03 23:35:31       4    FALSE
#> 51      720575940621039145         1755 2019-08-03 23:35:50       4    FALSE
#> 53      720575940621039145         1757 2019-08-03 23:36:12       4    FALSE
#> 55      720575940621039145         1759 2019-08-03 23:36:26       4    FALSE
#> 57      720575940621039145         1761 2019-08-03 23:36:45       4    FALSE
#> 59      720575940621039145         1763 2019-08-03 23:37:06       4    FALSE
#> 62      720575940621039145         1766 2019-08-03 23:37:33       4    FALSE
#> 64      720575940621039145         1768 2019-08-03 23:37:55       4    FALSE
#> 66      720575940621039145         1770 2019-08-03 23:38:14       4    FALSE
#> 67      720575940621039145         1771 2019-08-03 23:38:27       4    FALSE
#> 72...24 720575940621039145         1776 2019-08-03 23:39:39       4    FALSE
#> 74...25 720575940621039145         1778 2019-08-03 23:42:10       4    FALSE
#> 76...26 720575940621039145         1781 2019-08-03 23:47:50       4    FALSE
#> 77...27 720575940621039145         1782 2019-08-03 23:49:56       4     TRUE
#> 78      720575940621039145         1784 2019-08-03 23:55:56       4    FALSE
#> 79      720575940621039145         1785 2019-08-03 23:56:10       4    FALSE
#> 80...30 720575940621039145         1786 2019-08-03 23:56:36       4    FALSE
#> 81...31 720575940621039145         1787 2019-08-03 23:56:48       4    FALSE
#> 82...32 720575940621039145         1788 2019-08-03 23:57:10       4    FALSE
#> 94      720575940621039145         1896 2019-08-04 02:42:31       4    FALSE
#> 120     720575940621039145         2231 2019-08-08 16:45:15       4    FALSE
#> 151     720575940621039145         2300 2019-08-08 19:39:12       4    FALSE
#> 159     720575940621039145         2346 2019-08-09 21:51:43       4    FALSE
#> 164     720575940621039145         2377 2019-08-12 18:19:12       4    FALSE
#> 165     720575940621039145         2423 2019-08-14 17:16:15       4    FALSE
#> 166     720575940621039145         2742 2019-08-16 17:14:54       4     TRUE
#> 185     720575940621039145         8042 2019-09-05 23:10:59       4    FALSE
#> 186     720575940621039145         8046 2019-09-05 23:22:12       4    FALSE
#> 187     720575940621039145         8047 2019-09-05 23:26:27       4    FALSE
#> 188     720575940621039145         8049 2019-09-05 23:29:06       4    FALSE
#> 189     720575940621039145         8052 2019-09-05 23:32:15       4    FALSE
#> 204     720575940621039145         8085 2019-09-06 00:28:46       4    FALSE
#> 205     720575940621039145         8086 2019-09-06 00:30:04       4    FALSE
#> 207     720575940621039145         8091 2019-09-06 00:34:02       4     TRUE
#> 208     720575940621039145        11049 2019-09-16 11:57:40      13    FALSE
#> 209     720575940621039145        11194 2019-09-16 19:22:17      13    FALSE
#> 211     720575940621039145        20627 2019-10-18 18:04:25      16    FALSE
#> 213     720575940621039145        40816 2020-01-07 18:48:38      13    FALSE
#> 214     720575940621039145        40819 2020-01-07 18:50:06      13     TRUE
#> 216     720575940621039145        52005 2020-02-03 18:42:45      26     TRUE
#> 217     720575940621039145        52065 2020-02-03 19:42:24      26    FALSE
#> 220     720575940621039145        52135 2020-02-03 20:34:35      26     TRUE
#> 221     720575940621039145        52141 2020-02-03 20:35:32      26     TRUE
#> 222     720575940621039145        52142 2020-02-03 20:36:01      26     TRUE
#> 223     720575940621039145        52146 2020-02-03 20:40:29      26    FALSE
#> 1       720575940626877799         4729 2019-08-23 21:17:22      12     TRUE
#> 44...60 720575940626877799         6163 2019-08-28 20:48:48      12    FALSE
#> 68      720575940626877799        10877 2019-09-13 16:00:25      13    FALSE
#> 69      720575940626877799        10903 2019-09-13 16:52:47      13    FALSE
#> 70      720575940626877799        10904 2019-09-13 16:57:06      13     TRUE
#> 71      720575940626877799        10906 2019-09-13 16:58:52      13    FALSE
#> 72...65 720575940626877799        10907 2019-09-13 17:01:45      13    FALSE
#> 73      720575940626877799        10909 2019-09-13 17:11:08      13    FALSE
#> 74...67 720575940626877799        10912 2019-09-13 17:14:06      13    FALSE
#> 75      720575940626877799        10937 2019-09-13 17:37:41      13    FALSE
#> 76...69 720575940626877799        10939 2019-09-13 17:43:53      13     TRUE
#> 77...70 720575940626877799        10940 2019-09-13 17:44:28      13     TRUE
#> 80...71 720575940626877799        10947 2019-09-13 18:00:31      13    FALSE
#> 81...72 720575940626877799        10949 2019-09-13 18:04:35      13    FALSE
#> 82...73 720575940626877799        10952 2019-09-13 18:08:03      13    FALSE
#> 85      720575940626877799        10956 2019-09-13 18:12:46      13    FALSE
#> 88      720575940626877799        10967 2019-09-13 18:19:11      13    FALSE
#> 89      720575940626877799        10969 2019-09-13 18:20:48      13     TRUE
#> 90      720575940626877799        10971 2019-09-13 18:22:16      13    FALSE
#> 91      720575940626877799        10972 2019-09-13 18:24:20      13    FALSE
#> 92      720575940626877799        10981 2019-09-13 18:26:26      13     TRUE
#> 100     720575940626877799        11363 2019-09-17 14:53:32      17    FALSE
#> 101     720575940626877799        11414 2019-09-17 16:03:08      17    FALSE
#> 300     720575940626877799        33530 2019-12-13 19:16:56      84    FALSE
#> 369     720575940626877799        46345 2020-01-17 18:27:27      84    FALSE
#> 370     720575940626877799        46350 2020-01-17 18:28:49      84    FALSE
#> 371     720575940626877799        46351 2020-01-17 18:30:32      84    FALSE
#> 372     720575940626877799        46352 2020-01-17 18:33:23      84    FALSE
#> 507     720575940626877799       107453 2020-04-27 18:44:42     104    FALSE
#> 508     720575940626877799       107454 2020-04-27 18:44:46     104    FALSE
#> 509     720575940626877799       107455 2020-04-27 18:44:53     104    FALSE
#> 510     720575940626877799       107456 2020-04-27 18:45:00     104    FALSE
#> 513     720575940626877799       107460 2020-04-27 18:45:43     104     TRUE
#> 514     720575940626877799       107461 2020-04-27 18:45:46     104     TRUE
#> 519     720575940626877799       107466 2020-04-27 18:46:08     104     TRUE
#> 520     720575940626877799       107467 2020-04-27 18:46:14     104     TRUE
#> 521     720575940626877799       107468 2020-04-27 18:48:06     104    FALSE
#> 522     720575940626877799       107480 2020-04-27 19:03:54     104    FALSE
#> 523     720575940626877799       107482 2020-04-27 19:07:39     104     TRUE
#> 524     720575940626877799       107487 2020-04-27 19:17:14     104    FALSE
#> 525     720575940626877799       107494 2020-04-27 19:30:34     104    FALSE
#> 526     720575940626877799       107495 2020-04-27 19:30:47     104     TRUE
#> 527     720575940626877799       107557 2020-04-27 20:40:33     104    FALSE
#> 528     720575940626877799       107589 2020-04-27 21:10:52     104    FALSE
#> 529     720575940626877799       107590 2020-04-27 21:11:03     104    FALSE
#> 530     720575940626877799       107592 2020-04-27 21:11:47     104    FALSE
#> 531     720575940626877799       107603 2020-04-27 21:20:09     104     TRUE
#> 532     720575940626877799       107604 2020-04-27 21:20:13     104     TRUE
#> 533     720575940626877799       107635 2020-04-27 21:31:33     104    FALSE
#> 534     720575940626877799       107650 2020-04-27 22:15:31     104    FALSE
#> 535     720575940626877799       107651 2020-04-27 22:15:46     104    FALSE
#> 536     720575940626877799       107652 2020-04-27 22:15:51     104    FALSE
#> 537     720575940626877799       107653 2020-04-27 22:15:57     104    FALSE
#> 538     720575940626877799       107656 2020-04-27 22:27:41     104    FALSE
#> 539     720575940626877799       107657 2020-04-27 22:29:06     104    FALSE
#> 540     720575940626877799       107660 2020-04-27 22:31:44     104     TRUE
#> 541     720575940626877799       107661 2020-04-27 22:37:56     104    FALSE
#> 542     720575940626877799       107667 2020-04-27 22:57:59     104    FALSE
#> 543     720575940626877799       107668 2020-04-27 22:58:05     104    FALSE
#> 544     720575940626877799       107699 2020-04-27 23:36:50     104    FALSE
#> 545     720575940626877799       107710 2020-04-27 23:42:21     104    FALSE
#> 547     720575940626877799       107726 2020-04-28 00:43:21     104    FALSE
#> 548     720575940626877799       107729 2020-04-28 00:48:13     104    FALSE
#> 549     720575940626877799       107735 2020-04-28 01:00:38     104    FALSE
#> 550     720575940626877799       107737 2020-04-28 01:09:32     104    FALSE
#> 551     720575940626877799       107738 2020-04-28 01:11:31     104    FALSE
#> 552     720575940626877799       107739 2020-04-28 01:12:12     104    FALSE
#> 554     720575940626877799       107742 2020-04-28 01:21:57     104    FALSE
#> 555     720575940626877799       107766 2020-04-28 01:47:17     104    FALSE
#> 556     720575940626877799       107767 2020-04-28 01:47:22     104    FALSE
#> 557     720575940626877799       107783 2020-04-28 02:06:45     104    FALSE
#> 558     720575940626877799       107797 2020-04-28 02:55:10     104    FALSE
#> 559     720575940626877799       107798 2020-04-28 02:59:17     104    FALSE
#> 560     720575940626877799       107799 2020-04-28 03:01:32     104    FALSE
#> 561     720575940626877799       107800 2020-04-28 03:02:02     104    FALSE
#> 562     720575940626877799       107801 2020-04-28 03:03:27     104    FALSE
#> 563     720575940626877799       107803 2020-04-28 03:08:37     104    FALSE
#> 565     720575940626877799       107806 2020-04-28 03:22:03     104     TRUE
#> 566     720575940626877799       107807 2020-04-28 03:22:07     104     TRUE
#> 567     720575940626877799       107808 2020-04-28 03:22:49     104     TRUE
#> 568     720575940626877799       107809 2020-04-28 03:22:52     104     TRUE
#> 569     720575940626877799       107810 2020-04-28 03:22:56     104     TRUE
#> 570     720575940626877799       107948 2020-04-28 18:23:53     104    FALSE
#> 571     720575940626877799       107949 2020-04-28 18:24:11     104     TRUE
#> 572     720575940626877799       107950 2020-04-28 18:34:57     104    FALSE
#> 573     720575940626877799       107966 2020-04-28 19:01:09     104    FALSE
#> 574     720575940626877799       107981 2020-04-28 19:07:41     104    FALSE
#> 578     720575940626877799       108009 2020-04-28 19:33:43     104    FALSE
#> 579     720575940626877799       108010 2020-04-28 19:33:53     104    FALSE
#> 580     720575940626877799       108012 2020-04-28 19:34:00     104    FALSE
#> 581     720575940626877799       108013 2020-04-28 19:34:03     104    FALSE
#> 582     720575940626877799       108014 2020-04-28 19:34:06     104    FALSE
#> 587     720575940626877799       108034 2020-04-28 20:22:10     104    FALSE
#> 588     720575940626877799       108036 2020-04-28 20:31:23     104    FALSE
#> 589     720575940626877799       108037 2020-04-28 20:32:38     104     TRUE
#> 590     720575940626877799       108042 2020-04-28 20:40:41     104    FALSE
#> 593     720575940626877799       108046 2020-04-28 20:46:59     104    FALSE
#> 594     720575940626877799       108054 2020-04-28 20:56:18     104    FALSE
#> 595     720575940626877799       108056 2020-04-28 20:56:26     104    FALSE
#> 596     720575940626877799       108057 2020-04-28 20:59:15     104     TRUE
#> 597     720575940626877799       108058 2020-04-28 20:59:18     104     TRUE
#> 598     720575940626877799       108059 2020-04-28 21:01:17     104    FALSE
#> 599     720575940626877799       108060 2020-04-28 21:01:32     104     TRUE
#> 600     720575940626877799       108070 2020-04-28 21:26:49     104    FALSE
#> 601     720575940626877799       108071 2020-04-28 21:28:23     104     TRUE
#> 602     720575940626877799       108072 2020-04-28 21:28:24     104     TRUE
#> 603     720575940626877799       108073 2020-04-28 21:28:28     104     TRUE
#> 604     720575940626877799       108094 2020-04-28 21:58:38     104    FALSE
#> 605     720575940626877799       108095 2020-04-28 21:59:26     104    FALSE
#> 606     720575940626877799       108101 2020-04-29 00:47:33     104    FALSE
#> 607     720575940626877799       108102 2020-04-29 00:53:33     104    FALSE
#> 608     720575940626877799       108107 2020-04-29 01:16:53     104    FALSE
#> 609     720575940626877799       108109 2020-04-29 01:17:03     104    FALSE
#> 614     720575940626877799       108122 2020-04-29 01:32:15     104    FALSE
#> 615     720575940626877799       108128 2020-04-29 01:37:54     104    FALSE
#> 616     720575940626877799       108129 2020-04-29 01:38:22     104     TRUE
#> 617     720575940626877799       108130 2020-04-29 01:44:41     104    FALSE
#> 618     720575940626877799       108132 2020-04-29 01:57:21     104    FALSE
#> 619     720575940626877799       108133 2020-04-29 02:27:13     104    FALSE
#> 620     720575940626877799       108134 2020-04-29 02:36:06     104    FALSE
#> 621     720575940626877799       108135 2020-04-29 03:23:50     104    FALSE
#> 622     720575940626877799       108136 2020-04-29 03:28:56     104    FALSE
#> 623     720575940626877799       108137 2020-04-29 03:32:06     104    FALSE
#> 624     720575940626877799       108138 2020-04-29 03:32:24     104     TRUE
#> 625     720575940626877799       108139 2020-04-29 03:33:11     104    FALSE
#> 626     720575940626877799       108140 2020-04-29 03:33:23     104    FALSE
#> 627     720575940626877799       108259 2020-04-29 17:12:45     104    FALSE
#> 628     720575940626877799       108260 2020-04-29 17:12:49     104    FALSE
#> 631     720575940626877799       108263 2020-04-29 17:13:05     104    FALSE
#> 632     720575940626877799       108264 2020-04-29 17:13:12     104    FALSE
#> 633     720575940626877799       108266 2020-04-29 17:13:32     104    FALSE
#> 636     720575940626877799       108270 2020-04-29 17:14:03     104     TRUE
#> 639     720575940626877799       108273 2020-04-29 17:14:23     104     TRUE
#> 640     720575940626877799       108274 2020-04-29 17:14:25     104     TRUE
#> 641     720575940626877799       108275 2020-04-29 17:14:35     104     TRUE
#> 642     720575940626877799       108276 2020-04-29 17:14:44     104    FALSE
#> 643     720575940626877799       108277 2020-04-29 17:15:58     104     TRUE
#> 644     720575940626877799       108278 2020-04-29 17:16:15     104     TRUE
#> 645     720575940626877799       108279 2020-04-29 17:16:20     104     TRUE
#> 646     720575940626877799       108280 2020-04-29 17:16:33     104     TRUE
#> 647     720575940626877799       108281 2020-04-29 17:16:39     104     TRUE
#> 648     720575940626877799       108282 2020-04-29 17:16:43     104     TRUE
#> 649     720575940626877799       108283 2020-04-29 17:16:59     104     TRUE
#> 650     720575940626877799       108284 2020-04-29 17:17:25     104     TRUE
#> 651     720575940626877799       108285 2020-04-29 17:17:36     104     TRUE
#> 652     720575940626877799       108286 2020-04-29 17:17:44     104     TRUE
#> 653     720575940626877799       108287 2020-04-29 17:18:38     104     TRUE
#> 654     720575940626877799       108288 2020-04-29 17:18:50     104     TRUE
#> 655     720575940626877799       108289 2020-04-29 17:19:04     104     TRUE
#> 656     720575940626877799       108290 2020-04-29 17:19:16     104     TRUE
#> 657     720575940626877799       108291 2020-04-29 17:19:26     104     TRUE
#> 658     720575940626877799       108292 2020-04-29 17:21:44     104     TRUE
#> 659     720575940626877799       108293 2020-04-29 17:21:48     104     TRUE
#> 660     720575940626877799       108294 2020-04-29 17:22:01     104     TRUE
#> 661     720575940626877799       108295 2020-04-29 17:22:04     104     TRUE
#> 662     720575940626877799       108296 2020-04-29 17:22:40     104     TRUE
#> 664     720575940626877799       108298 2020-04-29 17:23:02     104     TRUE
#> 665     720575940626877799       108299 2020-04-29 17:23:05     104     TRUE
#> 667     720575940626877799       108301 2020-04-29 17:23:23     104     TRUE
#> 668     720575940626877799       108302 2020-04-29 17:23:38     104     TRUE
#> 669     720575940626877799       108303 2020-04-29 17:23:50     104     TRUE
#> 670     720575940626877799       108304 2020-04-29 17:23:53     104     TRUE
#> 671     720575940626877799       108305 2020-04-29 17:24:02     104     TRUE
#> 673     720575940626877799       108307 2020-04-29 17:24:23     104     TRUE
#> 674     720575940626877799       108308 2020-04-29 17:24:34     104     TRUE
#> 676     720575940626877799       108310 2020-04-29 17:24:47     104     TRUE
#> 678     720575940626877799       108312 2020-04-29 17:25:02     104    FALSE
#> 679     720575940626877799       108313 2020-04-29 17:25:29     104     TRUE
#> 680     720575940626877799       108314 2020-04-29 17:25:47     104     TRUE
#> 681     720575940626877799       108315 2020-04-29 17:25:54     104     TRUE
#> 682     720575940626877799       108316 2020-04-29 17:26:06     104     TRUE
#> 683     720575940626877799       108317 2020-04-29 17:26:21     104     TRUE
#> 684     720575940626877799       108318 2020-04-29 17:33:43     104     TRUE
#> 685     720575940626877799       108319 2020-04-29 17:33:51     104     TRUE
#> 686     720575940626877799       108320 2020-04-29 17:33:59     104     TRUE
#> 687     720575940626877799       108321 2020-04-29 17:34:05     104     TRUE
#> 688     720575940626877799       108322 2020-04-29 17:34:12     104     TRUE
#> 689     720575940626877799       108323 2020-04-29 17:34:21     104     TRUE
#> 690     720575940626877799       108324 2020-04-29 17:34:26     104     TRUE
#> 691     720575940626877799       108325 2020-04-29 17:34:33     104     TRUE
#> 692     720575940626877799       108326 2020-04-29 17:34:38     104     TRUE
#> 693     720575940626877799       108327 2020-04-29 17:34:44     104     TRUE
#> 694     720575940626877799       108328 2020-04-29 17:34:48     104     TRUE
#> 695     720575940626877799       108329 2020-04-29 17:34:58     104     TRUE
#> 697     720575940626877799       108331 2020-04-29 17:35:13     104     TRUE
#> 698     720575940626877799       108332 2020-04-29 17:35:19     104     TRUE
#> 699     720575940626877799       108333 2020-04-29 17:35:25     104     TRUE
#> 700     720575940626877799       108334 2020-04-29 17:35:32     104     TRUE
#> 701     720575940626877799       108335 2020-04-29 17:35:41     104     TRUE
#> 702     720575940626877799       108336 2020-04-29 17:35:46     104     TRUE
#> 703     720575940626877799       108337 2020-04-29 17:36:01     104     TRUE
#> 704     720575940626877799       108338 2020-04-29 17:37:18     104     TRUE
#> 705     720575940626877799       108479 2020-04-29 20:48:54     104    FALSE
#> 709     720575940626877799       111739 2020-05-06 16:16:58      84    FALSE
#> 710     720575940626877799       111740 2020-05-06 16:17:10      84    FALSE
#> 712     720575940626877799       111766 2020-05-06 16:48:07      84    FALSE
#> 713     720575940626877799       111819 2020-05-06 17:05:02      84    FALSE
#> 714     720575940626877799       111824 2020-05-06 17:13:58      84    FALSE
#> 715     720575940626877799       111836 2020-05-06 17:22:18      84    FALSE
#> 716     720575940626877799       111837 2020-05-06 17:22:28      84    FALSE
#> 721     720575940626877799       111842 2020-05-06 17:26:58      84    FALSE
#> 724     720575940626877799       111845 2020-05-06 17:29:53      84    FALSE
#> 725     720575940626877799       111849 2020-05-06 17:32:11      84    FALSE
#> 726     720575940626877799       112050 2020-05-06 18:47:44      84    FALSE
#> 729     720575940626877799       112095 2020-05-06 19:17:35      84    FALSE
#> 732     720575940626877799       112118 2020-05-06 19:50:26      84    FALSE
#> 733     720575940626877799       112136 2020-05-06 20:02:49      84    FALSE
#> 734     720575940626877799       112137 2020-05-06 20:03:05      84    FALSE
#> 735     720575940626877799       112140 2020-05-06 20:07:48      84    FALSE
#> 736     720575940626877799       112145 2020-05-06 20:17:07      84    FALSE
#> 737     720575940626877799       112176 2020-05-06 20:42:38      84    FALSE
#> 738     720575940626877799       112207 2020-05-06 21:10:54      84    FALSE
#> 739     720575940626877799       112208 2020-05-06 21:14:47      84    FALSE
#> 740     720575940626877799       112210 2020-05-06 21:15:08      84    FALSE
#> 741     720575940626877799       112211 2020-05-06 21:17:43      84    FALSE
#> 742     720575940626877799       112225 2020-05-06 21:41:35      84    FALSE
#> 743     720575940626877799       112226 2020-05-06 21:41:48      84    FALSE
#> 744     720575940626877799       112667 2020-05-07 14:11:01      84    FALSE
#> 748     720575940626877799       112682 2020-05-07 14:37:31      84     TRUE
#> 749     720575940626877799       112683 2020-05-07 14:37:34      84     TRUE
#> 750     720575940626877799       112685 2020-05-07 14:37:51      84     TRUE
#> 751     720575940626877799       112686 2020-05-07 14:38:30      84     TRUE
#> 752     720575940626877799       112687 2020-05-07 14:38:34      84     TRUE
#> 753     720575940626877799       112688 2020-05-07 14:38:48      84     TRUE
#> 754     720575940626877799       112689 2020-05-07 14:39:03      84     TRUE
#> 755     720575940626877799       112690 2020-05-07 14:39:05      84     TRUE
#> 756     720575940626877799       112691 2020-05-07 14:39:15      84     TRUE
#> 757     720575940626877799       112692 2020-05-07 14:39:29      84     TRUE
#> 760     720575940626877799       112695 2020-05-07 14:40:02      84     TRUE
#> 761     720575940626877799       112696 2020-05-07 14:40:16      84     TRUE
#>               user_name
#> 10      Claire McKellar
#> 11      Claire McKellar
#> 12      Claire McKellar
#> 31      Claire McKellar
#> 32      Claire McKellar
#> 33      Claire McKellar
#> 34      Claire McKellar
#> 35      Claire McKellar
#> 38      Claire McKellar
#> 40      Claire McKellar
#> 44...11 Claire McKellar
#> 46      Claire McKellar
#> 48      Claire McKellar
#> 49      Claire McKellar
#> 51      Claire McKellar
#> 53      Claire McKellar
#> 55      Claire McKellar
#> 57      Claire McKellar
#> 59      Claire McKellar
#> 62      Claire McKellar
#> 64      Claire McKellar
#> 66      Claire McKellar
#> 67      Claire McKellar
#> 72...24 Claire McKellar
#> 74...25 Claire McKellar
#> 76...26 Claire McKellar
#> 77...27 Claire McKellar
#> 78      Claire McKellar
#> 79      Claire McKellar
#> 80...30 Claire McKellar
#> 81...31 Claire McKellar
#> 82...32 Claire McKellar
#> 94      Claire McKellar
#> 120     Claire McKellar
#> 151     Claire McKellar
#> 159     Claire McKellar
#> 164     Claire McKellar
#> 165     Claire McKellar
#> 166     Claire McKellar
#> 185     Claire McKellar
#> 186     Claire McKellar
#> 187     Claire McKellar
#> 188     Claire McKellar
#> 189     Claire McKellar
#> 204     Claire McKellar
#> 205     Claire McKellar
#> 207     Claire McKellar
#> 208           Jay Gager
#> 209           Jay Gager
#> 211       Sandeep Kumar
#> 213           Jay Gager
#> 214           Jay Gager
#> 216      Sarah Morejohn
#> 217      Sarah Morejohn
#> 220      Sarah Morejohn
#> 221      Sarah Morejohn
#> 222      Sarah Morejohn
#> 223      Sarah Morejohn
#> 1         Christa Baker
#> 44...60   Christa Baker
#> 68            Jay Gager
#> 69            Jay Gager
#> 70            Jay Gager
#> 71            Jay Gager
#> 72...65       Jay Gager
#> 73            Jay Gager
#> 74...67       Jay Gager
#> 75            Jay Gager
#> 76...69       Jay Gager
#> 77...70       Jay Gager
#> 80...71       Jay Gager
#> 81...72       Jay Gager
#> 82...73       Jay Gager
#> 85            Jay Gager
#> 88            Jay Gager
#> 89            Jay Gager
#> 90            Jay Gager
#> 91            Jay Gager
#> 92            Jay Gager
#> 100      Austin T Burke
#> 101      Austin T Burke
#> 300        Merlin Moore
#> 369        Merlin Moore
#> 370        Merlin Moore
#> 371        Merlin Moore
#> 372        Merlin Moore
#> 507         Ryan Willie
#> 508         Ryan Willie
#> 509         Ryan Willie
#> 510         Ryan Willie
#> 513         Ryan Willie
#> 514         Ryan Willie
#> 519         Ryan Willie
#> 520         Ryan Willie
#> 521         Ryan Willie
#> 522         Ryan Willie
#> 523         Ryan Willie
#> 524         Ryan Willie
#> 525         Ryan Willie
#> 526         Ryan Willie
#> 527         Ryan Willie
#> 528         Ryan Willie
#> 529         Ryan Willie
#> 530         Ryan Willie
#> 531         Ryan Willie
#> 532         Ryan Willie
#> 533         Ryan Willie
#> 534         Ryan Willie
#> 535         Ryan Willie
#> 536         Ryan Willie
#> 537         Ryan Willie
#> 538         Ryan Willie
#> 539         Ryan Willie
#> 540         Ryan Willie
#> 541         Ryan Willie
#> 542         Ryan Willie
#> 543         Ryan Willie
#> 544         Ryan Willie
#> 545         Ryan Willie
#> 547         Ryan Willie
#> 548         Ryan Willie
#> 549         Ryan Willie
#> 550         Ryan Willie
#> 551         Ryan Willie
#> 552         Ryan Willie
#> 554         Ryan Willie
#> 555         Ryan Willie
#> 556         Ryan Willie
#> 557         Ryan Willie
#> 558         Ryan Willie
#> 559         Ryan Willie
#> 560         Ryan Willie
#> 561         Ryan Willie
#> 562         Ryan Willie
#> 563         Ryan Willie
#> 565         Ryan Willie
#> 566         Ryan Willie
#> 567         Ryan Willie
#> 568         Ryan Willie
#> 569         Ryan Willie
#> 570         Ryan Willie
#> 571         Ryan Willie
#> 572         Ryan Willie
#> 573         Ryan Willie
#> 574         Ryan Willie
#> 578         Ryan Willie
#> 579         Ryan Willie
#> 580         Ryan Willie
#> 581         Ryan Willie
#> 582         Ryan Willie
#> 587         Ryan Willie
#> 588         Ryan Willie
#> 589         Ryan Willie
#> 590         Ryan Willie
#> 593         Ryan Willie
#> 594         Ryan Willie
#> 595         Ryan Willie
#> 596         Ryan Willie
#> 597         Ryan Willie
#> 598         Ryan Willie
#> 599         Ryan Willie
#> 600         Ryan Willie
#> 601         Ryan Willie
#> 602         Ryan Willie
#> 603         Ryan Willie
#> 604         Ryan Willie
#> 605         Ryan Willie
#> 606         Ryan Willie
#> 607         Ryan Willie
#> 608         Ryan Willie
#> 609         Ryan Willie
#> 614         Ryan Willie
#> 615         Ryan Willie
#> 616         Ryan Willie
#> 617         Ryan Willie
#> 618         Ryan Willie
#> 619         Ryan Willie
#> 620         Ryan Willie
#> 621         Ryan Willie
#> 622         Ryan Willie
#> 623         Ryan Willie
#> 624         Ryan Willie
#> 625         Ryan Willie
#> 626         Ryan Willie
#> 627         Ryan Willie
#> 628         Ryan Willie
#> 631         Ryan Willie
#> 632         Ryan Willie
#> 633         Ryan Willie
#> 636         Ryan Willie
#> 639         Ryan Willie
#> 640         Ryan Willie
#> 641         Ryan Willie
#> 642         Ryan Willie
#> 643         Ryan Willie
#> 644         Ryan Willie
#> 645         Ryan Willie
#> 646         Ryan Willie
#> 647         Ryan Willie
#> 648         Ryan Willie
#> 649         Ryan Willie
#> 650         Ryan Willie
#> 651         Ryan Willie
#> 652         Ryan Willie
#> 653         Ryan Willie
#> 654         Ryan Willie
#> 655         Ryan Willie
#> 656         Ryan Willie
#> 657         Ryan Willie
#> 658         Ryan Willie
#> 659         Ryan Willie
#> 660         Ryan Willie
#> 661         Ryan Willie
#> 662         Ryan Willie
#> 664         Ryan Willie
#> 665         Ryan Willie
#> 667         Ryan Willie
#> 668         Ryan Willie
#> 669         Ryan Willie
#> 670         Ryan Willie
#> 671         Ryan Willie
#> 673         Ryan Willie
#> 674         Ryan Willie
#> 676         Ryan Willie
#> 678         Ryan Willie
#> 679         Ryan Willie
#> 680         Ryan Willie
#> 681         Ryan Willie
#> 682         Ryan Willie
#> 683         Ryan Willie
#> 684         Ryan Willie
#> 685         Ryan Willie
#> 686         Ryan Willie
#> 687         Ryan Willie
#> 688         Ryan Willie
#> 689         Ryan Willie
#> 690         Ryan Willie
#> 691         Ryan Willie
#> 692         Ryan Willie
#> 693         Ryan Willie
#> 694         Ryan Willie
#> 695         Ryan Willie
#> 697         Ryan Willie
#> 698         Ryan Willie
#> 699         Ryan Willie
#> 700         Ryan Willie
#> 701         Ryan Willie
#> 702         Ryan Willie
#> 703         Ryan Willie
#> 704         Ryan Willie
#> 705         Ryan Willie
#> 709        Merlin Moore
#> 710        Merlin Moore
#> 712        Merlin Moore
#> 713        Merlin Moore
#> 714        Merlin Moore
#> 715        Merlin Moore
#> 716        Merlin Moore
#> 721        Merlin Moore
#> 724        Merlin Moore
#> 725        Merlin Moore
#> 726        Merlin Moore
#> 729        Merlin Moore
#> 732        Merlin Moore
#> 733        Merlin Moore
#> 734        Merlin Moore
#> 735        Merlin Moore
#> 736        Merlin Moore
#> 737        Merlin Moore
#> 738        Merlin Moore
#> 739        Merlin Moore
#> 740        Merlin Moore
#> 741        Merlin Moore
#> 742        Merlin Moore
#> 743        Merlin Moore
#> 744        Merlin Moore
#> 748        Merlin Moore
#> 749        Merlin Moore
#> 750        Merlin Moore
#> 751        Merlin Moore
#> 752        Merlin Moore
#> 753        Merlin Moore
#> 754        Merlin Moore
#> 755        Merlin Moore
#> 756        Merlin Moore
#> 757        Merlin Moore
#> 760        Merlin Moore
#> 761        Merlin Moore

# }