Access precomputed flywire connectivity data

flywire_connectome_data_version returns the integer version number of the most recent available flywire connectome dump on your machine.

flywire_connectome_data(
  type = c("syn", "pre", "post"),
  version = NULL,
  cached = TRUE,
  ...
)

flywire_connectome_data_version()

Arguments

type

Character vector specifying the kind of data

version

Optional CAVE version. The default value of NULL uses the latest data dump available.

cached

When version is NULL whether to use a cached value (lasting 1 hour) of the latest available version.

...

Additional arguments passed to arrow::open_dataset.

Value

An arrow object that you can use with dplyr verbs like

filter in order to find neurons/connectivity data of interest.

Details

This depends on precomputed data dumps prepared periodically by Sven Dorkenwald. You must either download these to an appropriate location on your hard drive or add a link to your own Google drive. I have noticed unfortunately that some versions of arrow do not play nicely with the Google drive app on macosx, so I would generally recommend downloading a specific version to a location on your machine. The link to Sven's Google drive folder can be found in this Slack message.

Examples

# \donttest{
# latest available version/
syn=try(flywire_connectome_data('syn'), silent=TRUE)
syn450=try(flywire_connectome_data('syn', version=450), silent=TRUE)
if(!inherits(syn450, 'try-error')) {
syn450
syn450$metadata

dl4ds <- syn450 %>%
  filter(pre_pt_root_id==flywire_ids("DL4_adPN_R", version=450, integer64 = TRUE)) %>%
  collect()
}

# }
if (FALSE) {
  flywire_connectome_data_version()
}