Get predicted dense core vesicle locations in the flywire dataset

Preliminary dense core vesicle (DCV) detection results from the Lee group, Stephan Gerhard and Minsu Kim.

flywire_dcvs(
  rootid,
  url = "https://radagast.hms.harvard.edu/fafb_test/dcv/for_segment",
  simplify = TRUE,
  return = c("parsed", "text", "response"),
  token = NULL,
  simplifyVector = TRUE,
  include_headers = FALSE,
  OmitFailures = TRUE,
  ...
)

Arguments

rootid
flywire rootid/rootids

url
Full URL for brainmaps API endpoint

simplify
logical, if FALSE each rootid is a separate set of two data frames (one for DCV positions, one for the neuron's synapses). Else, a list of two combined data frames is returned.

return
One of "parsed", "text" (for raw JSON), or "response"

token
Optional chunkedgraph token (otherwise the default one for the current segmentation will be used).

simplifyVector
Whether to use jsonlite::simplifyVector

include_headers
Whether to include basic headers from the http request as attributes on the parsed JSON object (default TRUE) when parse.json=TRUE.

OmitFailures
logical, if TRUE then requests that result in a 500 error are dropped and a warning is displayed but not an error.

...
additional arguments passed to the httr::{RETRY} function. This may include a config list other named parameters etc.

Value

A list, where the first entry contains DCV locations and the second the synapses for the given rootid, with the nearest DCV precalculated for each synapse..

Examples

# \donttest{
data = flywire_dcvs("720575940629166904")

#> Request failed [400]. Retrying in 1.9 seconds...

#> Request failed [400]. Retrying in 1 seconds...

#> Request failed [400]. Retrying in 1.7 seconds...

#> Request failed [400]. Retrying in 3.3 seconds...

#> Request failed [400]. Retrying in 6 seconds...

#> Request failed [400]. Retrying in 11.9 seconds...

#> Request failed [400]. Retrying in 8.5 seconds...

#> Request failed [400]. Retrying in 41.7 seconds...

#> Request failed [400]. Retrying in 19.4 seconds...

#> Bad Request (HTTP 400).

#> Error in flywire_errorhandle(req): HTTPSConnectionPool(host='global.daf-apis.com', port=443): Max retries exceeded with url: /info/api/v2/datastack/full/flywire_fafb_production (Caused by NewConnectionError('<urllib3.connection.HTTPSConnection object at 0x7f0f570746d0>: Failed to establish a new connection: [Errno 16] Device or resource busy'))

dcv = data$dcv

#> Error in data$dcv: object of type 'closure' is not subsettable

synapse = data$syns

#> Error in data$syns: object of type 'closure' is not subsettable

# }