GET/POST from flywire servers with appropriate authorisation token

GET/POST from flywire servers with appropriate authorisation token

flywire_fetch(
  url,
  body = NULL,
  config = NULL,
  parse.json = TRUE,
  cache = FALSE,
  retry = 0L,
  include_headers = FALSE,
  simplifyVector = TRUE,
  ...
)

Arguments

url
Full URL for brainmaps API endpoint

body
an R list with parameters that will be converted with jsonlite::toJSON and then passed on to POST. You can also pass a JSON character vector to have more control of the JSON encoding.

parse.json
Whether or not to parse a JSON response to an R object (default TRUE)

cache
Whether or not to cache responses (default FALSE)

retry
The number of times to retry the operation (default 0, FALSE=>0 and TRUE=>3). See the documentation of the times argument of httr::RETRY for further details.

include_headers
Whether to include basic headers from the http request as attributes on the parsed JSON object (default TRUE) when parse.json=TRUE.

simplifyVector
Whether to use jsonlite::simplifyVector

...
additional arguments passed to the httr::{RETRY} function. This may include a config list other named parameters etc.

Value

Examples

# \donttest{
flywire_fetch("https://globalv1.flywire-daf.com/nglstate/5747205470158848")

#> $layers
#>                                                                        source
#> 1 precomputed://gs://microns-seunglab/drosophila_v0/alignment/image_rechunked
#> 2          graphene://https://prodv1.flywire-daf.com/segmentation/1.0/fly_v31
#>                      type   blend                               name   tab
#> 1                   image default                   Production-image  <NA>
#> 2 segmentation_with_graph    <NA> Production-segmentation_with_graph graph
#>   selectedAlpha  colorSeed
#> 1            NA         NA
#> 2          0.62 1358768360
#>                                                     segments
#> 1                                                       NULL
#> 2 720575940619527173, 720575940628130268, 720575940630484179
#>                           hiddenSegments skeletonRendering.mode2d
#> 1                                   NULL                     <NA>
#> 2 720575940634222734, 720575940636364870         lines_and_points
#>   skeletonRendering.mode3d graphOperationMarker pathFinder.color
#> 1                     <NA>                 NULL             <NA>
#> 2                    lines                 NULL          #ffff00
#>   pathFinder.pathObject.annotationPath.annotations
#> 1                                             NULL
#> 2                                             NULL
#>   pathFinder.pathObject.annotationPath.tags pathFinder.pathObject.hasPath
#> 1                                      NULL                            NA
#> 2                                      NULL                         FALSE
#> 
#> $navigation
#> $navigation$pose
#> $navigation$pose$position
#> $navigation$pose$position$voxelSize
#> [1]  4  4 40
#> 
#> $navigation$pose$position$voxelCoordinates
#> [1] 115198  55453   1537
#> 
#> 
#> 
#> $navigation$zoomFactor
#> [1] 4.145162
#> 
#> 
#> $perspectiveOrientation
#> [1] -0.22566605 -0.03401040 -0.07959923  0.97035152
#> 
#> $perspectiveZoom
#> [1] 4888.124
#> 
#> $showSlices
#> [1] FALSE
#> 
#> $jsonStateServer
#> [1] "https://globalv1.flywire-daf.com/nglstate/post"
#> 
#> $selectedLayer
#> $selectedLayer$layer
#> [1] "Production-segmentation_with_graph"
#> 
#> $selectedLayer$visible
#> [1] TRUE
#> 
#> 
#> $layout
#> [1] "xy-3d"
#> 

# }