Flexible specification of flywire ids (including from flytable types)

allows more flexible specification of flywire root ids compared with ngl_segments including by queries against cell types recorded in flytable.

flywire_ids(
  x,
  integer64 = FALSE,
  check_latest = FALSE,
  must_work = FALSE,
  na_ok = FALSE,
  unique = FALSE,
  version = NULL,
  ...
)

Arguments

x

A character or bit64::integer64 vector or a dataframe specifying ids directly or a string specifying a query (see examples) or a URL.

integer64

Whether to return ids as 64 bit integers - more compact than character vector, but can be more fragile (default FALSE).

check_latest

Whether to check if ids are up to date.

must_work

Whether ids must be valid

na_ok

whether NA ids are acceptable when must_work=TRUE

unique

Whether to return only unique ids

version

Integer materialisation version

...

Additional arguments passed to flytable_cell_types or ngl_segments.

Value

character (or integer64)) vector of segment ids

See also

Examples

flywire_ids(data.frame(root_id=1))
#> [1] "1"
flywire_ids(data.frame(root_id=1), integer64=TRUE)
#> integer64
#> [1] 1
# BA
flywire_ids(data.frame(root_id=-1))
#> [1] "0"
if (FALSE) {
# will error
flywire_ids(data.frame(root_id=-1), must_work = TRUE)
}
# DL1 olfactory PNs
flywire_ids("DL1_adPN")
#> [1] "720575940622368792" "720575940627042064" "720575940616079035"
#> [4] "720575940632167085"
# DL1 olfactory PNs but only on the RHS
flywire_ids("DL1_adPN_R")
#> [1] "720575940622368792" "720575940616079035"
# specifying materialisation version
flywire_ids("DL1_adPN_R", version=400)
#> Updating 3 ids
#> flywire_rootid_cached: Looking up 3 missing keys
#> [1] "720575940618302936" "720575940618757681"
# using SQL wild cards
flywire_ids("DA[12]_%PN_L")
#>  [1] "720575940604407468" "720575940623543881" "720575940639337461"
#>  [4] "720575940612355507" "720575940637469254" "720575940617229632"
#>  [7] "720575940628259407" "720575940624106442" "720575940622762995"
#> [10] "720575940622311704" "720575940610052266" "720575940621239679"
#> [13] "720575940623303108" "720575940630066007"

# note that side is defined by soma position (not arbour side)
flywire_ids("class:MBON_R", integer64=TRUE)
#> integer64
#>  [1] 720575940631318031 720575940638163428 720575940642142861 720575940617552340
#>  [5] 720575940617567206 720575940614595218 720575940627770628 720575940637902938
#>  [9] 720575940638774606 720575940626833021 720575940618008859 720575940635841143
#> [13] 720575940628779563 720575940632656173 720575940628581379 720575940617413957
#> [17] 720575940617573922 720575940613157329 720575940645304430 720575940623743415
#> [21] 720575940622855629 720575940630767959 720575940635082347 720575940628757547
#> [25] 720575940622473548 720575940626046154 720575940625596720 720575940644615716
#> [29] 720575940623377802 720575940622979277 720575940626872094 720575940643863496
#> [33] 720575940630075703 720575940652390134 720575940632535756 720575940643696288
#> [37] 720575940623201833 720575940616463477 720575940629981440 720575940623182847
#> [41] 720575940624539284 720575940620464321 720575940607183026 720575940614392999
#> [45] 720575940617302365 720575940624694503 720575940621164720 720575940623841463
#> [49] 720575940631177803 720575940609440324 720575940623199684