Flexible specification of flywire ids (including from flytable types)

allows more flexible specification of flywire root ids compared with ngl_segments including by queries against cell types recorded in flytable.

flywire_ids(
  x,
  integer64 = FALSE,
  check_latest = FALSE,
  must_work = FALSE,
  na_ok = FALSE,
  unique = FALSE,
  ...
)

Arguments

x

A character or bit64::integer64 vector or a dataframe specifying ids directly or a string specifying a query (see examples) or a URL.

integer64

Whether to return ids as 64 bit integers - more compact than character vector, but can be more fragile (default FALSE).

check_latest

Whether to check if ids are up to date.

must_work

Whether ids must be valid

na_ok

whether NA ids are acceptable when must_work=TRUE

unique

Whether to return only unique ids

...

Additional arguments passed to flytable_cell_types or ngl_segments.

Value

character (or integer64)) vector of segment ids

See also

Examples

flywire_ids(data.frame(root_id=1))
#> [1] "1"
flywire_ids(data.frame(root_id=1), integer64=TRUE)
#> integer64
#> [1] 1
# BA
flywire_ids(data.frame(root_id=-1))
#> [1] "0"
if (FALSE) {
# will error
flywire_ids(data.frame(root_id=-1), must_work = TRUE)
}
# DL1 olfactory PNs
flywire_ids("DL1_adPN")
#> [1] "720575940622368792" "720575940627042064" "720575940629656535"
#> [4] "720575940632167085"
# DL1 olfactory PNs but only on the RHS
flywire_ids("DL1_adPN_R")
#> [1] "720575940622368792" "720575940629656535"
# specifying materialisation version
flywire_ids("DL1_adPN_R", version=400)
#> Updating 3 ids
#> [1] "720575940618302936" "720575940618757681"
# using SQL wild cards
flywire_ids("DA[12]_%PN_L")
#>  [1] "720575940604407468" "720575940623543881" "720575940639337461"
#>  [4] "720575940640817627" "720575940637469254" "720575940617229632"
#>  [7] "720575940628259407" "720575940624106442" "720575940622762995"
#> [10] "720575940622311704" "720575940610052266" "720575940621239679"
#> [13] "720575940634379871" "720575940630066007"

# note that side is defined by soma position (not arbour side)
flywire_ids("class:MBON_R", integer64=TRUE)
#> integer64
#>  [1] 720575940630697303 720575940625991268 720575940619259352 720575940612772579
#>  [5] 720575940628764495 720575940616298635 720575940608410889 720575940624753336
#>  [9] 720575940621032052 720575940638774606 720575940625168528 720575940639693144
#> [13] 720575940620362945 720575940632656173 720575940628581379 720575940617413957
#> [17] 720575940617573922 720575940613157329 720575940627091210 720575940623743415
#> [21] 720575940638910141 720575940625850826 720575940630373113 720575940622281430
#> [25] 720575940637831487 720575940613666001 720575940617869083 720575940636311694
#> [29] 720575940630387820 720575940627584516 720575940622268374 720575940615186079
#> [33] 720575940612636694 720575940632028753 720575940617125081 720575940612637974
#> [37] 720575940654692001 720575940625994596 720575940628587779 720575940619271128
#> [41] 720575940619720277 720575940631031055 720575940624744462 720575940631940236
#> [45] 720575940623204548 720575940630390265 720575940630691415 720575940614720662
#> [49] 720575940623088937 720575940632393164 720575940621137334 720575940623199684