Find the input/output partners for a flywire neuron

flywire_partners is a low level function returning one row for every synapse.

flywire_partner_summary summarises the connectivity of one or more flywire neurons.

flywire_partners(
  rootid,
  partners = c("outputs", "inputs"),
  details = FALSE,
  roots = TRUE,
  cloudvolume.url = NULL,
  method = c("auto", "spine", "sqlite"),
  Verbose = TRUE,
  ...
)

flywire_partner_summary(
  rootid,
  partners = c("outputs", "inputs"),
  threshold = 0,
  remove_autapses = TRUE,
  Verbose = NA,
  ...
)

Arguments

rootid
flywire rootids identifying flywire neurons as a character vector or any format understandable by ngl_segments including neuroglancer scenes (ngscene).

partners
Whether to fetch input or output synapses

details
Whether to include additional details such as X Y Z location (default FALSE)

roots
Whether to fetch the flywire rootids of the partner neurons (default TRUE)

cloudvolume.url
The segmentation source URL for cloudvolume. Normally you can ignore this and rely on the default segmentation chosen by choose_segmentation

method
Whether to use a local SQLite database or remote spine service for synapse data. The default auto uses a local database when available (45GB but faster).

Verbose
Whether to print status messages

...
Additional arguments passed to pbsapply

threshold
For flywire_partner_summary only return partners with greater than this number of connections to the query neuron(s) (default of 0 returns all connections)

remove_autapses
For flywire_partner_summary whether to remove autapses (defaults to TRUE)

Details

FIXME behaviour when there are no partners is not well-defined.

Note that there are many duplicated connections in this raw output, false autapses (i.e. the same neuron connected to itself) and other false positives. See Buhmann et al for details and ideas about cleaning up the results.

See also

Examples

# \donttest{
pp=flywire_partners("720575940621039145")

#> Fetching supervoxel ids for id: 720575940621039145

#> Finding synapses for supervoxels

#> Fetching root ids

head(pp)

#>    offset          pre_svid         post_svid            post_id
#> 1  152196 81491472725509269 81491472725509199 720575940621039145
#> 2  828599 81561978908660664 81561978908653994 720575940628471127
#> 3  828625 81561978908669037 81561978908667521 720575940589754844
#> 4 1236913 81631042385463742 81631042385462154 720575940621039145
#> 5 1236914 81631042385459399 81631042385454876 720575940621039145
#> 6 1236915 81631042385462087 81631042385470852 720575940621039145
#>               pre_id
#> 1 720575940621039145
#> 2 720575940621039145
#> 3 720575940621039145
#> 4 720575940621039145
#> 5 720575940621039145
#> 6 720575940621039145

# }
# \donttest{
flywire_partner_summary("720575940621039145", partners='out')

#> Fetching supervoxel ids for id: 720575940621039145

#> Finding synapses for supervoxels

#> Fetching root ids

#> # A tibble: 1,262 x 3
#>    post_id            weight     n
#>    <chr>               <int> <int>
#>  1 720575940618522433    172     1
#>  2 720575940623301520     62     1
#>  3 720575940637390424     62     1
#>  4 720575940622541972     34     1
#>  5 720575940620466824     30     1
#>  6 720575940621370346     29     1
#>  7 720575940610168805     27     1
#>  8 720575940624743946     27     1
#>  9 720575940635067251     25     1
#> 10 720575940613535430     24     1
#> # … with 1,252 more rows

flywire_partner_summary("720575940621039145", partners='in')

#> Fetching supervoxel ids for id: 720575940621039145

#> Finding synapses for supervoxels

#> Fetching root ids

#> # A tibble: 684 x 3
#>    pre_id             weight     n
#>    <chr>               <int> <int>
#>  1 720575940613380914     90     1
#>  2 720575940623926970     88     1
#>  3 720575940621370346     77     1
#>  4 720575940602114016     74     1
#>  5 720575940637390424     69     1
#>  6 720575940610188697     56     1
#>  7 720575940619059120     56     1
#>  8 720575940623931710     53     1
#>  9 720575940613535430     51     1
#> 10 720575940621317610     51     1
#> # … with 674 more rows

flywire_partner_summary("720575940621039145")

#> Fetching supervoxel ids for id: 720575940621039145

#> Finding synapses for supervoxels

#> Fetching root ids

#> # A tibble: 1,262 x 3
#>    post_id            weight     n
#>    <chr>               <int> <int>
#>  1 720575940618522433    172     1
#>  2 720575940623301520     62     1
#>  3 720575940637390424     62     1
#>  4 720575940622541972     34     1
#>  5 720575940620466824     30     1
#>  6 720575940621370346     29     1
#>  7 720575940610168805     27     1
#>  8 720575940624743946     27     1
#>  9 720575940635067251     25     1
#> 10 720575940613535430     24     1
#> # … with 1,252 more rows

# summary for neuron at a XYZ location (in this case in raw coordinates)
flywire_partner_summary(flywire_xyz2id(cbind(155682, 58180, 3215),
  rawcoords = TRUE))

#> Fetching supervoxel ids for id: 720575940621039145

#> Finding synapses for supervoxels

#> Fetching root ids

#> # A tibble: 1,262 x 3
#>    post_id            weight     n
#>    <chr>               <int> <int>
#>  1 720575940618522433    172     1
#>  2 720575940623301520     62     1
#>  3 720575940637390424     62     1
#>  4 720575940622541972     34     1
#>  5 720575940620466824     30     1
#>  6 720575940621370346     29     1
#>  7 720575940610168805     27     1
#>  8 720575940624743946     27     1
#>  9 720575940635067251     25     1
#> 10 720575940613535430     24     1
#> # … with 1,252 more rows

if (FALSE) {
# Use Ctrl+Shift+J to share a flywire scene and then do this to get partner
# summary for that URL
flywire_partner_summary(clipr::read_clip())

}
# }