Find the root_id of a FlyWire segment / supervoxel.

Find the root_id of a FlyWire segment / supervoxel.

flywire_rootid(
  x,
  method = c("auto", "cloudvolume", "flywire"),
  cloudvolume.url = NULL,
  ...
)

Arguments

x
One or more FlyWire segment ids

method
Whether to use the flywire API (slow but no python required) OR cloudvolume (faster for many input ids, but requires python). "auto" (the default) will choose "flywire" for length 1 queries, "cloudvolume" otherwise.

cloudvolume.url
URL for CloudVolume to fetch segmentation image data. The default value of NULL chooses the flywire production segmentation dataset.

...
Additional arguments passed to pbsapply and eventually flywire_fetch when method="flywire" OR to cv$CloudVolume when method="cloudvolume"

Value

A vector of root ids as character vectors.

Details

The main purpose of this function is to convert a supervoxel into the current root id for the whole segment. If a segment id has been updated due to editing, calling with the original segment id will still return the same segment id (although calling it with a supervoxel would return the new segment id). If you wish to find the latest root id, then you can use flywire_latestid, but this is fairly slow (think 1 second per neuron). However in general it is best to select an XYZ location defining a neuron of interest, or the associated supervoxel id and store that, since these can be very rapidly looked up by flywire_xyz2id and flywire_rootid.

There are two methods. flywire is simpler but will be slower for many supervoxels since each id requires a separate http request.

See also

Examples

# \donttest{
flywire_rootid(c("81489548781649724", "80011805220634701"))

#> [1] "720575940625861628" "720575940615895581"

# same but using the flywire sandbox segmentation
with_segmentation('sandbox', {
flywire_rootid(c("81489548781649724", "80011805220634701"))
})

#> [1] "720575940626292279" "720575940622027055"

# }