Title

Title

flywire_xyz2id(
  xyz,
  rawcoords = FALSE,
  voxdims = c(4, 4, 40),
  cloudvolume.url = NULL,
  root = TRUE,
  ...
)

Arguments

xyz
One or more xyz locations as an Nx3 matrix or in any form compatible with xyzmatrix including neuron or mesh3d surface objects.

rawcoords
whether the input values are raw voxel indices or in nm

voxdims
voxel dimensions in nm used to convert the

cloudvolume.url
URL for CloudVolume to fetch segmentation image data. The default value of NULL choose the production segmentation dataset.

root
Whether to return the root id of the whole segment rather than the supervoxel id.

...
additional arguments passed to pbapply when looking up multiple positions.

Value

A character vector of segment ids, NA when lookup fails.

Details

Note that finding the supervoxel for a given XYZ location is order 3x faster than finding the root id for the agglomeration of all of the super voxels in a given object. Perhaps less intuitively, if you want to look up many root ids, it is actually quicker to do flywire_xyz2id(,root=F) followed by flywire_rootid since that function can look up many root ids in a single call to the ChunkedGraph server.

Examples

# \donttest{
if (FALSE) {
# example based on defining some sample points from real neurons
# sample dataset from FAFB catmaid
n=elmr::dense_core_neurons[[1]]
set.seed(42)
ss=sample(nvertices(n), size=10)
# first convert FAF14 points to FlyWire coordinate space
nx=xform_brain(elmr::dense_core_neurons[[1]], ref="FlyWire", sample="FAFB14")
pts=xyzmatrix(nx)[ss,]
}

# for simplicity just define those same sample points directly
pts = matrix(
c(428910.52, 110629.64, 174800, 410201.86, 110419.1, 180400,
  337136, 129926.28, 189280, 349981.85, 136041.53, 199280, 398361.74,
  110731.26, 182640, 382789.26, 118987.04, 189920, 358033.92, 171592.92,
  143960, 360990.48, 140277.42, 201400, 376258.06, 125201.51, 194400,
  331370.31, 128338.58, 181760),
ncol = 3, byrow = TRUE,
dimnames = list(NULL, c("X", "Y", "Z"))
)

# now find the ids for the selected xyz locations
flywire_xyz2id(pts)

#>  [1] "720575940637192819" "720575940637192819" "720575940637192819"
#>  [4] "720575940637192819" "720575940637192819" "720575940637192819"
#>  [7] "720575940637192819" "720575940637192819" "720575940637192819"
#> [10] "720575940637192819"

# we can also find the supervoxels - much faster
svids=flywire_xyz2id(pts, root=FALSE)

# now look up the root ids - very fast with method="cloudvolume", the default
flywire_rootid(svids)

#>    77970012195136907    77688537218572896    76422243487490672 
#> "720575940637192819" "720575940637192819" "720575940637192819" 
#>    76633418439845284    77477430986135693    77196093515061588 
#> "720575940637192819" "720575940637192819" "720575940637192819" 
#>    76774774202351196    76844593391918899    77055424746328456 
#> "720575940637192819" "720575940637192819" "720575940637192819" 
#>    76281437212920281 
#> "720575940637192819" 

# }