Graph representation of the agglomeration (merge) between raw segments

make_merge_graph generates a graph representation of all agglomeration merges.

merge_graph_components uses the full merge graph to find each connected subgraph corresponding to isolated merge groups.

make_merge_graph(x, n = NA)

merge_graph_components(g, n = NA)

Arguments

x
The mergeinfo data.frame generated by read_mergeinfo

n
For testing purposes, the number of rows of merge information to process - the default processes all rows.

g
Either a graph generated by make_merge_graph or the mergeinfo data.frame generated by read_mergeinfo.

Value

make_merge_graph returns an igraph object, merge_graph_components returns a list with elements describing the mapping between segment ids and merge groups.

Examples

# NOT RUN {
g=make_merge_graph(mergeinfo)
gc=merge_graph_components(g)
str(gc)
# }