Construct one or more neuroglancer annotation layers

Construct one or more neuroglancer annotation layers

ngl_annotation_layers(ann, rawcoords = NA, colpal = NULL)

Arguments

ann

An annotation dataframe (see details) or any object containing 3D vertices from which xyzmatrix can successfully extract points.

rawcoords

Whether points have been provided in raw (voxel) coordinates or in calibrated (nm) positions. The default of NA will try to infer this based on the coordinate values.

colpal

A function or character vector of colour names that will be used to set the colour for each layer.

Value

A list of additional class nglayers which can be added to an

ngscene object as produced by ngl_decode_scene.

Details

The ann arguments

See also

Examples

if (FALSE) {
psp=flywire_cave_query('proofreading_status_public_v1')
fwusers=googlesheets4::read_sheet('1G0zqA5DTrfd-a2LuebV4kcqNfl4q1ehlzHBrwT6ZMoc')
psp2=left_join(psp, fwusers, by=c("user_id"="id"))
psp2$layer=psp2$institution
al=ngl_annotation_layers(psp2[c("pt_position", "layer")])

}