Decode scene information from Neuroglancer URL or JSON block

Decode scene information from Neuroglancer URL or JSON block

ngl_decode_scene(
  x,
  return.json = FALSE,
  simplifyVector = TRUE,
  simplifyDataFrame = FALSE,
  ...
)

Arguments

x
Character vector containing single Neuroglancer URL or a json block

return.json
When TRUE extracts the JSON block in a URL does not parse it to an R list

simplifyVector
coerce JSON arrays containing only primitives into an atomic vector

simplifyDataFrame
coerce JSON arrays containing only records (JSON objects) into a data frame

...
arguments passed on to class specific print methods

Value

An R list with additional class ngscene describing the scene, or, when return.json=TRUE, a character vector.

See also

Examples

if (FALSE) {
ngl_decode_scene("<someneuroglancerurl>")

# decode scene from URL currently on clipboard
scene=ngl_decode_scene(clipr::read_clip())

# open a Neuroglancer URL in CATMAID
ngu="<someurl>"
library(elmr)
open_fafb(ngl_decode_scene(ngu))
# Or store the URL rather than opening it
cmu=open_fafb(ngl_decode_scene(ngu), open=FALSE)
}