Decode scene information from Neuroglancer URL or JSON block

+.ngscene adds segments to a neuroglancer scene

-.ngscene removes segments from a neuroglancer scene. It does not complain if the segment is not present.

# S3 method for ngscene
+(x, y)

# S3 method for ngscene
-(x, y)

ngl_decode_scene(
  x,
  return.json = FALSE,
  simplifyVector = TRUE,
  simplifyDataFrame = FALSE,
  ...
)

Arguments

x
Character vector containing single Neuroglancer URL or a json block

y
Segments to add or remove from a neuroglancer scene. Typically as character vectors or by applying ngl_segments to a more complex object. See examples.

return.json
When TRUE extracts the JSON block in a URL does not parse it to an R list

simplifyVector
coerce JSON arrays containing only primitives into an atomic vector

simplifyDataFrame
coerce JSON arrays containing only records (JSON objects) into a data frame

...
additional arguments passed to jsonlite::fromJSON

Value

An R list with additional class ngscene describing the scene, or, when return.json=TRUE, a character vector.

See also

Examples

# \donttest{
# get sample FlyWire URL
fw_url=with_segmentation('flywire', getOption('fafbseg.sampleurl'))
# only a 0 (dummy) segment id present
ngl_segments(fw_url)

#> [1] "0"

#
fw_sc=ngl_decode_scene(fw_url)
fw_sc

#> neuroglancer scene with 2 layers and 1 segments (of which 1 shown)
#>                                 name                    type visible nsegs
#> 1                   Production-image                   image    TRUE     0
#> 2 Production-segmentation_with_graph segmentation_with_graph    TRUE     1
#>   nhidden
#> 1       0
#> 2       0
#>                                                                        source
#> 1 precomputed://gs://microns-seunglab/drosophila_v0/alignment/image_rechunked
#> 2        graphene://https://prodv1.flywire-daf.com/segmentation/table/fly_v31

# add two segments
fw_sc=fw_sc+c("720575940621039145", "720575940626877799")
ngl_segments(fw_sc)

#> [1] "0"                  "720575940621039145" "720575940626877799"

# remove that 0 segment
fw_sc=fw_sc-0
ngl_segments(fw_sc)

#> [1] "720575940621039145" "720575940626877799"

# repeated segments are ignored i.e. no duplicates
ngl_segments(fw_sc+"720575940621039145")

#> [1] "720575940621039145" "720575940626877799"

# convert back to a URL, nb this depends on choose_segmentation
ngl_encode_url(fw_sc)

#> [1] "https://ngl.flywire.ai/#!%7B%22layers%22:[%7B%22source%22:%22precomputed://gs://microns-seunglab/drosophila_v0/alignment/image_rechunked%22,%22type%22:%22image%22,%22blend%22:%22default%22,%22shaderControls%22:%7B%7D,%22name%22:%22Production-image%22%7D,%7B%22source%22:%22graphene://https://prodv1.flywire-daf.com/segmentation/table/fly_v31%22,%22type%22:%22segmentation_with_graph%22,%22segments%22:[%22720575940621039145%22,%22720575940626877799%22],%22skeletonRendering%22:%7B%22mode2d%22:%22lines_and_points%22,%22mode3d%22:%22lines%22%7D,%22graphOperationMarker%22:[%7B%22annotations%22:[],%22tags%22:[]%7D,%7B%22annotations%22:[],%22tags%22:[]%7D],%22pathFinder%22:%7B%22color%22:%22#ffff00%22,%22pathObject%22:%7B%22annotationPath%22:%7B%22annotations%22:[],%22tags%22:[]%7D,%22hasPath%22:false%7D%7D,%22name%22:%22Production-segmentation_with_graph%22%7D],%22navigation%22:%7B%22pose%22:%7B%22position%22:%7B%22voxelSize%22:[4,4,40],%22voxelCoordinates%22:[108360,42086,3279]%7D%7D,%22zoomFactor%22:8%7D,%22perspectiveZoom%22:2230.6094,%22jsonStateServer%22:%22https://globalv1.flywire-daf.com/nglstate/post%22,%22selectedLayer%22:%7B%22layer%22:%22Production-segmentation_with_graph%22,%22visible%22:true%7D,%22layout%22:%22xy-3d%22%7D"

if (FALSE) {
# open in your default browser
browseURL(ngl_encode_url(fw_sc))
}
# }

if (FALSE) {
ngl_decode_scene("<someneuroglancerurl>")

# decode scene from URL currently on clipboard
scene=ngl_decode_scene(clipr::read_clip())

# open a Neuroglancer URL in CATMAID
ngu="<someurl>"
library(elmr)
open_fafb(ngl_decode_scene(ngu))
# Or store the URL rather than opening it
cmu=open_fafb(ngl_decode_scene(ngu), open=FALSE)
}