Encode scene information into a neuroglancer URL

ngl_encode_url converts an R list containing a neuroglancer scene into a URL that you can open in your browser.

as.character.ngscene is another way to convert a neuroglancer scene object to a URL.

ngl_encode_url(body, baseurl = NULL, auto_unbox = TRUE, ...)

# S3 method for ngscene
as.character(x, ...)

Arguments

body
A text file or character vector with JSON data or an R list object

baseurl
A URL specifying the neuroglancer server (if missing, uses options("fafbseg.sampleurl")). You can use any neuroglancer URL as will be appropriately truncated if it encodes scene information.

auto_unbox
automatically unbox all atomic vectors of length 1. It is usually safer to avoid this and instead use the unbox function to unbox individual elements. An exception is that objects of class AsIs (i.e. wrapped in I()) are not automatically unboxed. This is a way to mark single values as length-1 arrays.

...
Additional arguments for toJSON

x
the object to be encoded

Value

Character vector containing encoded URL

Details

We take pains to ensure that entries that neuroglancer expects to be JSON arrays are (including segments and hiddenSegments) are always mapped to a JSON array (even when length 1).

The default baseurl depends on the current segmentation chosen by choose_segmentation.

See also

Examples

# \donttest{
# get sample FlyWire URL
fw_url=with_segmentation('flywire', getOption('fafbseg.sampleurl'))
# only a 0 (dummy) segment id present
ngl_segments(fw_url)

#> [1] "0"

#
fw_sc=ngl_decode_scene(fw_url)
# add a segment
fw_sc$layers[[2]]$segments=union(fw_sc$layers[[2]]$segments,
  "720575940626877799")
# convert back to a URL, nb this depends on choose_segmentation
ngl_encode_url(fw_sc)

#> [1] "https://ngl.flywire.ai/#!%7B%22layers%22:[%7B%22source%22:%22precomputed://gs://microns-seunglab/drosophila_v0/alignment/image_rechunked%22,%22type%22:%22image%22,%22blend%22:%22default%22,%22shaderControls%22:%7B%7D,%22name%22:%22Production-image%22%7D,%7B%22source%22:%22graphene://https://prodv1.flywire-daf.com/segmentation/table/fly_v31%22,%22type%22:%22segmentation_with_graph%22,%22segments%22:[%220%22,%22720575940626877799%22],%22skeletonRendering%22:%7B%22mode2d%22:%22lines_and_points%22,%22mode3d%22:%22lines%22%7D,%22graphOperationMarker%22:[%7B%22annotations%22:[],%22tags%22:[]%7D,%7B%22annotations%22:[],%22tags%22:[]%7D],%22pathFinder%22:%7B%22color%22:%22#ffff00%22,%22pathObject%22:%7B%22annotationPath%22:%7B%22annotations%22:[],%22tags%22:[]%7D,%22hasPath%22:false%7D%7D,%22name%22:%22Production-segmentation_with_graph%22%7D],%22navigation%22:%7B%22pose%22:%7B%22position%22:%7B%22voxelSize%22:[4,4,40],%22voxelCoordinates%22:[108360,42086,3279]%7D%7D,%22zoomFactor%22:8%7D,%22perspectiveZoom%22:2230.6094,%22jsonStateServer%22:%22https://globalv1.flywire-daf.com/nglstate/post%22,%22selectedLayer%22:%7B%22layer%22:%22Production-segmentation_with_graph%22,%22visible%22:true%7D,%22layout%22:%22xy-3d%22%7D"

# another way to do this, which long time R users may find more intuitive
as.character(fw_sc)

#> [1] "https://ngl.flywire.ai/#!%7B%22layers%22:[%7B%22source%22:%22precomputed://gs://microns-seunglab/drosophila_v0/alignment/image_rechunked%22,%22type%22:%22image%22,%22blend%22:%22default%22,%22shaderControls%22:%7B%7D,%22name%22:%22Production-image%22%7D,%7B%22source%22:%22graphene://https://prodv1.flywire-daf.com/segmentation/table/fly_v31%22,%22type%22:%22segmentation_with_graph%22,%22segments%22:[%220%22,%22720575940626877799%22],%22skeletonRendering%22:%7B%22mode2d%22:%22lines_and_points%22,%22mode3d%22:%22lines%22%7D,%22graphOperationMarker%22:[%7B%22annotations%22:[],%22tags%22:[]%7D,%7B%22annotations%22:[],%22tags%22:[]%7D],%22pathFinder%22:%7B%22color%22:%22#ffff00%22,%22pathObject%22:%7B%22annotationPath%22:%7B%22annotations%22:[],%22tags%22:[]%7D,%22hasPath%22:false%7D%7D,%22name%22:%22Production-segmentation_with_graph%22%7D],%22navigation%22:%7B%22pose%22:%7B%22position%22:%7B%22voxelSize%22:[4,4,40],%22voxelCoordinates%22:[108360,42086,3279]%7D%7D,%22zoomFactor%22:8%7D,%22perspectiveZoom%22:2230.6094,%22jsonStateServer%22:%22https://globalv1.flywire-daf.com/nglstate/post%22,%22selectedLayer%22:%7B%22layer%22:%22Production-segmentation_with_graph%22,%22visible%22:true%7D,%22layout%22:%22xy-3d%22%7D"

if (FALSE) {
# open in your default browser
browseURL(ngl_encode_url(fw_sc))
# ... or
browseURL(as.character(fw_sc))
}
# }

if (FALSE) {
# copy JSON scene information from {} symbol at top right of neuroglancer
# now make a permanent URL for the scene
ngl_encode_url(clipr::read_clip())
}