Encode scene information into a neuroglancer URL

Encode scene information into a neuroglancer URL

ngl_encode_url(body, baseurl = getOption("fafbseg.baseurl"),
  auto_unbox = TRUE, fix_segments = TRUE, ...)

Arguments

body
A text file or character vector with JSON data or an R list object

baseurl
The base URL including the neuroglancer server

auto_unbox
automatically unbox all atomic vectors of length 1. It is usually safer to avoid this and instead use the unbox function to unbox individual elements. An exception is that objects of class AsIs (i.e. wrapped in I()) are not automatically unboxed. This is a way to mark single values as length-1 arrays.

fix_segments
Fix URL when only one segment in scene (see details)

...
Additional arguments for toJSON

Value

Character vector containing encoded URL

Details

When the neuroglancer URL scene refers to just one segment the only way we can currently ensure correctly formed JSON is to add a dummy 0 segment (thus forming a JSON array).

See also

Examples

if (FALSE) {
# copy JSON scene information from {} symbol at top right of neuroglancer
# now make a permanent URL for the scene
ngl_encode_url(clipr::read_clip())
}