Helper function to turn diverse inputs into neuroglancer segment ids

Helper function to turn diverse inputs into neuroglancer segment ids

ngl_segments(x, as_character = FALSE, include_hidden = TRUE)

Arguments

x
A neuroglancer scene specification form either in raw JSON format (character vector), the path to a file on disk, a neuroglancer scene URL (which embeds a JSON scene specification in a single URL), or an R list generated by parsing one of the above.

as_character
Whether to return segments as character rather than numeric vector.

include_hidden
Whether to include hiddenSegments (typically for flywire scenes).

Value

Numeric (or character) vector of segment ids

Examples

# no change
ngl_segments(c(10950626347, 10952282491, 13307888342))

#> [1] 10950626347 10952282491 13307888342

# just turns these into numeric
ngl_segments(c("10950626347", "10952282491", "13307888342"))

#> [1] 10950626347 10952282491 13307888342

if (FALSE) {
# from clipboard
ngl_segments(clipr::read_clip())

# URL
ngl_segments("<ngl-scene-url>")
# path to file on disk
ngl_segments("/path/to/scene.json")
# R list
ngl_segments(scenelist)
}