Construct Neuroglancer URL based on 3D location data

Construct Neuroglancer URL based on 3D location data

open_fafb_ngl(x, s = rgl::select3d(), zoomFactor = 8, sampleurl = NULL,
  open = interactive(), ...)

Arguments

x
A numeric vector OR any object compatible with xyzmatrix OR a CATMAID URL (see details)

s
Optional selection function of the type returned by select3d

zoomFactor
The Neuroglancer zoomFactor (bigger means zoomed out)

sampleurl
A sample URL that defines your neuroglancer dataset.

open
Whether or not to open the URL in a browser - this defaults to TRUE in interactive use.

...
Additional arguments passed to fromJSON to control JSON parsing.

Details

Neuroglancer scenes seem to be specified in a single URL that URLencodes a json object defining layers to display, position etc. This function works by taking a sample URL defining such a scene and then editing it to point to a new 3D location / adjust zoom.

This package comes with a default scene url. If you specify a different sample URL to the sampleurl argument it will be remember for the rest of the R session. If you regularly use a particular kind of scene URL, you can set options(fafbseg.sampleurl) in your Rprofile file.

Examples

# NOT RUN {
# Open a location in MB peduncle
open_fafb_ngl(c(433440, 168344, 131200))

# open a CATMAID URL in Neuroglancer
u=paste0("https://fafb.catmaid.virtualflybrain.org/?pid=2&zp=131280&",
"yp=170014.98879622458&xp=426584.81386896875&tool=navigator&sid0=2&s0=-1")
open_fafb_ngl(u)

# Set an existing scene URL (pointing to any old location) to act as
# the template for open_fafb_ngl
# nb the package sets one for you on startup if you haven't set yourself
options(fafbseg.sampleurl="https://<neuroglancerlurl>")
# Edit your R profile if you want to set a different default
usethis::edit_r_profile()
# }